Fixed Australia Blast furnace gas for 2010–2020

Australia's Blast furnaces have an undesirable characteristic that leads to singular matrices: From 2013 onward, the production of Blast furnace gas by Blast furnaces is consumed only by Blast furnaces. No other industry or energy production machine consumes Blast furnace gas. In fact, the problem is deeper, starting in 2010, the Iron and steel industry consumes no Blast furnace gas, in apparent contradiction to the IEA's own policies for reporting Blast furnace gas consumption. This data frame provides the "fixed" Blast furnace gas data for Australia for the 2010–2020 timeframe.

Usage

Fixed_AUS_bfg

Format

A data frame with 11 columns.

Details

The fix involves ensuring that the Iron and steel industry always consumes Blast furnace gas, according to the IEA's assumed efficiency of 40%. The function fix_AUS_bfg() makes use of these data.

Examples

Fixed_AUS_bfg
#> # A tibble: 22 × 11
#>    Country Method EnergyType LastStage LedgerSide FlowAggregationPoint    Flow  
#>    <chr>   <chr>  <chr>      <chr>     <chr>      <chr>                   <chr> 
#>  1 AUS     PCM    E          Final     Supply     Energy industry own use Blast…
#>  2 AUS     PCM    E          Final     Supply     Energy industry own use Blast…
#>  3 AUS     PCM    E          Final     Supply     Energy industry own use Blast…
#>  4 AUS     PCM    E          Final     Supply     Energy industry own use Blast…
#>  5 AUS     PCM    E          Final     Supply     Energy industry own use Blast…
#>  6 AUS     PCM    E          Final     Supply     Energy industry own use Blast…
#>  7 AUS     PCM    E          Final     Supply     Energy industry own use Blast…
#>  8 AUS     PCM    E          Final     Supply     Energy industry own use Blast…
#>  9 AUS     PCM    E          Final     Supply     Energy industry own use Blast…
#> 10 AUS     PCM    E          Final     Supply     Energy industry own use Blast…
#> # ℹ 12 more rows
#> # ℹ 4 more variables: Product <chr>, Unit <chr>, Year <dbl>, Edot <dbl>