Fixed Ghana Primary solid biofuels data

Ghana's Primary solid biofuels data show a very large and dramatic decline from 1999 to 2000 in IEA WEEB data. This decline is due to new survey data being used for the 2000 data. When we look at the PSB data on a per-capita basis, it is clear that a near-constant PSB/capita value was used to extrapolate per-capita usage in the late 1990s. When new survey data became available for the 2000 reporting year, the per-capita consumption of PSB obviously changed. Our approach to this problem is to smooth out the really big peak in PSB consumption by reducing the per-capita consumption of PSB, starting in 1991. This data frame contains the "fixed" data. The function fix_GHA_psb() makes use of these data.

Usage

Fixed_GHA_PSB

Format

A data frame with 11 columns.

Examples

Fixed_GHA_PSB
#> # A tibble: 36 × 11
#>    Country Method EnergyType LastStage LedgerSide FlowAggregationPoint     Flow 
#>    <chr>   <chr>  <chr>      <chr>     <chr>      <chr>                    <chr>
#>  1 GHA     PCM    E          Final     Supply     Total primary energy su… Prod…
#>  2 GHA     PCM    E          Final     Supply     Total primary energy su… Prod…
#>  3 GHA     PCM    E          Final     Supply     Total primary energy su… Prod…
#>  4 GHA     PCM    E          Final     Supply     Total primary energy su… Prod…
#>  5 GHA     PCM    E          Final     Supply     Total primary energy su… Prod…
#>  6 GHA     PCM    E          Final     Supply     Total primary energy su… Prod…
#>  7 GHA     PCM    E          Final     Supply     Total primary energy su… Prod…
#>  8 GHA     PCM    E          Final     Supply     Total primary energy su… Prod…
#>  9 GHA     PCM    E          Final     Supply     Total primary energy su… Prod…
#> 10 GHA     PCM    E          Final     Supply     Transformation processes Char…
#> # ℹ 26 more rows
#> # ℹ 4 more variables: Product <chr>, Unit <chr>, Year <dbl>, Edot <dbl>