Fixed Other non-OECD Americas Charcoal production 1971–2010

Other Non-OECD Americas has several years (1971–2010) in which Charcoal is produced but no Primary solid biofuels are consumed to create the Charcoal. This object contains (presumably) correct data. In particular, Charcoal production plants now consume Primary solid biofuels in all years, and Primary solid biofuels production is boosted accordingly. The efficiency of Charcoal production plants in 2011 was used to create the filled data. The function fix_OAMR_cpp() makes use of these data.

Usage

Fixed_OAMR_cpp

Format

A data frame with 11 columns.

Examples

Fixed_OAMR_cpp
#> # A tibble: 120 × 11
#>    Country Method EnergyType LastStage LedgerSide FlowAggregationPoint     Flow 
#>    <chr>   <chr>  <chr>      <chr>     <chr>      <chr>                    <chr>
#>  1 OAMR    PCM    E          Final     Supply     Total primary energy su… Prod…
#>  2 OAMR    PCM    E          Final     Supply     Total primary energy su… Prod…
#>  3 OAMR    PCM    E          Final     Supply     Total primary energy su… Prod…
#>  4 OAMR    PCM    E          Final     Supply     Total primary energy su… Prod…
#>  5 OAMR    PCM    E          Final     Supply     Total primary energy su… Prod…
#>  6 OAMR    PCM    E          Final     Supply     Total primary energy su… Prod…
#>  7 OAMR    PCM    E          Final     Supply     Total primary energy su… Prod…
#>  8 OAMR    PCM    E          Final     Supply     Total primary energy su… Prod…
#>  9 OAMR    PCM    E          Final     Supply     Total primary energy su… Prod…
#> 10 OAMR    PCM    E          Final     Supply     Total primary energy su… Prod…
#> # ℹ 110 more rows
#> # ℹ 4 more variables: Product <chr>, Unit <chr>, Year <dbl>, Edot <dbl>