Adds a nuclear industry

In the IEA World Energy Extended Balances, there is a "Nuclear industry" Energy industry own use flow, but there is no "Nuclear industry" in transformation processes flows, which prevents from defining a nuclear industry in the PSUT. However, using the World Energy Extended Balances documentation, one can deduce from the amount of nuclear fuel used by "Main activity producer electricity plants" and "Main activity producer CHP plants" the energy transformation due to the nuclear industry. This function performs that task. The function is called within the specify_all() function.

Usage

add_nuclear_industry(
  .tidy_iea_df,
  ascribe_eiou_to_nuclear = FALSE,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  product = IEATools::iea_cols$product,
  method = IEATools::iea_cols$method,
  ledger_side = IEATools::iea_cols$ledger_side,
  last_stage = IEATools::iea_cols$last_stage,
  energy_type = IEATools::iea_cols$energy_type,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  unit = IEATools::iea_cols$unit,
  eiou = IEATools::aggregation_flows$energy_industry_own_use,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  nuclear_industry = IEATools::eiou_flows$nuclear_industry,
  main_act_producer_elect = IEATools::main_act_plants$main_act_prod_elect_plants,
  main_act_producer_chp = IEATools::main_act_plants$main_act_prod_chp_plants,
  autoproducer_elect = IEATools::main_act_plants$autoprod_elect_plants,
  autoproducer_chp = IEATools::transformation_processes$autoproducer_CHP_plants,
  own_use_elect_chp_heat = IEATools::eiou_flows$own_use_elect_chp_heat_plants,
  nuclear = IEATools::nuclear_products$nuclear,
  electricity = IEATools::electricity_products$electricity,
  heat = IEATools::heat_products$heat,
  negzeropos = ".negzeropos",
  share_elect_output_From_Func = ".share_elect_output_From_Func",
  share_nuclear_output = ".share_nuclear_output",
  ratio_output_to_nuclear_fuel = 0.33
)

Arguments

.tidy_iea_df

The .tidy_iea_df which flows need to be specified.

ascribe_eiou_to_nuclear

A boolean defining whether a fraction of the EIOU of electricity, CHP and heat plants should be ascribed to the new nuclear industry. Default is FALSE.

flow_aggregation_point

The name of the flow aggregation point column in the .tidy_iea_df. Default is IEATools::iea_cols$flow_aggregation_point.

flow

The name of the flow column in the .tidy_iea_df. Default is IEATools::iea_cols$flow.

e_dot

The name of the energy column in the .tidy_iea_df. Default is IEATools::iea_cols$e_dot.

product

The name of the product column in the .tidy_iea_df. Default is IEATools::iea_cols$product.

method

The name of the method column in the .tidy_iea_df. Default is IEATools::iea_cols$method.

ledger_side

The name of the ledger side column in the .tidy_iea_df. Default is IEATools::iea_cols$ledger_side.

last_stage

The name of the last stage column in the .tidy_iea_df. Default is IEATools::iea_cols$last_stage.

energy_type

The name of the energy type column in the .tidy_iea_df. Default is IEATools::iea_cols$energy_type.

country

The name of the country column in the .tidy_iea_df. Default is IEATools::iea_cols$country.

year

The name of the year column in the .tidy_iea_df. Default is IEATools::iea_cols$year.

unit

The name of the unit column in the .tidy_iea_df. Default is IEATools::iea_cols$unit.

eiou

A string identifying the energy industry own use in the flow_aggregation_point column in the .tidy_iea_df. Default is IEATools::aggregation_flows$energy_industry_own_use.

transformation_processes

A string identifying the transformation processes in the flow_aggregation_point column in the .tidy_iea_df. Default is IEATools::aggregation_flows$transformation_processes.

nuclear_industry

A string identifying "Nuclear industry" in the flow column of the .tidy_iea_df. Default is IEATools::eiou_flows$nuclear_industry.

main_act_producer_elect

A string identifying "Main activity producer electricity plants" in the flow column of the .tidy_iea_df. Default is IEATools::main_act_plants$main_act_prod_elect_plants.

main_act_producer_chp

A string identifying "Main activity producer CHP plants" in the flow column of the .tidy_iea_df. Default is IEATools::main_act_plants$main_act_prod_chp_plants.

autoproducer_elect

A string identifying "Autoproducer electricity plants" in the flow column of the .tidy_iea_df. Default is IEATools::main_act_plants$autoprod_elect_plants.

autoproducer_chp

A string identifying "Autoproducer CHP plants" in the flow column of the .tidy_iea_df. Default is IEATools::transformation_processes$autoproducer_CHP_plants.

own_use_elect_chp_heat

A string identifying "Own use in electricity, CHP and heat plants" in the flow column of the .tidy_iea_df. Default is IEATools::eiou_flows$own_use_elect_chp_heat_plants.

nuclear

A string identifying the "Nuclear" product in the product column of the tidy_iea_df. Default is "Nuclear".

electricity

A string identifying the "Electricity" product in the product column of the tidy_iea_df. Default is IEATools::electricity_products$electricity.

heat

A string identifying the "Heat" product in the product column of the tidy_iea_df. Default is IEATools::heat_products$heat.

negzeropos

The name of a temporary column added to the data frame. Default is ".negzeropos".

share_elect_output_From_Func

The name of a temporary column added to the data frame. Default is ".share_elect_output_From_Func".

share_nuclear_output

The name of a temporary column added to the data frame. Default is ".share_nuclear_output".

ratio_output_to_nuclear_fuel

A parameter that describes the correspondance between input of nuclear fuel and output of electricity and/or heat. The IEA World Energy Extended Balances state that the value adopted in the balances is 0.33, which is therefore the default value of the parameter.

Value

A modified version of the .tidy_iea_df, with a nuclear industry added as an additional transformation process.

Details

The World Energy Extended Balances documentation states that "The primary energy equivalent of nuclear electricity is calculated from the gross generation by assuming a 33% conversion efficiency. The calculation to be carried out is the following: gross electricity generation in TWh x0.086 / 0.33 = primary energy equivalent in Mtoe."

Hence this function does the following:

the Nuclear fuel consumed by Main activity producer electricity & heat plants is ascribed to nuclear industry plants;

the output of Main activity producer electricity and heat plants to be directed to nuclear plants is determined by multiplying their nuclear fuel consumption per 0.33. In the case of CHP plants, that output is divided into heat and electricity according to the shares of output of each of these two products;

the output ascribed to nuclear plants is subtracted from Main activity producer electricity and heat plants.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
load_tidy_iea_df() %>% 
  add_nuclear_industry()
#> # A tibble: 404 × 11
#>    Country Method EnergyType LastStage  Year LedgerSide  FlowAggregationPoint
#>    <chr>   <chr>  <chr>      <chr>     <dbl> <chr>       <chr>               
#>  1 GHA     PCM    E          Final      1971 Consumption Industry            
#>  2 GHA     PCM    E          Final      1971 Consumption Industry            
#>  3 GHA     PCM    E          Final      1971 Consumption Industry            
#>  4 GHA     PCM    E          Final      1971 Consumption Industry            
#>  5 GHA     PCM    E          Final      1971 Consumption Industry            
#>  6 GHA     PCM    E          Final      1971 Consumption Industry            
#>  7 GHA     PCM    E          Final      1971 Consumption Industry            
#>  8 GHA     PCM    E          Final      1971 Consumption Non-energy use      
#>  9 GHA     PCM    E          Final      1971 Consumption Other               
#> 10 GHA     PCM    E          Final      1971 Consumption Other               
#> # ℹ 394 more rows
#> # ℹ 4 more variables: Flow <chr>, Product <chr>, Unit <chr>, Edot <dbl>