Add products to production industry names

The IEA specify all extraction processes as "Production". When combined with an input-output view of the world within the Physical Supply-Use Table (PSUT) framework, production of any one resource implies production of all other resources. That's clearly non-physical, as production of some Crude oil does not imply harvesting of Solar energy. The solution to this problem is to rename the Flow values to be Production (Product). For example, the industry that produces Crude oil will be renamed to Production (Crude oil).

add_production_details(.tidy_iea_df, flow = "Flow",
  production = "Production", product = "Product")

Arguments

.tidy_iea_df
an IEA data frame whose columns have been renamed by rename_iea_df_cols()

flow
the name of the flow column in .tidy_iea_df. Default is "Flow".

production
a string that identifies production in the flow column. Default is "Production".

product
the name of the product column in .iea_df. Default is "Product".

Value

.iea_df with detailed production industries

Details

If the suffix of the value of the Flow column is already "(Product)", no changes to the value in the Flow column will be made.

Examples

load_tidy_iea_df() %>%
  add_production_details()

#> # A tibble: 399 x 11
#>    Method Last.stage Country Year  Ledger.side Flow.aggregatio… Energy.type
#>    <chr>  <chr>      <chr>   <chr> <chr>       <chr>            <chr>      
#>  1 PCM    Final      GH      1971  Supply      Total primary e… E          
#>  2 PCM    Final      GH      1971  Supply      Total primary e… E          
#>  3 PCM    Final      GH      1971  Supply      Total primary e… E          
#>  4 PCM    Final      GH      1971  Supply      Total primary e… E          
#>  5 PCM    Final      GH      1971  Supply      Total primary e… E          
#>  6 PCM    Final      GH      1971  Supply      Total primary e… E          
#>  7 PCM    Final      GH      1971  Supply      Total primary e… E          
#>  8 PCM    Final      GH      1971  Supply      Total primary e… E          
#>  9 PCM    Final      GH      1971  Supply      Total primary e… E          
#> 10 PCM    Final      GH      1971  Supply      Total primary e… E          
#> # … with 389 more rows, and 4 more variables: Unit <chr>, Flow <chr>,
#> #   Product <chr>, E.dot <dbl>