The path to the default aggregation table

The path to the default aggregation table

Usage

default_aggregation_region_table_path(version = 2019)

Arguments

version

the desired version (expressed as the year of IEA data release) of the aggregation table. Options are 2019 (default), and 2020.

Value

the path to the default aggregation table for the desired IEA data version

Examples

default_aggregation_region_table_path()     # Assumes 2019
#> [1] "/private/var/folders/cy/_tgrtb4966v60qqmy2g2k3vr0000gq/T/RtmpHAmm72/temp_libpath150f91e28f1cb/IEATools/extdata/aggregation_table_iea_exiobase_2019.xlsx"
default_aggregation_region_table_path(2019) # Same
#> [1] "/private/var/folders/cy/_tgrtb4966v60qqmy2g2k3vr0000gq/T/RtmpHAmm72/temp_libpath150f91e28f1cb/IEATools/extdata/aggregation_table_iea_exiobase_2019.xlsx"
# Returns path for default aggregation table appropriate for 2020 IEA data release
default_aggregation_region_table_path(2020) 
#> [1] "/private/var/folders/cy/_tgrtb4966v60qqmy2g2k3vr0000gq/T/RtmpHAmm72/temp_libpath150f91e28f1cb/IEATools/extdata/aggregation_table_iea_exiobase_2020.xlsx"