Extract a unit summation matrix from a tidy data frame

Unit summation matrices have products in rows and units in columns, with 1s where a product is expressed in the unit and 0s otherwise.

extract_S_units_from_tidy(.tidy_iea_df, product = "Product",
  unit = "Unit", s_units = "S_units", val = ".val",
  rowtype = ".rowtype", coltype = ".coltype")

Arguments

.tidy_iea_df
the tidy data frame from which a unit summation S_units matrix is to be formed.

product
the name of the product column in .tidy_iea_df. Default is "Product".

unit
the name of the unit column in ".tidy_iea_df". Default is "Unit".

s_units
the name of the unit summation column to be added to .tidy_iea_df. Default is "S_unit".

val
the name of a temporary column to be created in .tidy_iea_df. Deafult is ".val".

rowtype
the name of a temporary rowype column created in .tidy_iea_df. Default is "rowtype".

coltype
the name of a temporary colype column created in .tidy_iea_df. Default is "coltype".

Value

a data frame containing grouping variables and a new column of unit summation matrices called s_unit.

Details

.tidy_iea_df should be grouped as needed, typically on Country, Year, Energy.type, Last.stage, etc., but not on Unit, Flow or Product. .tidy_iea_df is typically obtained from tidy_iea_df().

Examples

library(dplyr)
load_tidy_iea_df() %>%
  group_by(Country, Year, Energy.type) %>%
  extract_S_units_from_tidy()

#>   Country Year Energy.type
#> 1      GH 1971           E
#> 2      GH 2000           E
#> 3      ZA 1971           E
#> 4      ZA 2000           E
#>                                                                           S_units
#> 1                                        1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 2                                  1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 3                   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 4 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1