IEA order for combined flow aggregation point and flow

A string vector containing entries for a united flow aggregation point and flow column in IEA order.

Usage

fap_flows

Format

A string list with 103

Examples

fap_flows
#>   [1] "Total primary energy supply_Production"                             
#>   [2] "Total primary energy supply_Coal mines"                             
#>   [3] "Total primary energy supply_Oil and gas extraction"                 
#>   [4] "Total primary energy supply_Imports"                                
#>   [5] "Total primary energy supply_Exports"                                
#>   [6] "Total primary energy supply_International marine bunkers"           
#>   [7] "Total primary energy supply_International aviation bunkers"         
#>   [8] "Total primary energy supply_Stock changes"                          
#>   [9] "TFC compare_Transfers"                                              
#>  [10] "TFC compare_Statistical differences"                                
#>  [11] "Transformation processes_Coal mines"                                
#>  [12] "Transformation processes_Oil and gas extraction"                    
#>  [13] "Transformation processes_Nuclear industry"                          
#>  [14] "Transformation processes_Main activity producer electricity plants" 
#>  [15] "Transformation processes_Autoproducer electricity plants"           
#>  [16] "Transformation processes_Main activity producer CHP plants"         
#>  [17] "Transformation processes_Autoproducer CHP plants"                   
#>  [18] "Transformation processes_Main activity producer heat plants"        
#>  [19] "Transformation processes_Autoproducer heat plants"                  
#>  [20] "Transformation processes_Heat pumps"                                
#>  [21] "Transformation processes_Electric boilers"                          
#>  [22] "Transformation processes_Chemical heat for electricity production"  
#>  [23] "Transformation processes_Blast furnaces"                            
#>  [24] "Transformation processes_Gas works"                                 
#>  [25] "Transformation processes_Coke ovens"                                
#>  [26] "Transformation processes_Patent fuel plants"                        
#>  [27] "Transformation processes_BKB/peat briquette plants"                 
#>  [28] "Transformation processes_Oil refineries"                            
#>  [29] "Transformation processes_Petrochemical plants"                      
#>  [30] "Transformation processes_Coal liquefaction plants"                  
#>  [31] "Transformation processes_Gas-to-liquids (GTL) plants"               
#>  [32] "Transformation processes_For blended natural gas"                   
#>  [33] "Transformation processes_Charcoal production plants"                
#>  [34] "Transformation processes_Non-specified"                             
#>  [35] "Energy industry own use_Coal mines"                                 
#>  [36] "Energy industry own use_Oil and gas extraction"                     
#>  [37] "Energy industry own use_Blast furnaces"                             
#>  [38] "Energy industry own use_Gas works"                                  
#>  [39] "Energy industry own use_Gasification plants for biogases"           
#>  [40] "Energy industry own use_Coke ovens"                                 
#>  [41] "Energy industry own use_Patent fuel plants"                         
#>  [42] "Energy industry own use_BKB/peat briquette plants"                  
#>  [43] "Energy industry own use_Oil refineries"                             
#>  [44] "Energy industry own use_Coal liquefaction plants"                   
#>  [45] "Energy industry own use_Liquefaction (LNG) / regasification plants" 
#>  [46] "Energy industry own use_Gas-to-liquids (GTL) plants"                
#>  [47] "Energy industry own use_Own use in electricity, CHP and heat plants"
#>  [48] "Energy industry own use_Pumped storage plants"                      
#>  [49] "Energy industry own use_Nuclear industry"                           
#>  [50] "Energy industry own use_Main activity producer electricity plants"  
#>  [51] "Energy industry own use_Main activity producer CHP plants"          
#>  [52] "Energy industry own use_Main activity producer heat plants"         
#>  [53] "Energy industry own use_Oil extraction"                             
#>  [54] "Energy industry own use_Natural gas extraction"                     
#>  [55] "Energy industry own use_Charcoal production plants"                 
#>  [56] "Energy industry own use_Non-specified"                              
#>  [57] "TFC compare_Losses"                                                 
#>  [58] "Industry_Mining and quarrying"                                      
#>  [59] "Industry_Construction"                                              
#>  [60] "Industry_Iron and steel"                                            
#>  [61] "Industry_Chemical and petrochemical"                                
#>  [62] "Industry_Non-ferrous metals"                                        
#>  [63] "Industry_Non-metallic minerals"                                     
#>  [64] "Industry_Transport equipment"                                       
#>  [65] "Industry_Machinery"                                                 
#>  [66] "Industry_Food and tobacco"                                          
#>  [67] "Industry_Paper, pulp and printing"                                  
#>  [68] "Industry_Wood and wood products"                                    
#>  [69] "Industry_Textile and leather"                                       
#>  [70] "Industry_Industry not elsewhere specified"                          
#>  [71] "Industry_Non-specified (industry)"                                  
#>  [72] "Transport_World aviation bunkers"                                   
#>  [73] "Transport_Domestic aviation"                                        
#>  [74] "Transport_International aviation"                                   
#>  [75] "Transport_Road"                                                     
#>  [76] "Transport_Rail"                                                     
#>  [77] "Transport_Pipeline transport"                                       
#>  [78] "Transport_World marine bunkers"                                     
#>  [79] "Transport_Domestic navigation"                                      
#>  [80] "Transport_International navigation"                                 
#>  [81] "Transport_Transport not elsewhere specified"                        
#>  [82] "Transport_Non-specified (transport)"                                
#>  [83] "Other_Residential"                                                  
#>  [84] "Other_Commercial and public services"                               
#>  [85] "Other_Agriculture/forestry"                                         
#>  [86] "Other_Fishing"                                                      
#>  [87] "Other_Final consumption not elsewhere specified"                    
#>  [88] "Non-energy use_Non-energy use industry/transformation/energy"       
#>  [89] "Non-energy use_Non-energy use in construction"                      
#>  [90] "Non-energy use_Non-energy use in mining and quarrying"              
#>  [91] "Non-energy use_Non-energy use in iron and steel"                    
#>  [92] "Non-energy use_Non-energy use in chemical/petrochemical"            
#>  [93] "Non-energy use_Non-energy use in non-ferrous metals"                
#>  [94] "Non-energy use_Non-energy use in non-metallic minerals"             
#>  [95] "Non-energy use_Non-energy use in transport equipment"               
#>  [96] "Non-energy use_Non-energy use in machinery"                         
#>  [97] "Non-energy use_Non-energy use in food/beverages/tobacco"            
#>  [98] "Non-energy use_Non-energy use in paper/pulp and printing"           
#>  [99] "Non-energy use_Non-energy use in wood and wood products"            
#> [100] "Non-energy use_Non-energy use in textiles and leather"              
#> [101] "Non-energy use_Non-energy use in industry not elsewhere specified"  
#> [102] "Non-energy use_Non-energy use in transport"                         
#> [103] "Non-energy use_Non-energy use in other"