Final demand sectors

A string vector containing final demand sectors used for calculating Total Final Consumption (TFC)

fd_sectors

Format

A string list with 51

eiou_flows

The string vector identifying energy industry own use flows.

industry_net_flows

The string vector identifying non-eiou (net) Industry flows.

transport_domestic_flows

The string vector identifying domestic transport flows.

other_flows

The string vector identifying Other flows.

Examples

fd_sectors

#> $bkb_peat_briquette_plants
#> [1] "BKB/peat briquette plants"
#> 
#> $blast_furnaces
#> [1] "Blast furnaces"
#> 
#> $charcoal_plants
#> [1] "Charcoal production plants"
#> 
#> $coal_liquefaction_plants
#> [1] "Coal liquefaction plants"
#> 
#> $coal_mines
#> [1] "Coal mines"
#> 
#> $coke_ovens
#> [1] "Coke ovens"
#> 
#> $gas_work
#> [1] "Gas works"
#> 
#> $gas_to_liquids_plants
#> [1] "Gas-to-liquids (GTL) plants"
#> 
#> $gasification_plants
#> [1] "Gasification plants for biogases"
#> 
#> $liquefaction_regasification_plants
#> [1] "Liquefaction (LNG) / regasification plants"
#> 
#> $non_specified_eiou
#> [1] "Non-specified (energy)"
#> 
#> $nuclear_industry
#> [1] "Nuclear industry"
#> 
#> $oil_and_gas_extraction
#> [1] "Oil and gas extraction"
#> 
#> $oil_refineries
#> [1] "Oil refineries"
#> 
#> $own_use_elect_chp_heat_plants
#> [1] "Own use in electricity, CHP and heat plants"
#> 
#> $patent_fuel_plants
#> [1] "Patent fuel plants"
#> 
#> $pumped_storage_plants
#> [1] "Pumped storage plants"
#> 
#> $mining_and_quarrying
#> [1] "Mining and quarrying"
#> 
#> $construction
#> [1] "Construction"
#> 
#> $iron_and_steel
#> [1] "Iron and steel"
#> 
#> $chemical_and_petrochemical
#> [1] "Chemical and petrochemical"
#> 
#> $non_ferrous_metals
#> [1] "Non-ferrous metals"
#> 
#> $non_metallic_minerals
#> [1] "Non-metallic minerals"
#> 
#> $transport_equipment
#> [1] "Transport equipment"
#> 
#> $machinery
#> [1] "Machinery"
#> 
#> $food_and_tobacco
#> [1] "Food and tobacco"
#> 
#> $paper_pulp_and_print
#> [1] "Paper, pulp and print"
#> 
#> $paper_pulp_and_printing
#> [1] "Paper, pulp and printing"
#> 
#> $wood_and_wood_products
#> [1] "Wood and wood products"
#> 
#> $textile_and_leather
#> [1] "Textile and leather"
#> 
#> $non_specified_industry
#> [1] "Non-specified (industry)"
#> 
#> $industry_not_elsewhere_specified
#> [1] "Industry not elsewhere specified"
#> 
#> $oil_extraction
#> [1] "Oil extraction"
#> 
#> $natural_gas_extraction
#> [1] "Natural gas extraction"
#> 
#> $domestic_navigation
#> [1] "Domestic navigation"
#> 
#> $world_marine_bunkers
#> [1] "World marine bunkers"
#> 
#> $international_navigation
#> [1] "International navigation"
#> 
#> $domestic_aviation
#> [1] "Domestic aviation"
#> 
#> $world_aviation_bunkers
#> [1] "World aviation bunkers"
#> 
#> $international_aviation
#> [1] "International aviation"
#> 
#> $road
#> [1] "Road"
#> 
#> $rail
#> [1] "Rail"
#> 
#> $pipeline_transport
#> [1] "Pipeline transport"
#> 
#> $non_specified_transport
#> [1] "Non-specified (transport)"
#> 
#> $transport_not_elsewhere_specified
#> [1] "Transport not elsewhere specified"
#> 
#> $residential
#> [1] "Residential"
#> 
#> $commercial_and_public_services
#> [1] "Commercial and public services"
#> 
#> $agriculture_forestry
#> [1] "Agriculture/forestry"
#> 
#> $fishing
#> [1] "Fishing"
#> 
#> $non_specified_other
#> [1] "Non-specified (other)"
#> 
#> $final_consumption_not_elsewhere_specified
#> [1] "Final consumption not elsewhere specified"
#>