Find the point where Energy industry own use starts

Given a country's IEA extended energy balance data frame, find the row numbers that represent the transition between the Transformation processes and Energy industry own use. This function simply calls find_transformation_end(), because the dividing line between Transformation processes and Energy industry own use is also the point where Energy industry own use starts.

find_eiou_start(
  .ctry_tbl,
  flow = IEATools::iea_cols$flow,
  non_specified = IEATools::non_specified_flows$non_specified,
  eiou = IEATools::tfc_compare_flows$energy_industry_own_use,
  coal_mines = IEATools::industry_flows$coal_mines
)

Arguments

.ctry_tbl
a country's IEA data frame

flow
See IEATools::iea_cols$flow.

non_specified
See IEATools::non_specified_flows.

eiou
See IEATools::tfc_compare_flows.

coal_mines
the name for coal mines in the flow column. Default is "Coal mines".

Value

a pair of integers representing the rows that straddle the split between Transformation processes and Energy industry own use The second integer is the first row of Energy industry own use.

Details

Arguments should be supplied by the calling function.

An error is given if this function fails to find the location of the split between Transformation processes and Energy industry own use.