Create final-to-useful efficiency vectors (eta_fu) and exergy-to-energy ratio vectors (phi_u) from a final-to-useful efficiency table

Create final-to-useful efficiency vectors (eta_fu) and exergy-to-energy ratio vectors (phi_u) from a final-to-useful efficiency table

form_eta_fu_phi_u_vecs(
  .eta_fu_table,
  unit = IEATools::iea_cols$unit,
  year = IEATools::iea_cols$year,
  quantity = IEATools::template_cols$quantity,
  machine = IEATools::template_cols$machine,
  eu_product = IEATools::template_cols$eu_product,
  e_dot_machine = IEATools::template_cols$e_dot_machine,
  e_dot_machine_perc = IEATools::template_cols$e_dot_machine_perc,
  maximum_values = IEATools::template_cols$maximum_values,
  eta_fu = IEATools::template_cols$eta_fu,
  phi_u = IEATools::template_cols$phi_u,
  matnames = IEATools::mat_meta_cols$matnames,
  matvals = IEATools::mat_meta_cols$matvals,
  rownames = IEATools::mat_meta_cols$rownames,
  colnames = IEATools::mat_meta_cols$colnames,
  rowtypes = IEATools::mat_meta_cols$rowtypes,
  coltypes = IEATools::mat_meta_cols$coltypes,
  sep = " -> "
)

Arguments

.eta_fu_table
a final-to-useful efficiency table read by load_eta_fu_allocation_data(). A template for this table should have been created by eta_fu_table() and write_eta_fu_table().

unit, year
See IEATools::iea_cols.

quantity, machine, eu_product, e_dot_machine, e_dot_machine_perc, maximum_values, eta_fu, phi_u
See IEATools::template_cols.

matnames, matvals, rownames, colnames, rowtypes, coltypes
See IEATools::mat_meta_cols.

sep
The string separator between prefix and suffix of compound row and column names. Default is " -> ". The default value matches the default value for the sep argument of matsbyname::vectorize_byname(), because matsbyname::vectorize_byname() will be used for further manipulations.

Value

a tidy data frame with metadata columns (and year) along with matnames and matvals columns indicating and containing eta_fu and phi_u vectors, respectively.

Examples

load_eta_fu_data() %>%
  form_eta_fu_phi_u_vecs()

#>   Country Method Energy.type Last.stage matnames Year
#> 1     GHA    PCM           E      Final   eta.fu 1971
#> 2     GHA    PCM           E      Final   eta.fu 2000
#> 3     GHA    PCM           E      Final    phi.u 1971
#> 4     GHA    PCM           E      Final    phi.u 2000
#> 5     ZAF    PCM           E      Final   eta.fu 1971
#> 6     ZAF    PCM           E      Final   eta.fu 2000
#> 7     ZAF    PCM           E      Final    phi.u 1971
#> 8     ZAF    PCM           E      Final    phi.u 2000
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  matvals
#> 1                                                                                                                                                                                                                                                                                                                                                                                           0.30, 0.30, 0.14, 0.30, 0.30, 0.30, 0.85, 0.97, 0.20, 0.85, 0.70, 0.65, 0.10, 0.85, 0.85, 0.85, 0.16, 0.18, 0.30, 0.12, 0.85
#> 2                                                                                                                                                                                                                                                                                                                                                                                                 0.31, 0.31, 0.14, 0.30, 0.31, 0.30, 0.99, 0.20, 0.95, 0.70, 0.12, 0.86, 0.86, 0.85, 0.16, 0.18, 0.48, 0.31, 0.12, 0.86
#> 3                                                                                                                                                                                                                                                                                  1.0000000, 1.0000000, 0.2009916, 1.0000000, 1.0000000, 1.0000000, 0.6585352, 0.3698616, 0.1261201, 1.0000000, 0.2009916, 0.3698616, 1.0000000, 0.6585352, 0.3698616, 0.3698616, 0.2009916, 0.2009916, 1.0000000, 0.2009916, 0.3698616
#> 4                                                                                                                                                                                                                                                                                             1.0000000, 1.0000000, 0.2009916, 1.0000000, 1.0000000, 1.0000000, 0.3698616, 0.1830161, 1.0000000, 0.2009916, 1.0000000, 0.6585352, 0.3698616, 0.3698616, 0.2009916, 0.2009916, 0.3698616, 1.0000000, 0.2009916, 0.3698616
#> 5                                                                                                                                                                                                                                                                                                                               0.30, 0.30, 0.14, 0.16, 0.25, 0.30, 0.30, 0.45, 0.90, 0.40, 0.20, 0.87, 0.12, 0.80, 0.85, 0.40, 0.20, 0.85, 0.80, 0.30, 0.30, 0.14, 0.16, 0.30, 0.80, 0.40, 0.60, 0.30, 0.12, 0.85, 0.20
#> 6                                                                                                                                                                                                                                                                   0.31, 0.31, 0.31, 0.14, 0.16, 0.16, 0.31, 0.31, 0.50, 0.95, 0.45, 0.20, 0.96, 0.12, 0.85, 0.90, 0.85, 0.90, 0.45, 0.20, 0.31, 0.90, 0.85, 0.31, 0.31, 0.35, 0.20, 0.30, 0.16, 0.14, 0.16, 0.32, 0.16, 0.35, 0.85, 0.90, 0.62, 0.31, 0.12, 0.85, 0.25
#> 7                                                                                                                                                 1.00000000, 1.00000000, 0.20099156, 0.01677008, 0.36986157, 1.00000000, 1.00000000, 1.00000000, 0.65853519, 0.36986157, 0.12612006, 1.00000000, 1.00000000, 0.65853519, 0.01677008, 0.36986157, 0.12612006, 0.65853519, 0.36986157, 1.00000000, 0.36986157, 0.36986157, 0.36986157, 0.65853519, 0.36986157, 1.00000000, 1.00000000, 0.20099156, 0.01677008, 0.36986157
#> 8 1.00000000, 1.00000000, 1.00000000, 0.20099156, 0.01677008, 0.20099156, 1.00000000, 1.00000000, 1.00000000, 0.65853519, 0.36986157, 0.18301611, 1.00000000, 1.00000000, 0.65853519, 0.01677008, 0.36986157, 0.65853519, 0.20099156, 0.36986157, 0.18301611, 1.00000000, 0.65853519, 0.36986157, 1.00000000, 1.00000000, 0.36986157, 0.18301611, 0.01677008, 0.20099156, 0.36986157, 0.20099156, 0.01677008, 0.20099156, 0.36986157, 0.01677008, 0.36986157, 1.00000000, 1.00000000, 0.20099156, 0.01677008, 0.36986157