Gather main activity producer and autoproducer industries

The IEA extended energy balances include both main activity producer and autoproducer industries for electricity, heat, and CHP plants. See details for an explication of each. This function gathers main activity producer and autoproducer, for each of the three types of plants: electricity, heat, and CHP plants. This function is called within the specify_all() function.

gather_producer_autoproducer(
  .tidy_iea_df,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  negzeropos = ".negzeropos",
  autoproducer_elect = IEATools::main_act_plants$autoprod_elect_plants,
  autoproducer_chp = IEATools::transformation_processes$autoproducer_CHP_plants,
  autoproducer_heat = IEATools::transformation_processes$autoproducer_heat_plants,
  main_act_producer_elect = IEATools::main_act_plants$main_act_prod_elect_plants,
  main_act_producer_heat = IEATools::main_act_plants$main_act_prod_heat_plants,
  main_act_producer_chp = IEATools::main_act_plants$main_act_prod_chp_plants
)

Arguments

.tidy_iea_df
The .tidy_iea_df which flows need to be specified.

flow_aggregation_point
The name of the flow aggregation point column in the .tidy_iea_df. Default is IEATools::iea_cols$flow_aggregation_point.

flow
The name of the flow column in the .tidy_iea_df. Default is IEATools::iea_cols$flow.

e_dot
The name of the energy column in the .tidy_iea_df. Default is IEATools::iea_cols$flow.

transformation_processes
A string identifying transformation processes in the flow_aggregation_point column of the .tidy_iea_df Default is IEATools::aggregation_flows$flow_aggregation_point.

negzeropos
The name of a temporary column created in .tidy_iea_df. Default is ".negzeropos".

autoproducer_elect
A string identifying "Autoproducer electricity plants" in the flow column of the .tidy_iea_df. Default is IEATools::main_act_plants$autoprod_elect_plants.

autoproducer_chp
A string identifying "Autoproducer CHP plants" in the flow column of the .tidy_iea_df. Default is IEATools::transformation_processes$autoproducer_CHP_plants.

autoproducer_heat
A string identifying "Autoproducer heat plants" in the flow column of the .tidy_iea_df. Default is IEATools::transformation_processes$autoproducer_heat_plants.

main_act_producer_elect
A string identifying "Main activity producer electricity plants" in the flow column of the .tidy_iea_df. Default is IEATools::main_act_plants$main_act_prod_elect_plants.

main_act_producer_heat
A string identifying "Main activity producer heat plants" in the flow column of the .tidy_iea_df. Default is IEATools::main_act_plants$main_act_prod_heat_plants.

main_act_producer_chp
A string identifying "Main activity producer CHP plants" in the flow column of the .tidy_iea_df. Default is IEATools::main_act_plants$main_act_prod_chp_plants.

Value

The tidy_iea_df with autoproducer plants merged with main activity producer plants.

Details

Autoproducer plants are those that consume in-situ the energy they produce. For instance, an iron and steel plant that produces electricity and directly consumes it would be classified as an autoproducer electricity plant. Conversely, main activity producer plants are those that produce a product, be it electricity, heat, or both (CHP plants) and sell it to the market.

Examples

library(dplyr)
load_tidy_iea_df() %>% 
  gather_producer_autoproducer()

#> # A tibble: 397 x 11
#>    Country Method Energy.type Last.stage  Year Ledger.side Flow.aggregation.poi…
#>    <chr>   <chr>  <chr>       <chr>      <dbl> <chr>       <chr>                
#>  1 GHA     PCM    E           Final       1971 Consumption Industry             
#>  2 GHA     PCM    E           Final       1971 Consumption Industry             
#>  3 GHA     PCM    E           Final       1971 Consumption Industry             
#>  4 GHA     PCM    E           Final       1971 Consumption Industry             
#>  5 GHA     PCM    E           Final       1971 Consumption Industry             
#>  6 GHA     PCM    E           Final       1971 Consumption Industry             
#>  7 GHA     PCM    E           Final       1971 Consumption Industry             
#>  8 GHA     PCM    E           Final       1971 Consumption Non-energy use       
#>  9 GHA     PCM    E           Final       1971 Consumption Other                
#> 10 GHA     PCM    E           Final       1971 Consumption Other                
#> # … with 387 more rows, and 4 more variables: Flow <chr>, Product <chr>,
#> #   Unit <chr>, E.dot <dbl>