Industry flows

A string vector containing names of Industry Flows in the IEA extended energy balances database.

Usage

industry_flows

Format

A string vector with 20 entries.

construction

The string identifier for the construction industry.

iron_and_steel

The string identifier for the iron and steel industry.

chemical_and_petrochemical

The string identifier for the chemical and petrochemical industry.

non_ferrous_metals

The string identifier for the non-ferrous metals industry.

non_metallic_minerals

The string identifier for the non-metallic minerals industry.

transport_equipment

The string identifier for the transport equipment industry.

machinery

The string identifier for the machinery industry.

food_and_tobacco

The string identifier for the food and tobacco industry.

paper_pulp_and_print

The string identifier for the paper, pulp and print industry. (For 2018 and earlier versions of the IEA extended energy balance database.)

paper_pulp_and_printing

The string identifier for the paper, pulp and printing industry. (For 2019 and later versions of the IEA extended energy balance database.)

wood_and_wood_products

The string identifier for the wood and wood products industry.

textile_and_leather

The string identifier for the textile and leather industry.

non_specified_industry

The string identifier for non-specified (industry).

industry_not_elsewhere_specified

The string identifier for industry not elsewhere specified.

coal_mines

The string identifier for coal mines. This industry arises after specifying the dataset.

oil_and_gas_extraction

The string identifier for oil and gas extraction. This industry arises after specifying the dataset.

Examples

industry_flows
#> $mining_and_quarrying
#> [1] "Mining and quarrying"
#> 
#> $construction
#> [1] "Construction"
#> 
#> $manufacturing
#> [1] "Manufacturing"
#> 
#> $iron_and_steel
#> [1] "Iron and steel"
#> 
#> $chemical_and_petrochemical
#> [1] "Chemical and petrochemical"
#> 
#> $non_ferrous_metals
#> [1] "Non-ferrous metals"
#> 
#> $non_metallic_minerals
#> [1] "Non-metallic minerals"
#> 
#> $transport_equipment
#> [1] "Transport equipment"
#> 
#> $machinery
#> [1] "Machinery"
#> 
#> $food_and_tobacco
#> [1] "Food and tobacco"
#> 
#> $paper_pulp_and_print
#> [1] "Paper, pulp and print"
#> 
#> $paper_pulp_and_printing
#> [1] "Paper, pulp and printing"
#> 
#> $wood_and_wood_products
#> [1] "Wood and wood products"
#> 
#> $textile_and_leather
#> [1] "Textile and leather"
#> 
#> $non_specified_industry
#> [1] "Non-specified (industry)"
#> 
#> $industry_not_elsewhere_specified
#> [1] "Industry not elsewhere specified"
#> 
#> $coal_mines
#> [1] "Coal mines"
#> 
#> $oil_and_gas_extraction
#> [1] "Oil and gas extraction"
#> 
#> $oil_extraction
#> [1] "Oil extraction"
#> 
#> $natural_gas_extraction
#> [1] "Natural gas extraction"
#>