Load final-to-useful machine efficiency data

When performing extending an energy conversion chain from useful energy to final energy, efficiencies of final-to-useful energy conversion machines are defined by the analyst. The Excel file at path contains those allocations.

load_eta_fu_data(
  path = file.path("extdata", "GH-ZA-Efficiency-sample.xlsx") %>% system.file(package
    = "IEATools"),
  eta_fu_tab_name = "FU etas"
)

Arguments

path
the path from which final-to-useful machine efficiency data will be loaded. Default is the path to sample efficiency data supplied with this package.

eta_fu_tab_name
the tab in path that contains the final-to-useful machine efficiency data. Default is "FU etas".

Value

the eta_fu_tab_name tab in path as a data frame.

Details

A final-to-useful machine efficiencies template can be generated using eta_fu_template() and write_eta_fu_template().

A filled example can be loaded with the default value of path.

Examples

load_eta_fu_data()

#>     Country Method Energy.type Last.stage Unit                 Machine
#> 1       GHA    PCM           E      Final ktoe             Automobiles
#> 2       GHA    PCM           E      Final ktoe             Automobiles
#> 3       GHA    PCM           E      Final ktoe             Diesel cars
#> 4       GHA    PCM           E      Final ktoe             Diesel cars
#> 5       GHA    PCM           E      Final ktoe           Diesel trucks
#> 6       GHA    PCM           E      Final ktoe           Diesel trucks
#> 7       GHA    PCM           E      Final ktoe         Electric motors
#> 8       GHA    PCM           E      Final ktoe         Electric motors
#> 9       GHA    PCM           E      Final ktoe                  Trucks
#> 10      GHA    PCM           E      Final ktoe                  Trucks
#> 11      GHA    PCM           E      Final ktoe            Boat engines
#> 12      GHA    PCM           E      Final ktoe            Boat engines
#> 13      GHA    PCM           E      Final ktoe           Diesel trains
#> 14      GHA    PCM           E      Final ktoe           Diesel trains
#> 15      GHA    PCM           E      Final ktoe                    Fans
#> 16      GHA    PCM           E      Final ktoe                    Fans
#> 17      GHA    PCM           E      Final ktoe         Electric lights
#> 18      GHA    PCM           E      Final ktoe         Electric lights
#> 19      GHA    PCM           E      Final ktoe        Electric heaters
#> 20      GHA    PCM           E      Final ktoe        Electric heaters
#> 21      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 22      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 23      GHA    PCM           E      Final ktoe             Wood stoves
#> 24      GHA    PCM           E      Final ktoe             Wood stoves
#> 25      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 26      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 27      GHA    PCM           E      Final ktoe        Electric heaters
#> 28      GHA    PCM           E      Final ktoe        Electric heaters
#> 29      GHA    PCM           E      Final ktoe                   Irons
#> 30      GHA    PCM           E      Final ktoe                   Irons
#> 31      GHA    PCM           E      Final ktoe              LPG stoves
#> 32      GHA    PCM           E      Final ktoe              LPG stoves
#> 33      GHA    PCM           E      Final ktoe         Electric stoves
#> 34      GHA    PCM           E      Final ktoe         Electric stoves
#> 35      GHA    PCM           E      Final ktoe             Wood stoves
#> 36      GHA    PCM           E      Final ktoe             Wood stoves
#> 37      GHA    PCM           E      Final ktoe         Charcoal stoves
#> 38      GHA    PCM           E      Final ktoe         Charcoal stoves
#> 39      GHA    PCM           E      Final ktoe         Kerosene stoves
#> 40      GHA    PCM           E      Final ktoe         Kerosene stoves
#> 41      GHA    PCM           E      Final ktoe              LPG stoves
#> 42      GHA    PCM           E      Final ktoe              LPG stoves
#> 43      GHA    PCM           E      Final ktoe         Electric stoves
#> 44      GHA    PCM           E      Final ktoe         Electric stoves
#> 45      ZAF    PCM           E      Final ktoe  Steam engines (trains)
#> 46      ZAF    PCM           E      Final ktoe  Steam engines (trains)
#> 47      ZAF    PCM           E      Final ktoe         Electric motors
#> 48      ZAF    PCM           E      Final ktoe         Electric motors
#> 49      ZAF    PCM           E      Final ktoe                  Trucks
#> 50      ZAF    PCM           E      Final ktoe                  Trucks
#> 51      ZAF    PCM           E      Final ktoe             Automobiles
#> 52      ZAF    PCM           E      Final ktoe             Automobiles
#> 53      ZAF    PCM           E      Final ktoe  Industry static motors
#> 54      ZAF    PCM           E      Final ktoe  Industry static motors
#> 55      ZAF    PCM           E      Final ktoe           Diesel trucks
#> 56      ZAF    PCM           E      Final ktoe           Diesel trucks
#> 57      ZAF    PCM           E      Final ktoe      Diesel automobiles
#> 58      ZAF    PCM           E      Final ktoe      Diesel automobiles
#> 59      ZAF    PCM           E      Final ktoe                    Fans
#> 60      ZAF    PCM           E      Final ktoe                    Fans
#> 61      ZAF    PCM           E      Final ktoe               Airplanes
#> 62      ZAF    PCM           E      Final ktoe               Airplanes
#> 63      ZAF    PCM           E      Final ktoe Industry static engines
#> 64      ZAF    PCM           E      Final ktoe Industry static engines
#> 65      ZAF    PCM           E      Final ktoe    Electric automobiles
#> 66      ZAF    PCM           E      Final ktoe    Electric automobiles
#> 67      ZAF    PCM           E      Final ktoe            Boat engines
#> 68      ZAF    PCM           E      Final ktoe            Boat engines
#> 69      ZAF    PCM           E      Final ktoe              Generators
#> 70      ZAF    PCM           E      Final ktoe              Generators
#> 71      ZAF    PCM           E      Final ktoe         Electric lights
#> 72      ZAF    PCM           E      Final ktoe         Electric lights
#> 73      ZAF    PCM           E      Final ktoe         Kerosene lights
#> 74      ZAF    PCM           E      Final ktoe         Kerosene lights
#> 75      ZAF    PCM           E      Final ktoe              Gas lights
#> 76      ZAF    PCM           E      Final ktoe              Gas lights
#> 77      ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 78      ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 79      ZAF    PCM           E      Final ktoe        Electric heaters
#> 80      ZAF    PCM           E      Final ktoe        Electric heaters
#> 81      ZAF    PCM           E      Final ktoe             Gas heaters
#> 82      ZAF    PCM           E      Final ktoe             Gas heaters
#> 83      ZAF    PCM           E      Final ktoe             Oil heaters
#> 84      ZAF    PCM           E      Final ktoe             Oil heaters
#> 85      ZAF    PCM           E      Final ktoe            Gas furnaces
#> 86      ZAF    PCM           E      Final ktoe            Gas furnaces
#> 87      ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 88      ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 89      ZAF    PCM           E      Final ktoe             Wood stoves
#> 90      ZAF    PCM           E      Final ktoe             Wood stoves
#> 91      ZAF    PCM           E      Final ktoe                   Irons
#> 92      ZAF    PCM           E      Final ktoe                   Irons
#> 93      ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 94      ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 95      ZAF    PCM           E      Final ktoe             Oil heaters
#> 96      ZAF    PCM           E      Final ktoe             Oil heaters
#> 97      ZAF    PCM           E      Final ktoe             Gas heaters
#> 98      ZAF    PCM           E      Final ktoe             Gas heaters
#> 99      ZAF    PCM           E      Final ktoe        Electric heaters
#> 100     ZAF    PCM           E      Final ktoe        Electric heaters
#> 101     ZAF    PCM           E      Final ktoe              LPG stoves
#> 102     ZAF    PCM           E      Final ktoe              LPG stoves
#> 103     ZAF    PCM           E      Final ktoe             Coal stoves
#> 104     ZAF    PCM           E      Final ktoe             Coal stoves
#> 105     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 106     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 107     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 108     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 109     ZAF    PCM           E      Final ktoe         Wood cookstoves
#> 110     ZAF    PCM           E      Final ktoe         Wood cookstoves
#> 111     ZAF    PCM           E      Final ktoe         Charcoal stoves
#> 112     ZAF    PCM           E      Final ktoe         Charcoal stoves
#> 113     ZAF    PCM           E      Final ktoe          LPG cookstoves
#> 114     ZAF    PCM           E      Final ktoe          LPG cookstoves
#> 115     ZAF    PCM           E      Final ktoe              LPG stoves
#> 116     ZAF    PCM           E      Final ktoe              LPG stoves
#> 117     ZAF    PCM           E      Final ktoe             Coal stoves
#> 118     ZAF    PCM           E      Final ktoe             Coal stoves
#> 119     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 120     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 121     ZAF    PCM           E      Final ktoe             Gas heaters
#> 122     ZAF    PCM           E      Final ktoe             Gas heaters
#> 123     ZAF    PCM           E      Final ktoe             Coal stoves
#> 124     ZAF    PCM           E      Final ktoe             Coal stoves
#> 125     ZAF    PCM           E      Final ktoe           Wood furnaces
#> 126     ZAF    PCM           E      Final ktoe           Wood furnaces
#> 127     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 128     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 129     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 130     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 131     ZAF    PCM           E      Final ktoe              LPG stoves
#> 132     ZAF    PCM           E      Final ktoe              LPG stoves
#> 133     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 134     ZAF    PCM           E      Final ktoe            Oil furnaces
#>     Eu.product E.dot_machine_max.[%] Quantity                1971
#> 1           MD          1.154061e-01   eta.fu                 0.3
#> 2           MD          1.154061e-01    phi.u                   1
#> 3           MD          3.492921e-02   eta.fu                 0.3
#> 4           MD          3.492921e-02    phi.u                   1
#> 5           MD          3.492921e-02   eta.fu                 0.3
#> 6           MD          3.492921e-02    phi.u                   1
#> 7           MD          3.483607e-02   eta.fu                0.85
#> 8           MD          3.483607e-02    phi.u                   1
#> 9           MD          1.313338e-02   eta.fu                 0.3
#> 10          MD          1.313338e-02    phi.u                   1
#> 11          MD          1.173621e-02   eta.fu                 0.3
#> 12          MD          1.173621e-02    phi.u                   1
#> 13          MD          5.798222e-03   eta.fu                 0.3
#> 14          MD          5.798222e-03    phi.u                   1
#> 15          MD          3.548808e-03   eta.fu                 0.1
#> 16          MD          3.548808e-03    phi.u                   1
#> 17       Light          2.959203e-02   eta.fu                <NA>
#> 18       Light          2.959203e-02    phi.u                <NA>
#> 19   HTH.600.C          6.242752e-02   eta.fu                0.85
#> 20   HTH.600.C          6.242752e-02    phi.u   0.658535188684648
#> 21   HTH.600.C          7.730963e-03   eta.fu                0.85
#> 22   HTH.600.C          7.730963e-03    phi.u   0.658535188684648
#> 23   MTH.200.C          1.276080e-01   eta.fu                0.85
#> 24   MTH.200.C          1.276080e-01    phi.u   0.369861566099546
#> 25   MTH.200.C          5.257055e-02   eta.fu                0.85
#> 26   MTH.200.C          5.257055e-02    phi.u   0.369861566099546
#> 27   MTH.200.C          2.757079e-02   eta.fu                0.97
#> 28   MTH.200.C          2.757079e-02    phi.u   0.369861566099546
#> 29   MTH.200.C          3.548808e-03   eta.fu                0.85
#> 30   MTH.200.C          3.548808e-03    phi.u   0.369861566099546
#> 31   MTH.200.C          9.314456e-04   eta.fu                0.45
#> 32   MTH.200.C          9.314456e-04    phi.u   0.369861566099546
#> 33   MTH.200.C          3.865481e-04   eta.fu                0.65
#> 34   MTH.200.C          3.865481e-04    phi.u   0.369861566099546
#> 35   MTH.100.C          5.659065e-01   eta.fu                0.12
#> 36   MTH.100.C          5.659065e-01    phi.u   0.200991558354549
#> 37   MTH.100.C          9.612519e-02   eta.fu 0.14000000000000001
#> 38   MTH.100.C          9.612519e-02    phi.u   0.200991558354549
#> 39   MTH.100.C          3.247004e-02   eta.fu                0.16
#> 40   MTH.100.C          3.247004e-02    phi.u   0.200991558354549
#> 41   MTH.100.C          6.706408e-03   eta.fu                0.18
#> 42   MTH.100.C          6.706408e-03    phi.u   0.200991558354549
#> 43   MTH.100.C          3.548808e-03   eta.fu                 0.7
#> 44   MTH.100.C          3.548808e-03    phi.u   0.200991558354549
#> 45          MD          9.923227e-02   eta.fu                 0.6
#> 46          MD          9.923227e-02    phi.u                   1
#> 47          MD          9.640099e-02   eta.fu                0.87
#> 48          MD          9.640099e-02    phi.u                   1
#> 49          MD          7.668197e-02   eta.fu                 0.3
#> 50          MD          7.668197e-02    phi.u                   1
#> 51          MD          7.518937e-02   eta.fu                 0.3
#> 52          MD          7.518937e-02    phi.u                   1
#> 53          MD          4.658756e-02   eta.fu                 0.3
#> 54          MD          4.658756e-02    phi.u                   1
#> 55          MD          3.123251e-02   eta.fu                 0.3
#> 56          MD          3.123251e-02    phi.u                   1
#> 57          MD          3.121385e-02   eta.fu                 0.3
#> 58          MD          3.121385e-02    phi.u                   1
#> 59          MD          1.380275e-02   eta.fu                0.12
#> 60          MD          1.380275e-02    phi.u                   1
#> 61          MD          1.145565e-02   eta.fu                 0.3
#> 62          MD          1.145565e-02    phi.u                   1
#> 63          MD          9.683197e-03   eta.fu                 0.3
#> 64          MD          9.683197e-03    phi.u                   1
#> 65          MD          3.022501e-03   eta.fu                0.45
#> 66          MD          3.022501e-03    phi.u                   1
#> 67          MD          1.921713e-03   eta.fu                 0.3
#> 68          MD          1.921713e-03    phi.u                   1
#> 69          MD          6.343520e-04   eta.fu                 0.3
#> 70          MD          6.343520e-04    phi.u                   1
#> 71       Light          1.168588e-01   eta.fu                <NA>
#> 72       Light          1.168588e-01    phi.u                <NA>
#> 73       Light          2.779954e-03   eta.fu                <NA>
#> 74       Light          2.779954e-03    phi.u                <NA>
#> 75       Light          2.154931e-03   eta.fu                <NA>
#> 76       Light          2.154931e-03    phi.u                <NA>
#> 77   HTH.600.C          2.155050e-01   eta.fu                0.85
#> 78   HTH.600.C          2.155050e-01    phi.u   0.658535188684648
#> 79   HTH.600.C          3.141908e-02   eta.fu                 0.9
#> 80   HTH.600.C          3.141908e-02    phi.u   0.658535188684648
#> 81   HTH.600.C          1.962760e-02   eta.fu                 0.8
#> 82   HTH.600.C          1.962760e-02    phi.u   0.658535188684648
#> 83   HTH.600.C          1.746199e-02   eta.fu                 0.8
#> 84   HTH.600.C          1.746199e-02    phi.u   0.658535188684648
#> 85   HTH.600.C          3.171760e-03   eta.fu                 0.8
#> 86   HTH.600.C          3.171760e-03    phi.u   0.658535188684648
#> 87   MTH.200.C          6.451733e-02   eta.fu                 0.8
#> 88   MTH.200.C          6.451733e-02    phi.u   0.369861566099546
#> 89   MTH.200.C          2.826598e-02   eta.fu                 0.2
#> 90   MTH.200.C          2.826598e-02    phi.u   0.369861566099546
#> 91   MTH.200.C          2.300459e-02   eta.fu                 0.3
#> 92   MTH.200.C          2.300459e-02    phi.u   0.369861566099546
#> 93   MTH.200.C          1.685985e-02   eta.fu 0.14000000000000001
#> 94   MTH.200.C          1.685985e-02    phi.u   0.369861566099546
#> 95   MTH.200.C          1.577600e-02   eta.fu                 0.4
#> 96   MTH.200.C          1.577600e-02    phi.u   0.369861566099546
#> 97   MTH.200.C          1.070935e-02   eta.fu                 0.4
#> 98   MTH.200.C          1.070935e-02    phi.u   0.369861566099546
#> 99   MTH.200.C          8.916377e-03   eta.fu                 0.4
#> 100  MTH.200.C          8.916377e-03    phi.u   0.369861566099546
#> 101  MTH.200.C          3.612826e-03   eta.fu                 0.3
#> 102  MTH.200.C          3.612826e-03    phi.u   0.369861566099546
#> 103  MTH.200.C          2.318230e-03   eta.fu                0.25
#> 104  MTH.200.C          2.318230e-03    phi.u   0.369861566099546
#> 105  MTH.200.C          2.052315e-04   eta.fu                0.85
#> 106  MTH.200.C          2.052315e-04    phi.u   0.369861566099546
#> 107  MTH.200.C          5.597224e-05   eta.fu                0.85
#> 108  MTH.200.C          5.597224e-05    phi.u   0.369861566099546
#> 109  MTH.100.C          6.093511e-02   eta.fu                0.12
#> 110  MTH.100.C          6.093511e-02    phi.u   0.200991558354549
#> 111  MTH.100.C          1.513116e-02   eta.fu 0.14000000000000001
#> 112  MTH.100.C          1.513116e-02    phi.u   0.200991558354549
#> 113  MTH.100.C          2.705325e-03   eta.fu                0.16
#> 114  MTH.100.C          2.705325e-03    phi.u   0.200991558354549
#> 115  MTH.100.C          8.768984e-04   eta.fu                0.16
#> 116  MTH.100.C          8.768984e-04    phi.u   0.200991558354549
#> 117  MTH.100.C          8.209262e-04   eta.fu                0.16
#> 118  MTH.100.C          8.209262e-04    phi.u   0.200991558354549
#> 119  MTH.100.C          5.597224e-05   eta.fu                0.16
#> 120  MTH.100.C          5.597224e-05    phi.u   0.200991558354549
#> 121  MTH.100.C          9.328706e-06   eta.fu                0.85
#> 122  MTH.100.C          9.328706e-06    phi.u   0.200991558354549
#> 123   LTH.20.C          9.453560e-02   eta.fu                0.16
#> 124   LTH.20.C          9.453560e-02    phi.u  0.0167700821734027
#> 125   LTH.20.C          6.093511e-02   eta.fu                0.85
#> 126   LTH.20.C          6.093511e-02    phi.u  0.0167700821734027
#> 127   LTH.20.C          2.336294e-02   eta.fu                0.85
#> 128   LTH.20.C          2.336294e-02    phi.u  0.0167700821734027
#> 129   LTH.20.C          1.029889e-02   eta.fu                 0.3
#> 130   LTH.20.C          1.029889e-02    phi.u  0.0167700821734027
#> 131   LTH.20.C          2.705325e-03   eta.fu                0.32
#> 132   LTH.20.C          2.705325e-03    phi.u  0.0167700821734027
#> 133   LTH.20.C          2.612038e-04   eta.fu                0.85
#> 134   LTH.20.C          2.612038e-04    phi.u  0.0167700821734027
#>                    2000
#> 1                  0.31
#> 2                     1
#> 3                  0.31
#> 4                     1
#> 5                  0.31
#> 6                     1
#> 7                  0.95
#> 8                     1
#> 9                  0.31
#> 10                    1
#> 11                 0.31
#> 12                    1
#> 13                 0.31
#> 14                    1
#> 15                 0.12
#> 16                    1
#> 17                 <NA>
#> 18                 <NA>
#> 19                 0.86
#> 20    0.658535188684648
#> 21                 0.86
#> 22    0.658535188684648
#> 23                 0.86
#> 24    0.369861566099546
#> 25                 0.86
#> 26    0.369861566099546
#> 27                 0.99
#> 28    0.369861566099546
#> 29                 0.85
#> 30    0.369861566099546
#> 31                 0.48
#> 32    0.369861566099546
#> 33                 0.67
#> 34    0.369861566099546
#> 35                 0.12
#> 36    0.200991558354549
#> 37  0.14000000000000001
#> 38    0.200991558354549
#> 39                 0.16
#> 40    0.200991558354549
#> 41                 0.18
#> 42    0.200991558354549
#> 43                  0.7
#> 44    0.200991558354549
#> 45                 0.62
#> 46                    1
#> 47                 0.96
#> 48                    1
#> 49                 0.31
#> 50                    1
#> 51                 0.31
#> 52                    1
#> 53                 0.31
#> 54                    1
#> 55                 0.31
#> 56                    1
#> 57                 0.31
#> 58                    1
#> 59                 0.12
#> 60                    1
#> 61                 0.31
#> 62                    1
#> 63                 0.31
#> 64                    1
#> 65                  0.5
#> 66                    1
#> 67                 0.31
#> 68                    1
#> 69                 0.31
#> 70                    1
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                  0.9
#> 78    0.658535188684648
#> 79                 0.95
#> 80    0.658535188684648
#> 81                 0.85
#> 82    0.658535188684648
#> 83                 0.85
#> 84    0.658535188684648
#> 85                 0.85
#> 86    0.658535188684648
#> 87                 0.85
#> 88    0.369861566099546
#> 89                 0.25
#> 90    0.369861566099546
#> 91                 0.35
#> 92    0.369861566099546
#> 93  0.14000000000000001
#> 94    0.369861566099546
#> 95                 0.45
#> 96    0.369861566099546
#> 97                 0.45
#> 98    0.369861566099546
#> 99                 0.45
#> 100   0.369861566099546
#> 101                0.35
#> 102   0.369861566099546
#> 103                 0.3
#> 104   0.369861566099546
#> 105                 0.9
#> 106   0.369861566099546
#> 107                 0.9
#> 108   0.369861566099546
#> 109                0.12
#> 110   0.200991558354549
#> 111 0.14000000000000001
#> 112   0.200991558354549
#> 113                0.16
#> 114   0.200991558354549
#> 115                0.16
#> 116   0.200991558354549
#> 117                0.16
#> 118   0.200991558354549
#> 119                0.16
#> 120   0.200991558354549
#> 121                 0.9
#> 122   0.200991558354549
#> 123                0.16
#> 124  0.0167700821734027
#> 125                0.86
#> 126  0.0167700821734027
#> 127                0.86
#> 128  0.0167700821734027
#> 129                 0.3
#> 130  0.0167700821734027
#> 131                0.32
#> 132  0.0167700821734027
#> 133                0.85
#> 134  0.0167700821734027