Oil and oil products

A string vector containing names of products classified by the IEA as oil and oil products.

Usage

oil_and_oil_products

Format

A string vector with 24 entries.

primary_oil_products

The string identifier for primary_oil_products.

refinery_feedstocks

The string identifier for Refinery feedstocks.

refinery_gas

The string identifier for Refinery gas.

ethane

The string identifier for Ethane.

liquefied_petroleum_gases_lpg

The string identifier for Liquefied petroleum gases (LPG).

motor_gasoline_excl_biofuels

The string identifier for Motor gasoline excl. biofuels.

aviation_gasoline

The string identifier for Aviation gasoline.

gasoline_type_jet_fuel

The string identifier for Gasoline type jet fuel.

kerosene_type_jet_fuel_excl_biofuels

The string identifier for Kerosene type jet fuel excl. biofuels.

other_kerosene

The string identifier for Other kerosene.

gas_diesel_oil_excl_biofuels

The string identifier for Gas/diesel oil excl. biofuels.

fuel_oil

The string identifier for Fuel oil.

naptha

The string identifier for Naphtha.

white_spirit_SBP

The string identifier for White spirit & SBP.

lubricants

The string identifier for Lubricants.

bitumen

The string identifier for Bitumen.

paraffin_waxes

The string identifier for Paraffin waxes.

petroleum_coke

The string identifier for Petroleum coke.

other_oil_products

The string identifier for Other oil products.

Examples

oil_and_oil_products
#> $crude_ngl_feedstocks_if_no_detail
#> [1] "Crude/NGL/feedstocks (if no detail)"
#> 
#> $crude_oil
#> [1] "Crude oil"
#> 
#> $natural_gas_liquids
#> [1] "Natural gas liquids"
#> 
#> $additives_blending_components
#> [1] "Additives/blending components"
#> 
#> $other_hydrocarbons
#> [1] "Other hydrocarbons"
#> 
#> $oil_shales_and_oil_sands
#> [1] "Oil shale and oil sands"
#> 
#> $refinery_feedstocks
#> [1] "Refinery feedstocks"
#> 
#> $refinery_gas
#> [1] "Refinery gas"
#> 
#> $ethane
#> [1] "Ethane"
#> 
#> $liquefied_petroleum_gases_lpg
#> [1] "Liquefied petroleum gases (LPG)"
#> 
#> $motor_gasoline_excl_biofuels
#> [1] "Motor gasoline excl. biofuels"
#> 
#> $aviation_gasoline
#> [1] "Aviation gasoline"
#> 
#> $gasoline_type_jet_fuel
#> [1] "Gasoline type jet fuel"
#> 
#> $kerosene_type_jet_fuel_excl_biofuels
#> [1] "Kerosene type jet fuel excl. biofuels"
#> 
#> $other_kerosene
#> [1] "Other kerosene"
#> 
#> $gas_diesel_oil_excl_biofuels
#> [1] "Gas/diesel oil excl. biofuels"
#> 
#> $fuel_oil
#> [1] "Fuel oil"
#> 
#> $naptha
#> [1] "Naphtha"
#> 
#> $white_spirit_SBP
#> [1] "White spirit & SBP"
#> 
#> $lubricants
#> [1] "Lubricants"
#> 
#> $bitumen
#> [1] "Bitumen"
#> 
#> $paraffin_waxes
#> [1] "Paraffin waxes"
#> 
#> $petroleum_coke
#> [1] "Petroleum coke"
#> 
#> $other_oil_products
#> [1] "Other oil products"
#>