Country code overrides

A data frame containing 3-letter (and 4-letter) country codes and IEA country names. These code-country pairs are used as a default set of overrides (relative to countrycode::codelist) in the function use_iso_countries(). We use 4-letter codes when we override 3-letter ISO codes.

Usage

override_iso_codes_df

Format

A data frame with five rows and two columns.

PFU.code

The column of 3-letter country codes that to override those found in countrycode::codelist.

IEA.name

The column containing IEA country names.

Examples

override_iso_codes_df
#> # A tibble: 7 × 2
#>   PFU.code IEA.name                          
#>   <chr>    <chr>                             
#> 1 WRLD     World                             
#> 2 CHNM     People's Republic of China        
#> 3 HKG      Hong Kong (China)                 
#> 4 SUN      Former Soviet Union (if no detail)
#> 5 WMBK     World marine bunkers              
#> 6 WABK     World aviation bunkers            
#> 7 YUG      Former Yugoslavia (if no detail)