Primary oil products

A string vector containing names of products classified by the IEA as primary oil products.

Usage

primary_oil_products

Format

A string vector with 6 entries.

crude_ngl_feedstocks_if_no_detail

The string identifier for Crude/NGL/feedstocks (if no detail).

crude_oil

The string identifier for Crude oil.

natural_gas_liquids

The string identifier for Natural gas liquids.

additives_blending_components

The string identifier for Additives/blending components.

other_hydrocarbons

The string identifier for Other hydrocarbons.

oil_shales_and_oil_sands

The string identifier for Oil shale and oil sands.

Examples

primary_oil_products
#> $crude_ngl_feedstocks_if_no_detail
#> [1] "Crude/NGL/feedstocks (if no detail)"
#> 
#> $crude_oil
#> [1] "Crude oil"
#> 
#> $natural_gas_liquids
#> [1] "Natural gas liquids"
#> 
#> $additives_blending_components
#> [1] "Additives/blending components"
#> 
#> $other_hydrocarbons
#> [1] "Other hydrocarbons"
#> 
#> $oil_shales_and_oil_sands
#> [1] "Oil shale and oil sands"
#>