Identify all energy types supplied or consumed by Production, Transformation processes, or Energy industry own use

Sometimes, it is helpful to know all types of energy supplied or consumed by Production, Transformation processes, or Energy industry own use. This function (optionally) reads an IEA data file or loads an IEA data frame and builds a named list of energy types supplied or consumed by Production, Transformation processes, or Energy industry own use.

Usage

prod_tp_eiou_energy_carriers(
  file_path = sample_iea_data_path(),
  iea_df = IEATools::load_tidy_iea_df(file_path),
  side = c(IEATools::ledger_sides$consumption, IEATools::ledger_sides$supply),
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  production = IEATools::tpes_flows$production,
  transformation_processes = IEATools::tfc_compare_flows$transformation_processes,
  eiou = IEATools::tfc_compare_flows$energy_industry_own_use,
  stage = c(production, transformation_processes, eiou),
  e_dot = IEATools::iea_cols$e_dot,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product
)

Arguments

file_path

The path to the IEA data file (optional).

iea_df

A data frame containing IEA data. Default is IEATools::load_tidy_iea_df(file_path).

side

Refers to the "Consumption" or "Supply" side of Production, Transformation processes, or Energy industry own use. One of "Consumption" or "Supply". Default is "Consumption".

flow_aggregation_point

The flow aggregation point column in iea_df. Default is FlowAggregationPoint.

production

The string indicating the production flow. Default is "Production".

transformation_processes

The string indicating the transformation process stage. Default is "Transformation processes".

eiou

The string indicating the energy industry own use flow. Default is "Energy industry own use".

stage

The string indicating the stage for the analysis. One of production, transformation_processes, or eiou. Default is production.

e_dot

The energy flow rate column in iea_df. Default is "Edot".

flow

The flow column in iea_df. Default is "Flow".

product

The product column in iea_df. Default is "Product".

Value

a list of string vectors

Details

The names in the returned list are the Production, Transformation processes, or Energy industry own use industries in iea_df. The items in the returned list are vectors of energy types produced or consumed by the corresponding industries.

Examples

prod_tp_eiou_energy_carriers()
#> named list()
#> attr(,"ptype")
#> character(0)