PSUT matrix row and column types

PSUT matrices have row and column types. This list provides the typical names for the row and column types throughout the IEATools package.

Usage

row_col_types

Format

A string list with 6 entries.

resource

The type of entity that provides raw resources.

industry

The type of entity that receives inputs and makes outputs.

sector

The type of entity that absorbs final demand.

product

The inputs and outputs of industries.

unit

Units of physical measurement such as ktoe or TJ.

other

Used for, e.g., the 1-dimension of a vector.

Examples

row_col_types
#> $industry
#> [1] "Industry"
#> 
#> $resource
#> [1] "Industry"
#> 
#> $sector
#> [1] "Industry"
#> 
#> $product
#> [1] "Product"
#> 
#> $unit
#> [1] "Unit"
#> 
#> $other
#> [1] "Other"
#>