Specify all industries

This is a convenience function. It bundles several others:

specify_primary_production()

specify_tp_eiou()

specify_interface_industries()

tp_sinks_to_nonenergy()

Usage

specify_all(
  .tidy_iea_df,
  split_own_use_elect_chp_heat_using_shares_of = c("input", "output"),
  route_non_specified_eiou = TRUE,
  route_non_specified_tp = TRUE,
  specify_renewable_plants = FALSE,
  specify_electricity_grid = FALSE,
  specify_distribution_industries = FALSE,
  ascribe_eiou_to_renewable_plants = FALSE,
  ascribe_eiou_to_nuclear = FALSE
)

Arguments

.tidy_iea_df

A tidy data frame containing IEA extended energy balance data

split_own_use_elect_chp_heat_using_shares_of

Indicates whether the input or outputs to Main activity producer plants should be use for splitting the Own use in electricity, CHP and heat plants EIOU flow. Default is "input".

route_non_specified_eiou

Boolean stating whether non-specified EIOU flows should be routed to existing industries Default is TRUE.

route_non_specified_tp

Boolean stating whether non-specified transformation processes flows should be routed to existing industries Default is TRUE.

specify_renewable_plants

A boolean indicating whether renewable energy plants should be specified or not. Default is FALSE.

specify_electricity_grid

Boolean stating whether an electricity grid industry should be specified or not. Default is FALSE.

specify_distribution_industries

Boolean stating whether distribution industries should be specified or not. Default is FALSE.

ascribe_eiou_to_renewable_plants

A boolean defining whether a fraction of the EIOU of electricity, CHP and heat plants should be ascribed to the new renewable industries. Default is FALSE.

ascribe_eiou_to_nuclear

A boolean defining whether a fraction of the EIOU of electricity, CHP and heat plants should be ascribed to the new nuclear industry. Default is FALSE.

Value

An enhanced and corrected version of .tidy_iea_df That is ready for physical supply-use table (PSUT) analysis.

Details

Each bundled function is called in turn using default arguments. See examples for two ways to achieve the same result.

Examples

# Simple
load_tidy_iea_df() %>% 
  specify_all()
#> # A tibble: 429 × 11
#>    Country Method EnergyType LastStage  Year LedgerSide  FlowAggregationPoint
#>    <chr>   <chr>  <chr>      <chr>     <dbl> <chr>       <chr>               
#>  1 GHA     PCM    E          Final      1971 Consumption Industry            
#>  2 GHA     PCM    E          Final      1971 Consumption Industry            
#>  3 GHA     PCM    E          Final      1971 Consumption Industry            
#>  4 GHA     PCM    E          Final      1971 Consumption Industry            
#>  5 GHA     PCM    E          Final      1971 Consumption Industry            
#>  6 GHA     PCM    E          Final      1971 Consumption Industry            
#>  7 GHA     PCM    E          Final      1971 Consumption Industry            
#>  8 GHA     PCM    E          Final      1971 Consumption Other               
#>  9 GHA     PCM    E          Final      1971 Consumption Other               
#> 10 GHA     PCM    E          Final      1971 Consumption Other               
#> # ℹ 419 more rows
#> # ℹ 4 more variables: Flow <chr>, Product <chr>, Unit <chr>, Edot <dbl>
# Complicated
load_tidy_iea_df() %>% 
  specify_primary_production() %>% 
  specify_tp_eiou() %>% 
  specify_bunkers() %>% 
  specify_interface_industries() %>% 
  tp_sinks_to_nonenergy()
#> # A tibble: 429 × 11
#>    Country Method EnergyType LastStage  Year LedgerSide  FlowAggregationPoint
#>    <chr>   <chr>  <chr>      <chr>     <dbl> <chr>       <chr>               
#>  1 GHA     PCM    E          Final      1971 Consumption Industry            
#>  2 GHA     PCM    E          Final      1971 Consumption Industry            
#>  3 GHA     PCM    E          Final      1971 Consumption Industry            
#>  4 GHA     PCM    E          Final      1971 Consumption Industry            
#>  5 GHA     PCM    E          Final      1971 Consumption Industry            
#>  6 GHA     PCM    E          Final      1971 Consumption Industry            
#>  7 GHA     PCM    E          Final      1971 Consumption Industry            
#>  8 GHA     PCM    E          Final      1971 Consumption Other               
#>  9 GHA     PCM    E          Final      1971 Consumption Other               
#> 10 GHA     PCM    E          Final      1971 Consumption Other               
#> # ℹ 419 more rows
#> # ℹ 4 more variables: Flow <chr>, Product <chr>, Unit <chr>, Edot <dbl>