Change bunker flow names to reflect their purpose

The IEA extended energy balance data includes flows "International marine bunkers" and "International aviation bunkers" which are insufficiently specific for our purposes. These flows are actually exports to and imports at the regions "World marine bunkers" and "World aviation bunkers" (which we treat as a countries). This function changes the names of the flows to "Exports to World marine bunkers" and "Exports to World aviation bunkers".

Usage

specify_bunkers(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  imb = IEATools::tpes_flows$international_marine_bunkers,
  iab = IEATools::tpes_flows$international_aviation_bunkers,
  etwmb = IEATools::tpes_flows$exports_to_world_marine_bunkers,
  etwab = IEATools::tpes_flows$exports_to_world_aviation_bunkers,
  imports = IEATools::tpes_flows$imports,
  wmb_sector_long = IEATools::transport_flows$world_marine_bunkers,
  wab_sector_long = IEATools::transport_flows$world_aviation_bunkers,
  international_navigation = IEATools::transport_flows$international_navigation,
  international_aviation = IEATools::transport_flows$international_aviation,
  wmbk_3_letter = "WMBK",
  wabk_3_letter = "WABK"
)

Arguments

.tidy_iea_df

An IEA data frame whose columns have been renamed by rename_iea_df_cols()

country, flow, product

See IEATools::iea_cols$country.

imb, iab, etwmb, etwab, imports

See IEATools::tpes_flows.

wmb_sector_long, wab_sector_long, international_navigation, international_aviation

See IEATools::transport_flows.

wmbk_3_letter

The 3-letter code for the world marine bunkers country. Default is "WMB".

wabk_3_letter

The 3-letter code for the world aviation bunkers country. Default is "WAB".

Value

A .tidy_iea_df with bunker flows specified.

Examples

load_tidy_iea_df() %>% 
  specify_primary_production() %>% 
  specify_tp_eiou() %>% 
 specify_bunkers()   
#> # A tibble: 429 × 11
#>    Country Method EnergyType LastStage  Year LedgerSide  FlowAggregationPoint
#>    <chr>   <chr>  <chr>      <chr>     <dbl> <chr>       <chr>               
#>  1 GHA     PCM    E          Final      1971 Consumption Industry            
#>  2 GHA     PCM    E          Final      1971 Consumption Industry            
#>  3 GHA     PCM    E          Final      1971 Consumption Industry            
#>  4 GHA     PCM    E          Final      1971 Consumption Industry            
#>  5 GHA     PCM    E          Final      1971 Consumption Industry            
#>  6 GHA     PCM    E          Final      1971 Consumption Industry            
#>  7 GHA     PCM    E          Final      1971 Consumption Industry            
#>  8 GHA     PCM    E          Final      1971 Consumption Non-energy use      
#>  9 GHA     PCM    E          Final      1971 Consumption Other               
#> 10 GHA     PCM    E          Final      1971 Consumption Other               
#> # ℹ 419 more rows
#> # ℹ 4 more variables: Flow <chr>, Product <chr>, Unit <chr>, Edot <dbl>