Specifies distribution industries

For each product where losses are reported, creates a distribution industry which is specified by product (e.g., "Distribution [of Heat]"), and converts a product from a specific origin (e.g., "Heat [from Oil refineries]") into the final demand product (e.g., "Heat").

Usage

specify_distribution_losses(
  .tidy_iea_df,
  specify_distribution_industries = FALSE,
  supplying_industry_notation = RCLabels::from_notation,
  distribution_industry_notation = RCLabels::of_notation,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  e_dot = IEATools::iea_cols$e_dot,
  losses = IEATools::tfc_compare_flows$losses,
  distribution_industry = IEATools::distribution_industry,
  supply = IEATools::ledger_sides$supply,
  transformation_processes = IEATools::tfc_compare_flows$transformation_processes,
  negzeropos = ".negzeropos"
)

Arguments

.tidy_iea_df

The .tidy__iea_df for which an electricity grid industry should be added.

specify_distribution_industries

A boolean stating whether distribution industries should be added or not. Default is FALSE.

supplying_industry_notation

Notation to use to specify the supplying industries. Default is RCLabels::from_notation.

distribution_industry_notation

Notation to use to specify the distribution industries. Default is RCLabels::of_notation.

flow_aggregation_point, flow, e_dot, product, method, ledger_side, last_stage, energy_type, country, year, unit

See IEATools::iea_cols.

losses

The name of the "Losses" flows in the input data frame. Default is IEATools::tfc_compare_flows$losses.

distribution_industry

The name of the distribution industries to be added. Default is IEATools::distribution_industry.

supply

The name of the supply ledger side. Default is IEATools::ledger_sides$supply.

transformation_processes

The name of transformation processes in the flow aggregation point column. Default is IEATools::tfc_compare_flows$transformation_processes.

negzeropos

The name of a temporary column added to the data frame. Default is ".negzeropos".

Value

The .tidy__iea_df to which distribution industries have been added.

Examples

load_tidy_iea_df() %>% 
  specify_distribution_losses()
#> # A tibble: 403 × 11
#>    Country Method EnergyType LastStage  Year LedgerSide FlowAggregationPoint    
#>    <chr>   <chr>  <chr>      <chr>     <dbl> <chr>      <chr>                   
#>  1 GHA     PCM    E          Final      1971 Supply     Total primary energy su…
#>  2 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#>  3 GHA     PCM    E          Final      1971 Supply     Total primary energy su…
#>  4 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#>  5 GHA     PCM    E          Final      1971 Supply     Total primary energy su…
#>  6 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#>  7 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#>  8 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#>  9 GHA     PCM    E          Final      1971 Supply     Total primary energy su…
#> 10 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#> # ℹ 393 more rows
#> # ℹ 4 more variables: Flow <chr>, Product <chr>, Unit <chr>, Edot <dbl>