Specify primary production industries

The IEA extended energy balances include some Flows that identify Energy industry own use (EIOU), the consumption of energy by energy-producing industries. But some primary production industries that receive EIOU do not produce anything. For example, Coal mines receive electricity but there are no Coal mines that produce coal. Rather, the generic Production industry produces coal. This function solves that problem by replacing the generic Production industry with specific industries.

specify_primary_production(.tidy_iea_df,
  eiou_destinations = c("Coal mines", "Oil and gas extraction"),
  production_products = list(IEATools::coal_and_coal_products,
  c(IEATools::oil_and_oil_products, "Natural gas")),
  production_products_short_names = c("Coal", "Oil and natural gas"),
  flow_aggregation_point = "Flow.aggregation.point",
  eiou = "Energy industry own use",
  transformation_processes = "Transformation processes", flow = "Flow",
  eiou_suffix = "(energy)", resources = "Resources",
  production = "Production", e_dot = "E.dot", product = "Product")

Arguments

.tidy_iea_df
an IEA data frame whose columns have been renamed by rename_iea_df_cols()

eiou_destinations
a vector of destinations for EIOU for primary production of coal and coal products and oil and natural gas. Default is c("Coal mines", "Oil and gas extraction").

production_products
a list of products for which we want to specify primary industries. Default is list(coal_and_coal_products, c(oil_and_oil_products, "Natural gas")).

production_products_short_names
a vector of short names for primary industries. Default is c("Coal", "Oil and natural gas").

flow_aggregation_point
the name of the flow aggregation point column in .tidy_iea_df. Default is "Flow.aggregation.point".

eiou
a string identifying energy industry own use in the Flow.aggregation.point column of .tidy_iea_df. Default is "Energy industry own use".

transformation_processes
a string identifying transformation processes in the flow column of .tidy_iea_df. Default is "Transformation processes".

flow
the name of the flow column in .tidy_iea_df. Default is "Flow".

eiou_suffix
the string suffix identifying energy industry own use in the Flow column of .tidy_iea_df. Default is "(energy)".

resources
a string identifying resource industries to be added to .tidy_iea_df. Default is "Resources".

production
a string identifying production in the flow column. Default is "Production".

e_dot
the name of the energy column in .tidy_eia_df. Default is "E.dot".

product
the name of the product column in .tidy_iea_df. Default is "Product".

Value

.tidy_iea_df with adjusted production information for primary energy for both coal and coal products and oil and gas extraction

Details

By default, the following changes are made to .tidy_iea_df:

The Production industry for coal_and_coal_products is replaced by Coal mines.

A Resources (Coal) industry is created which becomes the source of all primary coal_and_coal_products in the energy conversion chain.

Flows from Resources (Coal) to Coal mines are added.

The Coal mines (energy) EIOU Flow is replaced by Coal mines.

The Production industry for oil_and_oil_products and Natural gas is replaced by Oil and gas extraction.

A Resources (Oil and natural gas) industry is created which becomes the source of all primary oil_and_oil_products and Natural gas in the energy conversion chain.

Flows from Resources (Oil and natural gas) to Oil and gas extraction are added.

The Oil and gas extraction (energy) EIOU Flow is replaced by Oil and gas extraction.

Users can specify other changes by adjusting the default argument values.

Examples

library(dplyr)
load_tidy_iea_df() %>%
  specify_primary_production() %>%
  add_psut_matnames() %>%
  filter(Flow %in% c("Resources (Coal)", "Coal mines")) %>%
  select(-Method, -Last.stage, -Ledger.side, -Unit)

#> # A tibble: 9 x 8
#>   Country Year  Flow.aggregation.… Energy.type Flow    Product     E.dot matname
#>   <chr>   <chr> <chr>              <chr>       <chr>   <chr>       <dbl> <chr>  
#> 1 ZA      1971  Total primary ene… E           Resour… Hard coa…   33064 R      
#> 2 ZA      2000  Total primary ene… E           Resour… Coking c…    2372 R      
#> 3 ZA      2000  Total primary ene… E           Resour… Other bi…  124554 R      
#> 4 ZA      1971  Transformation pr… E           Coal m… Hard coa…  -33064 U_excl…
#> 5 ZA      2000  Transformation pr… E           Coal m… Coking c…   -2372 U_excl…
#> 6 ZA      2000  Transformation pr… E           Coal m… Other bi… -124554 U_excl…
#> 7 ZA      1971  Transformation pr… E           Coal m… Hard coa…   33064 V      
#> 8 ZA      2000  Transformation pr… E           Coal m… Coking c…    2372 V      
#> 9 ZA      2000  Transformation pr… E           Coal m… Other bi…  124554 V