Specifies renewable electricity and heat

This function specifies hydro, geothermal, solar photovoltaic, solar thermal, oceanic, and wind power industries.

Usage

specify_renewable_plants(
  .tidy_iea_df,
  specify_renewable_plants = FALSE,
  ascribe_eiou_to_renewable_plants = FALSE,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  product = IEATools::iea_cols$product,
  method = IEATools::iea_cols$method,
  ledger_side = IEATools::iea_cols$ledger_side,
  last_stage = IEATools::iea_cols$last_stage,
  energy_type = IEATools::iea_cols$energy_type,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  unit = IEATools::iea_cols$unit,
  eiou = IEATools::aggregation_flows$energy_industry_own_use,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  main_act_producer_elect = IEATools::main_act_plants$main_act_prod_elect_plants,
  main_act_producer_chp = IEATools::main_act_plants$main_act_prod_chp_plants,
  main_act_producer_heat = IEATools::main_act_plants$main_act_prod_heat_plants,
  autoproducer_elect = IEATools::main_act_plants$autoprod_elect_plants,
  autoproducer_chp = IEATools::transformation_processes$autoproducer_CHP_plants,
  autoproducer_heat = IEATools::main_act_plants$autoprod_heat_plants,
  own_use_elect_chp_heat = IEATools::eiou_flows$own_use_elect_chp_heat_plants,
  geothermal = IEATools::renewable_products$geothermal,
  hydro = IEATools::renewable_products$hydro,
  solar_pv = IEATools::renewable_products$solar_photovoltaics,
  solar_th = IEATools::renewable_products$solar_thermal,
  oceanic = IEATools::renewable_products$tide_wave_and_ocean,
  wind = IEATools::renewable_products$wind,
  electricity = IEATools::electricity_products$electricity,
  heat = IEATools::heat_products$heat,
  ratio_solar_th_elec = 0.33,
  ratio_solar_th_heat = 1,
  ratio_geothermal_elec = 0.1,
  ratio_geothermal_heat = 0.5,
  ratio_other_renewable_elec = 1,
  geothermal_plants = IEATools::renewable_industries$geothermal_plants,
  hydro_plants = IEATools::renewable_industries$hydro_plants,
  solar_pv_plants = IEATools::renewable_industries$solar_pv_plants,
  solar_th_plants = IEATools::renewable_industries$solar_th_plants,
  oceanic_plants = IEATools::renewable_industries$oceanic_plants,
  wind_power_plants = IEATools::renewable_industries$wind_power_plants,
  negzeropos = ".negzeropos",
  ratio_elec_to_heat = ".ratio_elec_to_heat",
  .share_industry = ".share_industry",
  .share = ".share"
)

Arguments

.tidy_iea_df

The .tidy_iea_df which flows need to be specified.

specify_renewable_plants

A boolean indicating whether renewable energy plants should be specified or not. Default is FALSE.

ascribe_eiou_to_renewable_plants

A boolean defining whether a fraction of the EIOU of electricity, CHP and heat plants should be ascribed to the new renewable industries. Default is FALSE.

flow_aggregation_point, flow, e_dot, product, method, ledger_side, last_stage, energy_type, country, year, unit

See IEATools::iea_cols.

eiou

A string identifying the energy industry own use in the flow_aggregation_point column in the .tidy_iea_df. Default is IEATools::aggregation_flows$energy_industry_own_use.

transformation_processes

A string identifying the transformation processes in the flow_aggregation_point column in the .tidy_iea_df. Default is IEATools::aggregation_flows$transformation_processes.

main_act_producer_elect

A string identifying "Main activity producer electricity plants" in the flow column of the .tidy_iea_df. Default is IEATools::main_act_plants$main_act_prod_elect_plants.

main_act_producer_chp

A string identifying "Main activity producer CHP plants" in the flow column of the .tidy_iea_df. Default is IEATools::main_act_plants$main_act_prod_chp_plants.

main_act_producer_heat

A string identifying "Main activity producer heat plants" in the flow column of the .tidy_iea_df. Default is IEATools::main_act_plants$main_act_prod_heat_plants.

autoproducer_elect

A string identifying "Autoproducer electricity plants" in the flow column of the .tidy_iea_df. Default is IEATools::main_act_plants$autoprod_elect_plants.

autoproducer_chp

A string identifying "Autoproducer CHP plants" in the flow column of the .tidy_iea_df. Default is IEATools::transformation_processes$autoproducer_CHP_plants.

autoproducer_heat

A string identifying "Autoproducer CHP plants" in the flow column of the .tidy_iea_df. Default is IEATools::transformation_processes$autoprod_heat_plants.

own_use_elect_chp_heat

A string identifying the "Own use in electricity, CHP and heat plants" EIOU flow in the .tidy_iea_df. Default is IEATools::eiou_flows$own_use_elect_chp_heat_plants.

geothermal, hydro, solar_pv, solar_th, oceanic, wind

Renewable energy product names. See IEATools::renewable_products.

electricity

The name of the electricity product. Default is IEATools::electricity_products$electricity.

heat

The name of the heat product. Default is IEATools::heat_products$heat.

ratio_solar_th_elec

The ratio of primary energy to electricity to use for solar thermal. Default is 0.33 as this is the value assumed in the IEA's energy balances.

ratio_solar_th_heat

The ratio of primary energy to heat to use for solar thermal. Default is 1 as this is the value assumed in the IEA's energy balances.

ratio_geothermal_elec

The ratio of primary energy to electricity to use for geothermal. Default is 0.1 as this is the value assumed in the IEA's energy balances.

ratio_geothermal_heat

The ratio of primary energy to heat to use for geothermal. Default is 0.5 as this is the value assumed in the IEA's energy balances.

ratio_other_renewable_elec

The ratio of primary energy to electricity to use for hydro, solar photovoltaic, oceanic, and wind power. Default is 1 as this is the value assumed in the IEA's energy balances.

geothermal_plants, hydro_plants, solar_pv_plants, solar_th_plants, oceanic_plants, wind_power_plants

Names of renewable industries added. See IEATools::renewable_industries.

negzeropos

The name of a temporary column added to the data frame. Default is ".negzeropos".

ratio_elec_to_heat

A temporary column added to the data frame. Default is ".ratio_elec_to_heat".

.share_industry

The name of a temporary column added to specify the renewable industry for which the share of output is calculated. Default is ".share_industry".

.share

The name of a temporary column where the share of output is calculated for each renewable industry. Default is ".share".

Value

Returns a .tidy_iea_df with renewable electricity and heat from geothermal, hydro, solar thermal, solar photovoltaic, wind, and oceanic power specified.

Details

The primary energy use of hydro, geothermal, solar photovoltaic, solar thermal, oceanic, and wind power energy by main activity and autoproducer plants are used to create new renewable industries that produce electricity and heat (heat only in the case of geothermal and solar thermal). The physical content method is used to derive the electricity produced by renewable industries, except in the case of geothermal and solar thermal, for which the IEA uses other factors in its balances. In the case of CHP plants (which can be relevant for geothermal and solar thermal), the output of the new renewable industry follows the same heat vs electricity breakdown as the main industry from which it is derived.

Examples

library(dplyr)
load_tidy_iea_df() %>% 
  specify_renewable_plants()
#> # A tibble: 403 × 11
#>    Country Method EnergyType LastStage  Year LedgerSide FlowAggregationPoint    
#>    <chr>   <chr>  <chr>      <chr>     <dbl> <chr>      <chr>                   
#>  1 GHA     PCM    E          Final      1971 Supply     Total primary energy su…
#>  2 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#>  3 GHA     PCM    E          Final      1971 Supply     Total primary energy su…
#>  4 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#>  5 GHA     PCM    E          Final      1971 Supply     Total primary energy su…
#>  6 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#>  7 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#>  8 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#>  9 GHA     PCM    E          Final      1971 Supply     Total primary energy su…
#> 10 GHA     PCM    E          Final      2000 Supply     Total primary energy su…
#> # ℹ 393 more rows
#> # ℹ 4 more variables: Flow <chr>, Product <chr>, Unit <chr>, Edot <dbl>