Specify destinations for energy industry own use flows into transformation processes

The extended energy balance data from the IEA includes Energy industry own use (EIOU) for many transformation processes. Unfortunately, in some cases the EIOU flows into industries that aren't included in transformation processes. For example, Electricity is consumed by Own use in electricity, CHP and heat plants, which is not a transformation process. We have to make some decisions to ensure that EIOU is routed to actual transformation processes. See details for a list of changes made to the .tidy_iea_df data frame.

specify_tp_eiou(.tidy_iea_df,
  flow_aggregation_point = "Flow.aggregation.point",
  eiou = "Energy industry own use", flow = "Flow",
  own_use_elect_chp_heat = "Own use in electricity, CHP and heat plants",
  pumped_storage = "Pumped storage plants",
  nuclear_industry = "Nuclear industry",
  non_spec_energy = "Non-specified (energy)",
  main_act_producer_elect = "Main activity producer electricity plants",
  nonspecenergy_reclassify = "Non-specified (transformation)")

Arguments

.tidy_iea_df
an IEA data frame whose columns have been renamed by rename_iea_df_cols()

flow_aggregation_point
the name of the flow aggregation point column in .tidy_iea_df. Default is "Flow.aggregation.point".

eiou
a string identifying energy industry own use in the flow aggregation point column. Default is "Energy industry own use".

flow
the name of the flow column in .tidy_iea_df. Default is "Flow".

own_use_elect_chp_heat
a string identifying own use in electricity, CHP and heat plants in the flow column. Default is "Own use in electricity, CHP and heat plants".

pumped_storage
a string identifying pumped storage plants in the flow column. Default is "Pumped storage plants".

nuclear_industry
a string identifying nuclear plants in the flow column. Default is "Nuclear industry".

non_spec_energy
a string identifying non-specified energy in the flow solumn. Default is "Non-specified (energy)".

main_act_producer_elect
a string identifying main activity producer electricity plants. Default is "Main activity producter electricity plants".

nonspecenergy_reclassify
a string identifying the reclassified non-specified (energy) industry. Default is "Non-specified (transformation)".

Value

a modified version of .tidy_iea_df

Details

The following changes are made to the .tidy_iea_df data frame:

EIOU classified as own_use_elect_chp_heat is sent to main_act_producer_elect.

EIOU classified as pumped_storage is sent to main_act_producer_elect.

EIOU classified as nuclear_industry is sent to main_act_producer_elect.

EIOU classified as non_spec_energy is sent to nonspecenergy_reclassify.

Examples

library(dplyr)
load_tidy_iea_df() %>%
  specify_tp_eiou() %>%
  filter(Flow.aggregation.point == "Energy industry own use" &
           Flow == "Main activity producer electricity plants")

#> # A tibble: 5 x 11
#>   Method Last.stage Country  Year Ledger.side Flow.aggregatio… Energy.type Unit 
#>   <chr>  <chr>      <chr>   <dbl> <chr>       <chr>            <chr>       <chr>
#> 1 PCM    Final      GHA      1971 Supply      Energy industry… E           ktoe 
#> 2 PCM    Final      ZAF      1971 Supply      Energy industry… E           ktoe 
#> 3 PCM    Final      GHA      2000 Supply      Energy industry… E           ktoe 
#> 4 PCM    Final      ZAF      2000 Supply      Energy industry… E           ktoe 
#> 5 PCM    Final      ZAF      2000 Supply      Energy industry… E           ktoe 
#> # … with 3 more variables: Flow <chr>, Product <chr>, E.dot <dbl>