Transport domestic flows

A string vector containing names of Transport Flows in the IEA extended energy balances database for territorial transport energy consumption.

Usage

transport_domestic_flows

Format

A string vector with 11 entries.

domestic_aviation

The string identifying the Domestic aviation sector.

road

The string identifying the Road transport sector

rail

The string identifying the Rail transport sector

pipeline_transport

The string identifying the Pipeline transport sector

domestic_navigation

The string identifying the Domestic navigation sector

non_specified_transport

The string identifying the Non-specified transport sector

transport_not_elsewhere_specified

The string identifying the Transport not elsewhere specified sector. This string replaced non_specified_transport in the IEA's 2020 extended energy balances.

Examples

transport_domestic_flows
#> $domestic_navigation
#> [1] "Domestic navigation"
#> 
#> $world_marine_bunkers
#> [1] "World marine bunkers"
#> 
#> $international_navigation
#> [1] "International navigation"
#> 
#> $domestic_aviation
#> [1] "Domestic aviation"
#> 
#> $world_aviation_bunkers
#> [1] "World aviation bunkers"
#> 
#> $international_aviation
#> [1] "International aviation"
#> 
#> $road
#> [1] "Road"
#> 
#> $rail
#> [1] "Rail"
#> 
#> $pipeline_transport
#> [1] "Pipeline transport"
#> 
#> $non_specified_transport
#> [1] "Non-specified (transport)"
#> 
#> $transport_not_elsewhere_specified
#> [1] "Transport not elsewhere specified"
#>