Add a column of matrix names to tidy data frame

This function adds a column of matrix names to a tidy data frame wherein each row of .tidy_iea_df is a single value in an energy conversion chain. The default argument values assume that .tidy_iea_df uses IEA-style nomenclature and terminology, although .tidy_iea_df does not necessarily need to contain IEA data.

add_psut_matnames(
  .tidy_iea_df,
  ledger_side = "Ledger.side",
  supply = "Supply",
  consumption = "Consumption",
  flow_aggregation_point = "Flow.aggregation.point",
  flow = "Flow",
  production = "Production",
  resources = "Resources",
  product = "Product",
  e_dot = "E.dot",
  eiou = "Energy industry own use",
  neg_supply_in_fd = c("Exports", "International aviation bunkers",
    "International marine bunkers", "Losses", "Statistical differences", "Stock changes"),
  matnames = "matnames",
  R = "R",
  U_excl_EIOU = "U_excl_EIOU",
  U_EIOU = "U_EIOU",
  V = "V",
  Y = "Y"
)

Arguments

.tidy_iea_df
a data frame with ledger_side, flow_aggregation_point, flow, and e_dot columns.

ledger_side
the name of the column in .tidy_iea_df that contains ledger side (a string). Default is "Ledger.side".

supply
the identifier for items on the supply side of the ledger (a string). Default is "Supply".

consumption
the identifier for items on the consumption side of the ledger (a string). Default is "Consumption".

flow_aggregation_point
the name of the column in .tidy_iea_df that contains flow aggregation point information. Default is "Flow.aggregation.point".

flow
the name of the column in .tidy_iea_df that contains flow information. Default is "Flow".

production
a string identifying production in the flow column. Default is Production.

resources
a string identifying resources in the flow column. Default is Resources.

product
the name of the column in .tidy_iea_df that contains flow information. Default is "Product".

e_dot
the name of the column in .tidy_iea_df that contains energy and exergy values (a string). Default is "E.dot".

eiou
the identifier for items that are energy industry own use. Default is "Energy industry own use".

neg_supply_in_fd
identifiers for flow items that, when negative, are entries in the final demand (Y) matrix.

matnames
the name of the output column containing the name of the matrix to which a row's value belongs (a string). Default is "matnames".

R
the name for the resource matrix (a string). Default is "R".

U_excl_EIOU
the name for the use matrix that excludes energy industry own use (a string). Default is "U_excl_EIOU".

U_EIOU
the name for the energy industry own use matrix. Default is "U_EIOU".

V
the name for the make matrix (a string). Default is "V".

Y
the name for the final demand matrix (a string). Default is "Y".

Value

.tidy_iea_df with an added column matnames.

Details

In a reasonable workflow, this function would be followed by a call to add_row_col_meta() and matsindf::collapse_to_matrices().

This function respects groups when identifying entries in the resource matrix (R). So be sure to group .tidy_iea_df before calling this function.

Internally, this function adds a temporary column to .tidy_iea_df called ".R". An error will occur if .tidy_iea_df already has a column named ".R".

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

load_tidy_iea_df() %>%
  add_psut_matnames() %>%
  glimpse()

#> Observations: 399
#> Variables: 12
#> $ Country                <chr> "GHA", "GHA", "GHA", "GHA", "GHA", "GHA", "GHA…
#> $ Method                 <chr> "PCM", "PCM", "PCM", "PCM", "PCM", "PCM", "PCM…
#> $ Energy.type            <chr> "E", "E", "E", "E", "E", "E", "E", "E", "E", "…
#> $ Last.stage             <chr> "Final", "Final", "Final", "Final", "Final", "…
#> $ Year                   <dbl> 1971, 1971, 1971, 1971, 1971, 1971, 1971, 1971…
#> $ Ledger.side            <chr> "Supply", "Supply", "Supply", "Supply", "Suppl…
#> $ Flow.aggregation.point <chr> "Total primary energy supply", "Total primary …
#> $ Flow                   <chr> "Production", "Production", "Imports", "Import…
#> $ Product                <chr> "Primary solid biofuels", "Hydro", "Crude oil"…
#> $ Unit                   <chr> "ktoe", "ktoe", "ktoe", "ktoe", "ktoe", "ktoe"…
#> $ E.dot                  <dbl> 2088, 250, 916, 1, 21, 1, 18, -4, -4, -178, -2…
#> $ matnames               <chr> "R", "R", "V", "V", "V", "V", "V", "Y", "Y", "…