Add row, column, row type, and column type metadata

After calling add_psut_matnames(), call this function to add row, column, row type, and column type information to .tidy_iea_df.

add_row_col_meta(.tidy_iea_df, matname = "matname",
  product = "Product", flow = "Flow", U = "U", U_EIOU = "U_EIOU",
  R = "R", V = "V", Y = "Y", industry_type = "Industry",
  product_type = "Product", sector_type = "Industry",
  resource_type = "Industry", rowname = "rowname",
  colname = "colname", rowtype = "rowtype", coltype = "coltype")

Arguments

.tidy_iea_df
a data frame containing matname

matname
the name of the column in .tidy_iea_df that contains names of matrices (a string). Default is "matname".

product
the name of the column in .tidy_iea_df where Product names is found (a string). Default is "Product".

flow
the name of the column in .tidy_iea_df where Flow names is found (a string). Default is "Flow".

U
the name for use matrices (a string). Default is "U".

U_EIOU
the name for energy industry own use matrices (a string). Default is "U_EIOU".

R
the name for resource matrices (a string). Default is "R".

V
the name for make matrices (a string). Default is "V".

Y
the name for final demand matrices (a string). Default is "Y".

industry_type
the name that identifies production industries and and transformation processes (a string). Default is "Industry".

product_type
the name that identifies energy carriers (a string). Default is "Product".

sector_type
the name that identifies final demand sectors (a string). Default is "Industry".

resource_type
the name that identifies resource sectors (a string). Default is "Industry".

rowname
the name of the output column that contains row names for matrices (a string). Default is "rowname".

colname
the name of the output column that contains column names for matrices (a string). Default is "colname".

rowtype
the name of the output column that contains row types for matrices (a string). Default is "rowtype".

coltype
the name of the output column that contains column types for matrices (a string). Default is "coltype".

Value

.tidy_iea_df with additional columns named rowname, colname, rowtype, and coltype.

Examples

library(dplyr)
load_tidy_iea_df() %>%
  add_psut_matnames() %>%
  add_row_col_meta()

#> # A tibble: 399 x 16
#>    Method Last.stage Country  Year Ledger.side Flow.aggregatio… Energy.type
#>    <chr>  <chr>      <chr>   <dbl> <chr>       <chr>            <chr>      
#>  1 PCM    Final      GHA      1971 Supply      Total primary e… E          
#>  2 PCM    Final      GHA      1971 Supply      Total primary e… E          
#>  3 PCM    Final      GHA      1971 Supply      Total primary e… E          
#>  4 PCM    Final      GHA      1971 Supply      Total primary e… E          
#>  5 PCM    Final      GHA      1971 Supply      Total primary e… E          
#>  6 PCM    Final      GHA      1971 Supply      Total primary e… E          
#>  7 PCM    Final      GHA      1971 Supply      Total primary e… E          
#>  8 PCM    Final      GHA      1971 Supply      Total primary e… E          
#>  9 PCM    Final      GHA      1971 Supply      Total primary e… E          
#> 10 PCM    Final      GHA      1971 Supply      Total primary e… E          
#> # … with 389 more rows, and 9 more variables: Unit <chr>, Flow <chr>,
#> #   Product <chr>, E.dot <dbl>, matname <chr>, rowname <chr>, colname <chr>,
#> #   rowtype <chr>, coltype <chr>