Augment an IEA data frame

This function prepares an IEA data frame created by iea_df() for use in R.

augment_iea_df(
  .iea_df,
  country = "Country",
  ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point",
  flow = "Flow",
  product = "Product",
  energy_type = "Energy.type",
  energy_type_val = "E",
  method = "Method",
  method_val = "PCM",
  last_stage = "Last.stage",
  last_stage_val = "Final",
  unit = "Unit",
  unit_val = "ktoe",
  supply = "Supply",
  consumption = "Consumption",
  tpes = "Total primary energy supply",
  tpes_flows = c("Production", "Imports", "Exports", "International marine bunkers",
    "International aviation bunkers", "Stock changes"),
  tfc_compare = "TFC compare",
  tfc_compare_flows = c("Total primary energy supply", "Transfers",
    "Statistical differences", "Transformation processes", "Energy industry own use",
    "Losses"),
  transfers = "Transfers",
  statistical_differences = "Statistical differences",
  losses = "Losses",
  transformation_processes = "Transformation processes",
  tp_flows_suffix = "(transf.)",
  nstp_flows_suffix = "(transformation)",
  eiou = "Energy industry own use",
  eiou_flows_suffix = "(energy)",
  tfc = "Total final consumption",
  tfc_flows = c("Industry", "Transport", "Other", "Non-energy use"),
  industry = "Industry",
  industry_flows = c("Iron and steel", "Chemical and petrochemical",
    "Non-ferrous metals", "Non-metallic minerals", "Transport equipment", "Machinery",
    "Mining and quarrying", "Food and tobacco", "Paper, pulp and print",
    "Wood and wood products", "Construction", "Textile and leather",
    "Non-specified (industry)"),
  transport = "Transport",
  transport_flows = c("World aviation bunkers", "Domestic aviation", "Road", "Rail",
    "Pipeline transport", "World marine bunkers", "Domestic navigation",
    "Non-specified (transport)"),
  other = "Other",
  other_flows = c("Residential", "Commercial and public services",
    "Agriculture/forestry", "Fishing", "Non-specified (other)"),
  non_energy = "Non-energy use",
  non_energy_prefix = "Non-energy use",
  electricity_output = "Electricity output (GWh)",
  electricity_output_flows_prefix = "Electricity output (GWh)-",
  heat_output = "Heat output",
  heat_output_flows_prefix = "Heat output-",
  .rownum = ".rownum"
)

Arguments

.iea_df
a data frame produced by the iea_df() function

country
the name of the country column in .iea_df. Default is "Country".

ledger_side
the name of the ledger side column to be added to .iea_df. Default is "Ledger.side".

flow_aggregation_point
the name of the flow aggregation point column to be added to .iea_df. Default is "Flow.aggregation.point".

flow
the name of the flow column in .iea_df. Default is "Flow".

product
the name of the product column in .iea_df. Default is "Product".

energy_type
the name of the energy type column to be added to .iea_df. Default is "Energy.type.

energy_type_val
the value to put in the energy_type column. Default is "E".

method
the name of the method column to be added to .iea_df. Default is "Method.

method_val
the value to put in the method column. Default is "PCM" (Physical Content Method, which is used by the IEA).

last_stage
the name of the last stage column to be added to .iea_df. Default is "Last.stage.

last_stage_val
the value to put in the last_stage column. Default is "Final" (which is the last stage supplied by the IEA).

unit
the name of the unit column to be added to .iea_df. Default is "Unit".

unit_val
the value to put in the unit column. Default is "ktoe" for kilotons of oil equivalent.

supply
the string that identifies supply Ledger.side. Default is "Supply".

consumption
the string that identifies consumption Ledger.side. Default is "Consumption".

tpes
the string that identifies total primary energy supply Flow.aggregation.point. Default is "Total primary energy supply.

tpes_flows
a vector of strings that give flows that are aggregated to Total primary energy supply.

tfc_compare
a string that identifies the TFC compare flow aggregation point. Default is TFC compare.

tfc_compare_flows
a vector of strings that give Flows that are aggregated to TFC compare.

transfers
= a string that identifies transfers in the flow column. Default is "Transfers".

statistical_differences
a string that identifies statistical differences in flow column. Default is "Statistical differences".

losses
the string that indicates losses in the Flow column. Default is "Losses".

transformation_processes
the string that indicates transformation processes in the Flow column. Default is "Transformation processes".

tp_flows_suffix
the suffix for transformation processes in the Flow column. Default is "(transf.)".

nstp_flows_suffix
the suffix for non-specified transformation processes in the Flow column. Default is "(transformation)".

eiou
the string that identifies energy industry own use in the Flow column. Default is "Energy industry own use".

eiou_flows_suffix
the suffix for energy industry own use in the Flow column. Default is "(energy)".

tfc
the string that identifies total final consumption in the Flow column. Default is "Total final consumption".

tfc_flows
a vector of strings that give total final consumption in the Flow column.

industry
a string that names the industry Flow.aggregation.point. Default is "Industry".

industry_flows
a vector of strings representing Flows to be aggregated in the Industry Flow.aggregation.point.

transport
a string that names the transport Flow.aggregation.point. Default is "Transport".

transport_flows
a vector of strings representing Flows to be aggregated in the Transport Flow.aggregation.point.

other
a string that names the other Flow.aggregation.point. Default is "Other".

other_flows
a vector of strings representing Flows to be aggregated in the Other Flow.aggregation.point.

non_energy
a string that names the non-energy Flow.aggregation.point. Default is "Non-energy use".

non_energy_prefix
a string prefix for Flows to be aggregated in the Non-energy use Flow.aggregation.point.

electricity_output
a string that names the electricity output Flow. Default is "Electricity output (GWh)".

electricity_output_flows_prefix
a string prefix for Flows to be aggregated in electricity output. Default is "Electricity output (GWh)-".

heat_output
a string that names the heat output Flow. Default is "Heat output".

heat_output_flows_prefix
a string prefix for Flows to be aggregated in heat output. Default is "Heat output-".

.rownum
the name of a column created (and destroyed) internally by this function. The .rownum column temporarily holds row numbers for internal calculations. The .rownum column is deleted before returning.

Value

.iea_df with additional columns named ledger_side, flow_aggregation_point, energy_type, and unit.

Details

This function solves several problems. The first problem is that metadata in the COUNTRY, FLOW, and PRODUCT columns of an IEA data table are not unique. A second problem is that the FLOW column contains both industries to which energy is flowing and the type of flow that is involved. (E.g., the suffix "(energy)" means that the flow is an own use by the energy industry. The "(transf.)" suffix means that a flow is involved in a transformation process between primary and final energy.

To solve these problems, two additional columns are added: Ledger.side and Flow.aggregation.point. Ledger.side can be one of "Supply" or "Consumption", corresponding to the top or bottom of the IEA's tables, respectively. Flow.aggregation.point indicates the next level of aggregation for these data. Flow.aggregation.point can be one of "Total primary energy supply", "Transformation processes", "Energy industry own use", or "TFC compare" on the Supply side of the ledger. On the Consumption side of the ledger, Flow.aggregation.point can be one of "Industry", "Transport", "Other", or "Non-energy use". When the Flow.aggregation.point column is present, the need for the "(energy)" and "(transf.)" suffixes is eliminated, so they are deleted.

The third problem this function solves is that energy type and units are not specified in IEA data. An Energy.type column is added with the value of energy_type_val. (Default is E, for energy, as opposed to X, which would be exergy.) A Unit column is added with the value of unit_val. (Default is ktoe, although any string can be specified in unit_val.)

Note that this function decides where to divide Supply from Consumption. To do so, it first looks for rows in which Flow is "Losses". The last "Losses" row is the last row of the Supply side of the ledger. If "Losses" rows are not found, the function looks for rows in which Flow is "Total final consumption". The first "Total final consumption" row is the first row of the Consumption side of the ledger. If neither "Losses" nor "Total final consumption" Flows are present, an error is generated.

Examples

iea_df(text = paste0(",,TIME,1960,1961\n",
                       "COUNTRY,FLOW,PRODUCT\n",
                       "World,Production,Hard coal (if no detail),42,43\n",
                       "World,Losses,Hard coal (if no detail),1,2")) %>%
  rename_iea_df_cols() %>%
  augment_iea_df()

#> # A tibble: 2 x 11
#>   Country Method Energy.type Last.stage Ledger.side Flow.aggregatio… Flow 
#>   <chr>   <chr>  <chr>       <chr>      <chr>       <chr>            <chr>
#> 1 World   PCM    E           Final      Supply      Total primary e… Prod…
#> 2 World   PCM    E           Final      Supply      TFC compare      Loss…
#> # … with 4 more variables: Product <chr>, Unit <chr>, `1960` <dbl>,
#> #   `1961` <dbl>