Calculate balances on a tidy IEA data frame

It is important to know whether energy flows are balanced before proceeding with further analyses. This function calculates energy balances by groups in .tidy_iea_df. So be sure to group .tidy_iea_df by appropriate variables before calling this function. Grouping should definitely be done on the Product column. Typically, grouping is also done on Country, Year, Energy.type, Last.stage, etc. columns. Grouping should not be done on the Ledger.side column or the Flow column. To test whether all balances are OK, use the tidy_iea_df_balanced() function.

calc_tidy_iea_df_balances(.tidy_iea_df, ledger_side = "Ledger.side",
  e_dot = "E.dot", unit = "Unit", supply = "Supply",
  consumption = "Consumption", supply_sum = "supply_sum",
  consumption_sum = "consumption_sum",
  supply_minus_consumption = "supply_minus_consumption",
  balance_OK = "balance_OK", err = "err", tol = 1e-06)

Arguments

.tidy_iea_df
an IEA-style data frame containing a ledger_side, product, flow, and energy rate (e_dot) columns along with grouping columns, typically Country, Year, Product, etc. a Ledger.side column.

ledger_side
the name of the column in .tidy_iea_data that contains ledger side information (a string). Default is "Ledger.side".

e_dot
the name of the column in .tidy_iea_data that contains energy flow data. Default is "E.dot".

unit
the name of the colum in .tidy_iea_data that contains the units for the energy flow data. Default is "Unit".

supply
the identifier for supply data in the ledger_side column (a string). Default is "Supply".

consumption
the identifier for consumption data in the ledger_side column (a string). Default is "Consumption".

supply_sum
the name of a new column that will contain the sum of all supply for that group. Default is "supply_sum".

consumption_sum
the name of a new column that will contain the sum of all consumption for that group. Default is "consumption_sum".

supply_minus_consumption
the name of a new column that will contain the difference between supply and consumption for that group. Default is "supply_minus_consumption".

balance_OK
the name of a new logical column that tells whether a row's energy balance is OK. Default is "balance_OK".

err
the name of a new column that indicates the energy balance error for each group. Default is "err".

tol
if the difference between supply and consumption is greater than tol, balance_OK will be set to FALSE. Default is 1e-6.

Value

.tidy_iea_df with additional columns supply_sum, consumption_sum, supply_minus_consumption, balance_OK, and err.

Details

Supply side and consumption side energy flows are aggregated to a supply_sum and a consumption_sum column. There are two possibilities:

A Product appears only on the supply side, because it is completely transformed before reaching the consumption side of the ledger. In this case, the consumption_sum column will have an NA value, and the supply_minus_consumption column will also have an NA value.

A Product appears on both the supply and the demand sides of the ledger and, therefore, is not NA in the consumption_sum column and the supply_minus_consumption column. The column balance_OK is calculated as follows:

For the first situation, consumption_sum will be 0 (within tol) if the Product is balanced and balance_OK will have a value of TRUE. If not, balance_OK will have a value of FALSE.

In the second situation, the difference between supply_sum and consumption_sum is calculated (supply_minus_consumption). If the product is balanced, supply_minus_consumption will be 0 (within tol) and balance_OK will be TRUE. If not, balance_OK will be FALSE.

Examples

library(dplyr)
Ebal <- file.path("extdata", "GH-ZA-ktoe-Extended-Energy-Balances-sample.csv") %>%
  system.file(package = "IEATools") %>%
  iea_df() %>%
  rename_iea_df_cols() %>%
  remove_agg_memo_flows() %>%
  augment_iea_df() %>%
  tidy_iea_df() %>%
  group_by(Country, Year, Energy.type, Unit, Product) %>%
  calc_tidy_iea_df_balances()
head(Ebal, 5)

#> # A tibble: 5 x 10
#> # Groups:   Country, Year, Energy.type, Unit [1]
#>   Country Year  Energy.type Unit  Product supply_sum consumption_sum
#>   <chr>   <chr> <chr>       <chr> <chr>        <dbl>           <dbl>
#> 1 Ghana   1971  E           ktoe  Aviati…          0              NA
#> 2 Ghana   1971  E           ktoe  Charco…        119             119
#> 3 Ghana   1971  E           ktoe  Crude …          0              NA
#> 4 Ghana   1971  E           ktoe  Electr…        236             235
#> 5 Ghana   1971  E           ktoe  Fuel o…         94              95
#> # … with 3 more variables: supply_minus_consumption <dbl>, balance_OK <lgl>,
#> #   err <dbl>