Remove specification strings from a column

Flow and Product columns of IEA data frames may have been "specified" with one of the functions of specify_all(). The specifying makes it difficult to sort the columns in IEA order (with sort_iea_df()), as the Flow and Product columns now contain non-IEA flows and products. To enable sorting, this function de-specifies a column in .df.

despecify_col(
  .df,
  col,
  despecified_col,
  notation = IEATools::specify_notation,
  .open = notation$open,
  .close = notation$close
)

Arguments

.df
the data frame in which col exists

col
the string name of the column in .df to be de-specified

despecified_col
the string name of the column in .df to contain the de-specified version of col

notation
a list of specification notations. Default is IEATools::specify_notation.

.open
the opening string for specification decoration. Default is IEATools::specify_notation$open.

.close
the closing string for specification decoration. Default is IEATools::specify_notation$close.

Value

a de-specified version of .df

Details

De-specifying includes the following changes: * Any "Resource" flows are replaced by "Production". E.g., "Resources [Coal]" becomes "Production". * All parenthetical decorations are removed. E.g., "Other bituminous coal [of Coal mines]" becomes "Other bituminous coal".

Identification of parenthetical notation delimiters is determined by the specify_notation object of this package.

Examples

library(dplyr)
load_tidy_iea_df() %>%
  specify_all() %>%
  despecify_col(col = "Flow", despecified_col = "clean_Flow") %>%
  select(Flow, Product, E.dot, clean_Flow) %>%
  filter(endsWith(Flow, specify_notation$close))

#> # A tibble: 83 x 4
#>    Flow                            Product               E.dot clean_Flow       
#>    <chr>                           <chr>                 <dbl> <chr>            
#>  1 Exports [of Fuel oil]           Fuel oil            -178.   Exports          
#>  2 Exports [of Motor gasoline exc… Motor gasoline exc…   -4.28 Exports          
#>  3 Exports [of Other kerosene]     Other kerosene        -4.18 Exports          
#>  4 Imports [of Aviation gasoline]  Aviation gasoline      1.07 Imports          
#>  5 Imports [of Crude oil [of Oil … Crude oil [of Oil …  916.   Imports          
#>  6 Imports [of Kerosene type jet … Kerosene type jet …   21.3  Imports          
#>  7 Imports [of Lubricants]         Lubricants            18.1  Imports          
#>  8 Imports [of Other kerosene]     Other kerosene         1.05 Imports          
#>  9 International aviation bunkers… Aviation gasoline     -1.07 International av…
#> 10 International aviation bunkers… Kerosene type jet …  -43.7  International av…
#> # … with 73 more rows