Energy industry own use (EIOU) flows

A string vector containing names of Energy industry own use Flows in the IEA extended energy balances database.

eiou_flows

Format

A string vector with 17 entries.

bkb_peat_briquette_plants

The string identifier for BKB/peat briquette plants.

blast_furnaces

The string identifier for Blast furnaces.

charcoal_plants

The string identifier for Charcoal production plants.

coal_liquefaction_plants

The string identifier for Coal liquefaction plants.

coal_mines

The string identifier for Coal mines.

coke_ovens

The string identifier for Coke ovens.

gas_work

The string identifier for Gas works.

gas_to_liquids_plants

The string identifier for Gas-to-liquids (GTL) plants.

gasification_plants

The string identifier for Gasification plants for biogases.

liquefaction_regasification_plants

The string identifier for Liquefaction (LNG) / regasification plants.

non_specified_eiou

The string identifier for Non-specified (energy).

nuclear_industry

The string identifier for Nuclear industry.

oil_and_gas_extraction

The string identifier for Oil and gas extraction.

oil_refineries

The string identifier for Oil refineries.

own_use_elect_chp_heat_plants

The string identifier for Own use in electricity, CHP and heat plants.

patent_fuel_plants

The string identifier for Patent fuel plants.

pumped_storage_plants

The string identifier for Pumped storage plants.

Examples

eiou_flows

#> $bkb_peat_briquette_plants
#> [1] "BKB/peat briquette plants"
#> 
#> $blast_furnaces
#> [1] "Blast furnaces"
#> 
#> $charcoal_plants
#> [1] "Charcoal production plants"
#> 
#> $coal_liquefaction_plants
#> [1] "Coal liquefaction plants"
#> 
#> $coal_mines
#> [1] "Coal mines"
#> 
#> $coke_ovens
#> [1] "Coke ovens"
#> 
#> $gas_work
#> [1] "Gas works"
#> 
#> $gas_to_liquids_plants
#> [1] "Gas-to-liquids (GTL) plants"
#> 
#> $gasification_plants
#> [1] "Gasification plants for biogases"
#> 
#> $liquefaction_regasification_plants
#> [1] "Liquefaction (LNG) / regasification plants"
#> 
#> $non_specified_eiou
#> [1] "Non-specified (energy)"
#> 
#> $nuclear_industry
#> [1] "Nuclear industry"
#> 
#> $oil_and_gas_extraction
#> [1] "Oil and gas extraction"
#> 
#> $oil_refineries
#> [1] "Oil refineries"
#> 
#> $own_use_elect_chp_heat_plants
#> [1] "Own use in electricity, CHP and heat plants"
#> 
#> $patent_fuel_plants
#> [1] "Patent fuel plants"
#> 
#> $pumped_storage_plants
#> [1] "Pumped storage plants"
#>