Extract a unit summation matrix from a tidy data frame

Unit summation matrices have products in rows and units in columns, with 1s where a product is expressed in the unit and 0s otherwise.

extract_S_units_from_tidy(
  .tidy_iea_df,
  ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point",
  flow = "Flow",
  product = "Product",
  e_dot = "E.dot",
  unit = "Unit",
  s_units = "S_units",
  val = ".val",
  rowtype = ".rowtype",
  coltype = ".coltype"
)

Arguments

.tidy_iea_df
the tidy data frame from which a unit summation S_units matrix is to be formed.

ledger_side
the name of the ledger side column in .tidy_iea_df. Default is "Ledger.side".

flow_aggregation_point
the name of the flow aggregation point column in .tidy_iea_df. Default is "Flow.aggregation.point".

flow
the name of the flow column in .tidy_iea_df. Default is "Flow".

product
the name of the product column in .tidy_iea_df. Default is "Product".

e_dot
the name of the energy flow rate column in .tidy_iea_df. Default is "E.dot".

unit
the name of the unit column in ".tidy_iea_df". Default is "Unit".

s_units
the name of the unit summation column to be added to .tidy_iea_df. Default is "S_unit".

val
the name of a temporary column to be created in .tidy_iea_df. Default is ".val".

rowtype
the name of a temporary rowtype column created in .tidy_iea_df. Default is "rowtype".

coltype
the name of a temporary coltype column created in .tidy_iea_df. Default is "coltype".

Value

a data frame containing grouping variables and a new column of unit summation matrices called s_unit.

Details

.tidy_iea_df should be grouped as needed, typically on Country, Year, Energy.type, Last.stage, etc., but not on Unit, Flow or Product. .tidy_iea_df is typically obtained from tidy_iea_df().

Examples

library(dplyr)
load_tidy_iea_df() %>%
  extract_S_units_from_tidy()

#>   Country Method Energy.type Last.stage Year
#> 1     GHA    PCM           E      Final 1971
#> 2     GHA    PCM           E      Final 2000
#> 3     ZAF    PCM           E      Final 1971
#> 4     ZAF    PCM           E      Final 2000
#>                                                                           S_units
#> 1                                        1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 2                                  1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 3                   1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 4 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1