IEA order for combined flow aggregation point and flow

A string vector containing entries for a united flow aggregation point and flow column in IEA order.

fap_flows

Format

A string list with 81

Examples

fap_flows

#>  [1] "Total primary energy supply_Production"                             
#>  [2] "Total primary energy supply_Coal mines"                             
#>  [3] "Total primary energy supply_Oil and gas extraction"                 
#>  [4] "Total primary energy supply_Imports"                                
#>  [5] "Total primary energy supply_Exports"                                
#>  [6] "Total primary energy supply_International marine bunkers"           
#>  [7] "Total primary energy supply_International aviation bunkers"         
#>  [8] "Total primary energy supply_Stock changes"                          
#>  [9] "TFC compare_Transfers"                                              
#> [10] "TFC compare_Statistical differences"                                
#> [11] "Transformation processes_Coal mines"                                
#> [12] "Transformation processes_Oil and gas extraction"                    
#> [13] "Transformation processes_Main activity producer electricity plants" 
#> [14] "Transformation processes_Autoproducer electricity plants"           
#> [15] "Transformation processes_Main activity producer CHP plants"         
#> [16] "Transformation processes_Autoproducer CHP plants"                   
#> [17] "Transformation processes_Main activity producer heat plants"        
#> [18] "Transformation processes_Autoproducer heat plants"                  
#> [19] "Transformation processes_Heat pumps"                                
#> [20] "Transformation processes_Electric boilers"                          
#> [21] "Transformation processes_Chemical heat for electricity production"  
#> [22] "Transformation processes_Blast furnaces"                            
#> [23] "Transformation processes_Gas works"                                 
#> [24] "Transformation processes_Coke ovens"                                
#> [25] "Transformation processes_Patent fuel plants"                        
#> [26] "Transformation processes_BKB/peat briquette plants"                 
#> [27] "Transformation processes_Oil refineries"                            
#> [28] "Transformation processes_Petrochemical plants"                      
#> [29] "Transformation processes_Coal liquefaction plants"                  
#> [30] "Transformation processes_Gas-to-liquids (GTL) plants"               
#> [31] "Transformation processes_For blended natural gas"                   
#> [32] "Transformation processes_Charcoal production plants"                
#> [33] "Transformation processes_Non-specified"                             
#> [34] "Energy industry own use_Coal mines"                                 
#> [35] "Energy industry own use_Oil and gas extraction"                     
#> [36] "Energy industry own use_Blast furnaces"                             
#> [37] "Energy industry own use_Gas works"                                  
#> [38] "Energy industry own use_Gasification plants for biogases"           
#> [39] "Energy industry own use_Coke ovens"                                 
#> [40] "Energy industry own use_Patent fuel plants"                         
#> [41] "Energy industry own use_BKB/peat briquette plants"                  
#> [42] "Energy industry own use_Oil refineries"                             
#> [43] "Energy industry own use_Coal liquefaction plants"                   
#> [44] "Energy industry own use_Liquefaction (LNG) / regasification plants" 
#> [45] "Energy industry own use_Gas-to-liquids (GTL) plants"                
#> [46] "Energy industry own use_Own use in electricity, CHP and heat plants"
#> [47] "Energy industry own use_Pumped storage plants"                      
#> [48] "Energy industry own use_Nuclear industry"                           
#> [49] "Energy industry own use_Main activity producer electricity plants"  
#> [50] "Energy industry own use_Charcoal production plants"                 
#> [51] "Energy industry own use_Non-specified"                              
#> [52] "TFC compare_Losses"                                                 
#> [53] "Industry_Mining and quarrying"                                      
#> [54] "Industry_Construction"                                              
#> [55] "Industry_Iron and steel"                                            
#> [56] "Industry_Chemical and petrochemical"                                
#> [57] "Industry_Non-ferrous metals"                                        
#> [58] "Industry_Non-metallic minerals"                                     
#> [59] "Industry_Transport equipment"                                       
#> [60] "Industry_Machinery"                                                 
#> [61] "Industry_Food and tobacco"                                          
#> [62] "Industry_Paper, pulp and printing"                                  
#> [63] "Industry_Wood and wood products"                                    
#> [64] "Industry_Textile and leather"                                       
#> [65] "Industry_Industry not elsewhere specified"                          
#> [66] "Transport_World aviation bunkers"                                   
#> [67] "Transport_Domestic aviation"                                        
#> [68] "Transport_Road"                                                     
#> [69] "Transport_Rail"                                                     
#> [70] "Transport_Pipeline transport"                                       
#> [71] "Transport_World marine bunkers"                                     
#> [72] "Transport_Domestic navigation"                                      
#> [73] "Transport_Non-specified (transport)"                                
#> [74] "Other_Residential"                                                  
#> [75] "Other_Commercial and public services"                               
#> [76] "Other_Agriculture/forestry"                                         
#> [77] "Other_Fishing"                                                      
#> [78] "Other_Final consumption not elsewhere specified"                    
#> [79] "Non-energy use_Non-energy use industry/transformation/energy"       
#> [80] "Non-energy use_Non-energy use in transport"                         
#> [81] "Non-energy use_Non-energy use in other"