Fix IEA energy balances

IEA extended energy balance data are sometimes not quite balanced. In fact, supply and consumption are often wrong by a few ktoe for any given Product in a Country in a Year. This function ensures that the balance is perfect by adjusting the Statistical differences flow on a per-product basis.

fix_tidy_iea_df_balances(.tidy_iea_df, ledger_side = "Ledger.side",
  supply = "Supply", consumption = "Consumption",
  flow_aggregation_point = "Flow.aggregation.point",
  tfc_compare = "TFC compare", flow = "Flow",
  statistical_differences = "Statistical differences",
  product = "Product", e_dot = "E.dot", err = ".err",
  remove_zeroes = TRUE)

Arguments

.tidy_iea_df
a tidy data frame containing IEA data, typically the output

ledger_side
the name of the ledger side column in .tidy_iea_df. Default is "Ledger.side".

supply
a string indicating the supply side of the ledger in the ledger_side column. Default is "Supply".

consumption
a string indicating the consumption side of the ledger in the ledger_side column. Default is "Consumption".

flow_aggregation_point
the name of the flow aggregation point column in .tidy_iea_df. Default is "Flow.aggregation.point".

tfc_compare
a string indicating the Total final consumption comparison flow aggregation point. Default is "TFC compare".

flow
the name of the flow column in .tidy_iea_df. Default is "Flow".

statistical_differences
a string indicating statistical differences in the flow column. Default is "Statistical differences".

product
the name of the product column in .tidy_iea_df. Default is "Product".

e_dot
the name of the energy flow column in .tidy_iea_df. Default is "E.dot".

err
the name of a temporary error column added to .tidy_iea_df. Default is ".err".

remove_zeroes
a logical telling whether to remove 0s after balancing.

Value

.tidy_iea_df with adjusted Statistical differences flows such that the data for each product are in perfect energy balance.

Details

This function assumes that .tidy_iea_df is grouped appropriately. So be sure to group .tidy_iea_df by appropriate variables before calling this function. Grouping should definitely be done on the Product column. Typically, grouping is also done on Country, Year, Energy.type, Last.stage, etc. columns. Grouping should not be done on the flow column or the ledger_side column.

Internally, this function calls calc_tidy_iea_df_balances() and adjusts Statistical differences by any imbalances that are present.

Examples

library(dplyr)
unbalanced <- load_tidy_iea_df()
# This will not be balanced, because the IEA data are not in perfect balance
unbalanced %>%
  group_by(Country, Year, Energy.type, Unit, Product) %>%
  calc_tidy_iea_df_balances() %>%
  tidy_iea_df_balanced()

#> [1] FALSE

# This will be balanced, becasue we fix the imbalances.
unbalanced %>%
  group_by(Country, Year, Energy.type, Unit, Product) %>%
  fix_tidy_iea_df_balances() %>%
  calc_tidy_iea_df_balances() %>%
  tidy_iea_df_balanced()

#> [1] TRUE