Fix IEA energy balances

IEA extended energy balance data are sometimes not quite balanced. In fact, supply and consumption are often wrong by a few ktoe for any given Product in a Country in a Year. This function ensures that the balance is perfect by adjusting the Statistical differences flow on a per-product basis.

fix_tidy_iea_df_balances(
  .tidy_iea_df,
  max_fix = 5,
  remove_zeroes = TRUE,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  supply = IEATools::ledger_sides$supply,
  consumption = IEATools::ledger_sides$consumption,
  tfc_compare = IEATools::aggregation_flows$tfc_compare,
  statistical_differences = IEATools::tfc_compare_flows$statistical_differences,
  .err = ".err"
)

Arguments

.tidy_iea_df
a tidy data frame containing IEA extended energy balance data

max_fix
the maximum energy balance that will be fixed without giving an error. Default is 5.

remove_zeroes
a logical telling whether to remove 0s after balancing. Default is TRUE.

ledger_side, flow_aggregation_point, flow, product, e_dot
See IEATools::iea_cols.

supply, consumption
See IEATools::ledger_sides.

tfc_compare
See IEATools::aggregation_flows.

statistical_differences
See IEATools::tfc_compare_flows.

.err
the name of a temporary error column added to .tidy_iea_df. Default is ".err".

Value

.tidy_iea_df with adjusted statistical_differences flows such that the data for each product are in perfect energy balance.

Details

This function assumes that .tidy_iea_df is grouped appropriately prior to passing into this function. The Product column should definitely be included in grouping_vars, but any other grouping level is fine. Typically, grouping should be done by Country, Year, Energy.type, Last.stage, Product, etc. columns. Grouping should not be done on the flow_aggregation_point, Flow, or ledger_side columns.

Internally, this function calls calc_tidy_iea_df_balances() and adjusts the value of the statistical_differences column to compensate for any imbalances that are present.

If energy balance for any product is greater than max_fix (default 5), an error will be emitted, and execution will halt. This behavior is intended to identify any places where there are gross energy imbalances that should be investigated prior to further analysis.

Examples

library(dplyr)
# Balances are calculated for each group.
# Remember that grouping should _not_ be done on
# the `flow_aggregation_point`, `Flow`, or `ledger_side` columns.
grouped_iea_df <- load_tidy_iea_df() %>%
  group_by(Country, Method, Energy.type, Last.stage, Year, Product)
# unbalanced will not be balanced, because the IEA data are not in perfect balance.
# Because we have grouped by key variables, 
# `calc_tidy_iea_df_balances` provides energy balances 
# on a per-product basis.
# The `err` column shows the magnitude of the imbalances.
unbalanced <- grouped_iea_df %>%
  calc_tidy_iea_df_balances()
unbalanced

#> # A tibble: 78 x 12
#>    Country Method Energy.type Last.stage  Year Product Unit  supply_sum
#>    <chr>   <chr>  <chr>       <chr>      <dbl> <chr>   <chr>      <dbl>
#>  1 GHA     PCM    E           Final       1971 Aviati… ktoe    0.      
#>  2 GHA     PCM    E           Final       1971 Charco… ktoe    1.19e+ 2
#>  3 GHA     PCM    E           Final       1971 Crude … ktoe   -3.55e-14
#>  4 GHA     PCM    E           Final       1971 Electr… ktoe    2.35e+ 2
#>  5 GHA     PCM    E           Final       1971 Fuel o… ktoe    9.51e+ 1
#>  6 GHA     PCM    E           Final       1971 Gas/di… ktoe    2.18e+ 2
#>  7 GHA     PCM    E           Final       1971 Hydro   ktoe    0.      
#>  8 GHA     PCM    E           Final       1971 Kerose… ktoe   -3.55e-15
#>  9 GHA     PCM    E           Final       1971 Liquef… ktoe    3.39e+ 0
#> 10 GHA     PCM    E           Final       1971 Lubric… ktoe    1.81e+ 1
#> # … with 68 more rows, and 4 more variables: consumption_sum <dbl>,
#> #   supply_minus_consumption <dbl>, balance_OK <lgl>, err <dbl>

# The `tidy_iea_df_balanced` function returns `TRUE` if and only if `all` of the groups (rows) 
# in `unbalanced` are balanced.
unbalanced %>%
  tidy_iea_df_balanced()

#> [1] FALSE

# Fix the imbalances.
balanced <- grouped_iea_df %>%
  fix_tidy_iea_df_balances() %>%
  calc_tidy_iea_df_balances()
balanced

#> # A tibble: 78 x 12
#>    Country Method Energy.type Last.stage  Year Product Unit  supply_sum
#>    <chr>   <chr>  <chr>       <chr>      <dbl> <chr>   <chr>      <dbl>
#>  1 GHA     PCM    E           Final       1971 Aviati… ktoe        0   
#>  2 GHA     PCM    E           Final       1971 Charco… ktoe      119.  
#>  3 GHA     PCM    E           Final       1971 Crude … ktoe        0   
#>  4 GHA     PCM    E           Final       1971 Electr… ktoe      235.  
#>  5 GHA     PCM    E           Final       1971 Fuel o… ktoe       95.1 
#>  6 GHA     PCM    E           Final       1971 Gas/di… ktoe      218.  
#>  7 GHA     PCM    E           Final       1971 Hydro   ktoe        0   
#>  8 GHA     PCM    E           Final       1971 Kerose… ktoe        0   
#>  9 GHA     PCM    E           Final       1971 Liquef… ktoe        3.39
#> 10 GHA     PCM    E           Final       1971 Lubric… ktoe       18.1 
#> # … with 68 more rows, and 4 more variables: consumption_sum <dbl>,
#> #   supply_minus_consumption <dbl>, balance_OK <lgl>, err <dbl>

balanced %>%
  tidy_iea_df_balanced()

#> [1] TRUE