Create eta_fu and phi_u vectors

This function creates vectors from a filled final-to-useful efficiency table created by write_fu_allocation_template(). The two vectors are:

eta_fu: a vector of final-to-useful energy efficiencies, and

phi_u: a vector of useful exergy-to-useful energy ratios.

form_eta_fu_phi_u_vecs(
  .eta_fu_table,
  unit = IEATools::iea_cols$unit,
  year = IEATools::iea_cols$year,
  quantity = IEATools::template_cols$quantity,
  machine = IEATools::template_cols$machine,
  eu_product = IEATools::template_cols$eu_product,
  e_dot_machine = IEATools::template_cols$e_dot_machine,
  e_dot_machine_perc = IEATools::template_cols$e_dot_machine_perc,
  maximum_values = IEATools::template_cols$maximum_values,
  eta_fu = IEATools::template_cols$eta_fu,
  phi_u = IEATools::template_cols$phi_u,
  matnames = IEATools::mat_meta_cols$matnames,
  matvals = IEATools::mat_meta_cols$matvals,
  rownames = IEATools::mat_meta_cols$rownames,
  colnames = IEATools::mat_meta_cols$colnames,
  rowtypes = IEATools::mat_meta_cols$rowtypes,
  coltypes = IEATools::mat_meta_cols$coltypes,
  product = IEATools::row_col_types$product,
  industry = IEATools::row_col_types$industry,
  arrow_note = IEATools::arrow_notation,
  from_note = IEATools::from_notation
)

Arguments

.eta_fu_table
a final-to-useful efficiency table read by load_eta_fu_allocation_data(). A template for this table should have been created by eta_fu_table() and write_eta_fu_table().

unit, year
See IEATools::iea_cols.

quantity, machine, eu_product, e_dot_machine, e_dot_machine_perc, maximum_values, eta_fu, phi_u
See IEATools::template_cols.

matnames, matvals, rownames, colnames, rowtypes, coltypes
See IEATools::mat_meta_cols.

product, industry
See IEATools::row_col_types.

arrow_note, from_note
Notation vectors used for creating the eta_fu and phi vectors. See matsbyname::notation_vec(). Defaults are arrow_notation and ``from_notation`, respectively.

Value

a wide data frame with metadata columns (and year) along with columns for eta_fu and phi_u vectors.

Details

The vectors eta_fu and phi_u have special rownames that indicate sources and types of useful energy flows. Row names have the pattern "machine -> useful energy form" to indicate the energy efficiency of "machine" for making "useful energy form" or the exergy-to-energy ratio of the useful energy form created by machine.

Columns are named by the variable in the vector: "eta_fu" for final-to-useful efficiencies and "phi_u" for useful exergy-to-useful energy ratios.

Examples

load_eta_fu_data() %>%
  form_eta_fu_phi_u_vecs()

#> # A tibble: 4 x 7
#>   Country Method Energy.type Last.stage  Year eta.fu            phi.u           
#>   <chr>   <chr>  <chr>       <chr>      <dbl> <list>            <list>          
#> 1 GHA     PCM    E           Final       1971 <dbl[,1] [22 × 1… <dbl[,1] [22 × …
#> 2 GHA     PCM    E           Final       2000 <dbl[,1] [23 × 1… <dbl[,1] [23 × …
#> 3 ZAF     PCM    E           Final       1971 <dbl[,1] [33 × 1… <dbl[,1] [32 × …
#> 4 ZAF     PCM    E           Final       2000 <dbl[,1] [47 × 1… <dbl[,1] [47 × …