IEA data frame column names

A string list containing named names of columns in IEA data frames. The IEA data frames can be tidy (with one row for each data point) or wide (with years spread to the right). Items in the list provide default values for column name function arguments throughout the IEATools package.

Usage

iea_cols

Format

A string list with 11 entries.

country

The name of a column containing countries.

method

The name of a column containing methods for calculating primary energy equivalent of renewable electricity. See IEATools::methods.

energy_type

The name of a column containing energy types. See IEATools::energy_types.

last_stage

The name of a column containing last stages of energy conversion chain analysis. See IEATools::last_stages.

year

The name of a column containing years in wide data frames.

ledger_side

The name of a column containing ledger sides. See IEATools::ledger_sides.

flow_aggregation_point

The name of a column containing flow aggregation points. See IEATools::flow_aggregation_points.

flow

The name of a column containing IEA flow information. See IEATools::flows.

product

The name of a column containing IEA products. See IEATools::products.

unit

The name of a column containing units for energy flows.

e_dot

The name of a column containing energy flow rates.

Examples

iea_cols
#> $country
#> [1] "Country"
#> 
#> $method
#> [1] "Method"
#> 
#> $energy_type
#> [1] "Energy.type"
#> 
#> $last_stage
#> [1] "Last.stage"
#> 
#> $year
#> [1] "Year"
#> 
#> $ledger_side
#> [1] "Ledger.side"
#> 
#> $flow_aggregation_point
#> [1] "Flow.aggregation.point"
#> 
#> $flow
#> [1] "Flow"
#> 
#> $product
#> [1] "Product"
#> 
#> $unit
#> [1] "Unit"
#> 
#> $e_dot
#> [1] "E.dot"
#>