Load IEA data from an extended enegy balances csv file

This function reads an IEA extended energy balances file and converts it to a data frame with appropriately-labelled columns. One of iea_file or text must be specified, but not both. The first line of iea_file or text is expected to start with expected_start_1st_line, and the second line is expected to start with expected_2nd_line_start, and it may have any number of commas appended. (The extra commas might come from opening and re-saving the file in Excel.) If those conditions are not met, execution is halted, and an error message is given.

iea_df(
  .iea_file = NULL,
  text = NULL,
  expected_1st_line_start = ",,TIME",
  expected_2nd_line_start = "COUNTRY,FLOW,PRODUCT",
  missing_data = "..",
  not_applicable_data = "x",
  confidential_data = "c",
  estimated_year = "E"
)

Arguments

.iea_file
a string containing the path to a .csv file of extended energy balances from the IEA. Default is the path to a sample IEA file provided in this package.

text
a character string that can be parsed as IEA extended energy balances. (This argument is useful for testing.)

expected_1st_line_start
the expected start of the first line of iea_file. Default is ",,TIME".

expected_2nd_line_start
the expected start of the second line of iea_file. Default is "COUNTRY,FLOW,PRODUCT".

missing_data
a string that identifies missing data. Default is "..". Entries of "missing_data" are coded as `0`` in output.

not_applicable_data
a string that identifies not-applicable data. Default is "x". Entries of "not_applicable_data" are coded as 0 in output.

confidential_data
a string that identifies confidential data. Default is "c". Entries of "confidential_data" are coded as 0 in output.

estimated_year
a string that identifies an estimated year. Default is "E". E.g., in "2014E", the "E" indicates that data for 2014 are estimated. Data from estimated years is removed from output.

Value

a data frame containing the IEA extended energy balances data

Details

This function is designed to work even as more years are added in columns at the right of .iea_file, because column names in the output are constructed from the first two lines of iea_file (which contain years and country, flow, product information).

The IEA data have indicators for not applicable values ("x") and for unavailable values (".."). (See "World Energy Balances: Database Documentation (2018 edition)" at http://wds.iea.org/wds/pdf/worldbal_documentation.pdf.) R has three concepts that could be used for "x" and "..": 0 would indicate value known to be zero. NULL would indicate an undefined value. NA would indicate a value that is unavailable. In theory, mapping from the IEA's indicators to R should proceed as follows: ".." (unavailable) in the IEA data would be converted to NA in R. "x" (not applicable) in the IEA data would be converted to 0 in R. "NULL" would not be used. However, the IEA are not consistent with their coding. In some places ".." is used for not applicable values (e.g., World Anthracite supply in 1971). World Anthracite supply in 1971 is actually not applicable, because Anthracite was classified under "Hard coal (if no detail)" in 1971. On the other hand, ".." is used for data in the most recent year when those data have not yet been incorporated into the database. In the face of IEA’s inconsistencies, the only rational way to proceed is to convert both "x" and ".." to "0" in this function. Furthermore, confidential data (coded by the IEA as "c") is also interpreted as 0. (What else can we do?)

The data frame returned from this function is not ready to be used in R, because rows are not unique. To further prepre the data frame for use, call augment_iea_data(), passing the output of this function in the .iea_df argument of augment_iea_data().

Examples

# Original file format
iea_df(text = ",,TIME,1960,1961\nCOUNTRY,FLOW,PRODUCT\nWorld,Production,Hard coal,42,43")

#>   COUNTRY       FLOW   PRODUCT 1960 1961
#> 1   World Production Hard coal   42   43

# With extra commas on the 2nd line
iea_df(text = ",,TIME,1960,1961\nCOUNTRY,FLOW,PRODUCT,,\nWorld,Production,Hard coal,42,43")

#>   COUNTRY       FLOW   PRODUCT 1960 1961
#> 1   World Production Hard coal   42   43