Perform quality assurance on a raw IEA data file

When starting to work with an IEA data file, it is important to verify its integrity. This function performs some validation tests on .iea_file.

iea_file_OK(
  .iea_file = NULL,
  text = NULL,
  country = "COUNTRY",
  flow = "FLOW",
  product = "PRODUCT",
  rowid = "rowid"
)

Arguments

.iea_file
the path to the raw IEA data file for which quality assurance is desired

text
a string containing text to be parsed as an IEA file.

country
the name of the country column. Default is "COUNTRY".

flow
the name of the flow column. Default is "FLOW".

product
the name of the product column. Default is "PRODUCT".

rowid
the name of a row number column added internally to .iea_file per country. Default is "rowid".

Value

TRUE if .iea_file passes all checks. Errors are thrown when a verification step fails.

Details

At this time, the only verification step performed by this function is confirming that every country has the same flow and product rows in the same order. The approach is to add a per-country row number column to the data frame and delete all the data in year columns. Then, the resulting data frame is queried for duplicate row numbers. If none are found, the function returns the data frame read from the file.

Note that .iea_file is read internally with data.table::fread() without stripping white space.

Examples

library(magrittr)

#> 
#> Attaching package: ‘magrittr’

#> The following objects are masked from ‘package:testthat’:
#> 
#>     equals, is_less_than, not

sample_iea_data_path() %>%
 iea_file_OK()

#> [1] TRUE