Load final-to-useful machine efficiency data

When performing extending an energy conversion chain from useful energy to final energy, efficiencies of final-to-useful energy conversion machines are defined by the analyst. The Excel file at path contains those allocations.

Usage

load_eta_fu_data(
  path = sample_eta_fu_table_path(),
  eta_fu_tab_name = IEATools::fu_analysis_file_info$eta_fu_tab_name
)

Arguments

path

the path from which final-to-useful machine efficiency data will be loaded. Default is the path to sample efficiency data supplied with this package.

eta_fu_tab_name

the tab in path that contains the final-to-useful machine efficiency data. Default is "FU etas".

Value

the eta_fu_tab_name tab in path as a data frame.

Details

A final-to-useful machine efficiencies template can be generated using eta_fu_template() and write_eta_fu_template().

A filled example can be loaded with the default value of path.

Examples

load_eta_fu_data()
#>     Country Method Energy.type Last.stage Unit                 Machine
#> 1       GHA    PCM           E      Final ktoe             Automobiles
#> 2       GHA    PCM           E      Final ktoe             Automobiles
#> 3       GHA    PCM           E      Final ktoe             Automobiles
#> 4       GHA    PCM           E      Final ktoe             Automobiles
#> 5       GHA    PCM           E      Final ktoe             Diesel cars
#> 6       GHA    PCM           E      Final ktoe             Diesel cars
#> 7       GHA    PCM           E      Final ktoe             Diesel cars
#> 8       GHA    PCM           E      Final ktoe             Diesel cars
#> 9       GHA    PCM           E      Final ktoe           Diesel trucks
#> 10      GHA    PCM           E      Final ktoe           Diesel trucks
#> 11      GHA    PCM           E      Final ktoe           Diesel trucks
#> 12      GHA    PCM           E      Final ktoe           Diesel trucks
#> 13      GHA    PCM           E      Final ktoe         Electric motors
#> 14      GHA    PCM           E      Final ktoe         Electric motors
#> 15      GHA    PCM           E      Final ktoe         Electric motors
#> 16      GHA    PCM           E      Final ktoe         Electric motors
#> 17      GHA    PCM           E      Final ktoe                  Trucks
#> 18      GHA    PCM           E      Final ktoe                  Trucks
#> 19      GHA    PCM           E      Final ktoe                  Trucks
#> 20      GHA    PCM           E      Final ktoe                  Trucks
#> 21      GHA    PCM           E      Final ktoe            Boat engines
#> 22      GHA    PCM           E      Final ktoe            Boat engines
#> 23      GHA    PCM           E      Final ktoe            Boat engines
#> 24      GHA    PCM           E      Final ktoe            Boat engines
#> 25      GHA    PCM           E      Final ktoe           Diesel trains
#> 26      GHA    PCM           E      Final ktoe           Diesel trains
#> 27      GHA    PCM           E      Final ktoe           Diesel trains
#> 28      GHA    PCM           E      Final ktoe           Diesel trains
#> 29      GHA    PCM           E      Final ktoe                    Fans
#> 30      GHA    PCM           E      Final ktoe                    Fans
#> 31      GHA    PCM           E      Final ktoe                    Fans
#> 32      GHA    PCM           E      Final ktoe                    Fans
#> 33      GHA    PCM           E      Final ktoe         Electric lights
#> 34      GHA    PCM           E      Final ktoe         Electric lights
#> 35      GHA    PCM           E      Final ktoe         Electric lights
#> 36      GHA    PCM           E      Final ktoe         Electric lights
#> 37      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 38      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 39      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 40      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 41      GHA    PCM           E      Final ktoe        Electric heaters
#> 42      GHA    PCM           E      Final ktoe        Electric heaters
#> 43      GHA    PCM           E      Final ktoe        Electric heaters
#> 44      GHA    PCM           E      Final ktoe        Electric heaters
#> 45      GHA    PCM           E      Final ktoe             Wood stoves
#> 46      GHA    PCM           E      Final ktoe             Wood stoves
#> 47      GHA    PCM           E      Final ktoe             Wood stoves
#> 48      GHA    PCM           E      Final ktoe             Wood stoves
#> 49      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 50      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 51      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 52      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 53      GHA    PCM           E      Final ktoe        Electric heaters
#> 54      GHA    PCM           E      Final ktoe        Electric heaters
#> 55      GHA    PCM           E      Final ktoe        Electric heaters
#> 56      GHA    PCM           E      Final ktoe        Electric heaters
#> 57      GHA    PCM           E      Final ktoe                   Irons
#> 58      GHA    PCM           E      Final ktoe                   Irons
#> 59      GHA    PCM           E      Final ktoe                   Irons
#> 60      GHA    PCM           E      Final ktoe                   Irons
#> 61      GHA    PCM           E      Final ktoe              LPG stoves
#> 62      GHA    PCM           E      Final ktoe              LPG stoves
#> 63      GHA    PCM           E      Final ktoe              LPG stoves
#> 64      GHA    PCM           E      Final ktoe              LPG stoves
#> 65      GHA    PCM           E      Final ktoe         Electric stoves
#> 66      GHA    PCM           E      Final ktoe         Electric stoves
#> 67      GHA    PCM           E      Final ktoe         Electric stoves
#> 68      GHA    PCM           E      Final ktoe         Electric stoves
#> 69      GHA    PCM           E      Final ktoe         Wood cookstoves
#> 70      GHA    PCM           E      Final ktoe         Wood cookstoves
#> 71      GHA    PCM           E      Final ktoe         Wood cookstoves
#> 72      GHA    PCM           E      Final ktoe         Wood cookstoves
#> 73      GHA    PCM           E      Final ktoe         Charcoal stoves
#> 74      GHA    PCM           E      Final ktoe         Charcoal stoves
#> 75      GHA    PCM           E      Final ktoe         Charcoal stoves
#> 76      GHA    PCM           E      Final ktoe         Charcoal stoves
#> 77      GHA    PCM           E      Final ktoe         Kerosene stoves
#> 78      GHA    PCM           E      Final ktoe         Kerosene stoves
#> 79      GHA    PCM           E      Final ktoe         Kerosene stoves
#> 80      GHA    PCM           E      Final ktoe         Kerosene stoves
#> 81      GHA    PCM           E      Final ktoe              LPG stoves
#> 82      GHA    PCM           E      Final ktoe              LPG stoves
#> 83      GHA    PCM           E      Final ktoe              LPG stoves
#> 84      GHA    PCM           E      Final ktoe              LPG stoves
#> 85      GHA    PCM           E      Final ktoe         Electric stoves
#> 86      GHA    PCM           E      Final ktoe         Electric stoves
#> 87      GHA    PCM           E      Final ktoe         Electric stoves
#> 88      GHA    PCM           E      Final ktoe         Electric stoves
#> 89      GHA    PCM           E      Final ktoe  Non-energy consumption
#> 90      GHA    PCM           E      Final ktoe  Non-energy consumption
#> 91      GHA    PCM           E      Final ktoe  Non-energy consumption
#> 92      GHA    PCM           E      Final ktoe  Non-energy consumption
#> 93      ZAF    PCM           E      Final ktoe  Steam engines (trains)
#> 94      ZAF    PCM           E      Final ktoe  Steam engines (trains)
#> 95      ZAF    PCM           E      Final ktoe  Steam engines (trains)
#> 96      ZAF    PCM           E      Final ktoe  Steam engines (trains)
#> 97      ZAF    PCM           E      Final ktoe         Electric motors
#> 98      ZAF    PCM           E      Final ktoe         Electric motors
#> 99      ZAF    PCM           E      Final ktoe         Electric motors
#> 100     ZAF    PCM           E      Final ktoe         Electric motors
#> 101     ZAF    PCM           E      Final ktoe                  Trucks
#> 102     ZAF    PCM           E      Final ktoe                  Trucks
#> 103     ZAF    PCM           E      Final ktoe                  Trucks
#> 104     ZAF    PCM           E      Final ktoe                  Trucks
#> 105     ZAF    PCM           E      Final ktoe             Automobiles
#> 106     ZAF    PCM           E      Final ktoe             Automobiles
#> 107     ZAF    PCM           E      Final ktoe             Automobiles
#> 108     ZAF    PCM           E      Final ktoe             Automobiles
#> 109     ZAF    PCM           E      Final ktoe  Industry static motors
#> 110     ZAF    PCM           E      Final ktoe  Industry static motors
#> 111     ZAF    PCM           E      Final ktoe  Industry static motors
#> 112     ZAF    PCM           E      Final ktoe  Industry static motors
#> 113     ZAF    PCM           E      Final ktoe           Diesel trucks
#> 114     ZAF    PCM           E      Final ktoe           Diesel trucks
#> 115     ZAF    PCM           E      Final ktoe           Diesel trucks
#> 116     ZAF    PCM           E      Final ktoe           Diesel trucks
#> 117     ZAF    PCM           E      Final ktoe      Diesel automobiles
#> 118     ZAF    PCM           E      Final ktoe      Diesel automobiles
#> 119     ZAF    PCM           E      Final ktoe      Diesel automobiles
#> 120     ZAF    PCM           E      Final ktoe      Diesel automobiles
#> 121     ZAF    PCM           E      Final ktoe                    Fans
#> 122     ZAF    PCM           E      Final ktoe                    Fans
#> 123     ZAF    PCM           E      Final ktoe                    Fans
#> 124     ZAF    PCM           E      Final ktoe                    Fans
#> 125     ZAF    PCM           E      Final ktoe               Airplanes
#> 126     ZAF    PCM           E      Final ktoe               Airplanes
#> 127     ZAF    PCM           E      Final ktoe               Airplanes
#> 128     ZAF    PCM           E      Final ktoe               Airplanes
#> 129     ZAF    PCM           E      Final ktoe Industry static engines
#> 130     ZAF    PCM           E      Final ktoe Industry static engines
#> 131     ZAF    PCM           E      Final ktoe Industry static engines
#> 132     ZAF    PCM           E      Final ktoe Industry static engines
#> 133     ZAF    PCM           E      Final ktoe    Electric automobiles
#> 134     ZAF    PCM           E      Final ktoe    Electric automobiles
#> 135     ZAF    PCM           E      Final ktoe    Electric automobiles
#> 136     ZAF    PCM           E      Final ktoe    Electric automobiles
#> 137     ZAF    PCM           E      Final ktoe            Boat engines
#> 138     ZAF    PCM           E      Final ktoe            Boat engines
#> 139     ZAF    PCM           E      Final ktoe            Boat engines
#> 140     ZAF    PCM           E      Final ktoe            Boat engines
#> 141     ZAF    PCM           E      Final ktoe              Generators
#> 142     ZAF    PCM           E      Final ktoe              Generators
#> 143     ZAF    PCM           E      Final ktoe              Generators
#> 144     ZAF    PCM           E      Final ktoe              Generators
#> 145     ZAF    PCM           E      Final ktoe         Electric lights
#> 146     ZAF    PCM           E      Final ktoe         Electric lights
#> 147     ZAF    PCM           E      Final ktoe         Electric lights
#> 148     ZAF    PCM           E      Final ktoe         Electric lights
#> 149     ZAF    PCM           E      Final ktoe         Kerosene lights
#> 150     ZAF    PCM           E      Final ktoe         Kerosene lights
#> 151     ZAF    PCM           E      Final ktoe         Kerosene lights
#> 152     ZAF    PCM           E      Final ktoe         Kerosene lights
#> 153     ZAF    PCM           E      Final ktoe              Gas lights
#> 154     ZAF    PCM           E      Final ktoe              Gas lights
#> 155     ZAF    PCM           E      Final ktoe              Gas lights
#> 156     ZAF    PCM           E      Final ktoe              Gas lights
#> 157     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 158     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 159     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 160     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 161     ZAF    PCM           E      Final ktoe        Electric heaters
#> 162     ZAF    PCM           E      Final ktoe        Electric heaters
#> 163     ZAF    PCM           E      Final ktoe        Electric heaters
#> 164     ZAF    PCM           E      Final ktoe        Electric heaters
#> 165     ZAF    PCM           E      Final ktoe             Gas heaters
#> 166     ZAF    PCM           E      Final ktoe             Gas heaters
#> 167     ZAF    PCM           E      Final ktoe             Gas heaters
#> 168     ZAF    PCM           E      Final ktoe             Gas heaters
#> 169     ZAF    PCM           E      Final ktoe             Oil heaters
#> 170     ZAF    PCM           E      Final ktoe             Oil heaters
#> 171     ZAF    PCM           E      Final ktoe             Oil heaters
#> 172     ZAF    PCM           E      Final ktoe             Oil heaters
#> 173     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 174     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 175     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 176     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 177     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 178     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 179     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 180     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 181     ZAF    PCM           E      Final ktoe             Wood stoves
#> 182     ZAF    PCM           E      Final ktoe             Wood stoves
#> 183     ZAF    PCM           E      Final ktoe             Wood stoves
#> 184     ZAF    PCM           E      Final ktoe             Wood stoves
#> 185     ZAF    PCM           E      Final ktoe                   Irons
#> 186     ZAF    PCM           E      Final ktoe                   Irons
#> 187     ZAF    PCM           E      Final ktoe                   Irons
#> 188     ZAF    PCM           E      Final ktoe                   Irons
#> 189     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 190     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 191     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 192     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 193     ZAF    PCM           E      Final ktoe             Oil heaters
#> 194     ZAF    PCM           E      Final ktoe             Oil heaters
#> 195     ZAF    PCM           E      Final ktoe             Oil heaters
#> 196     ZAF    PCM           E      Final ktoe             Oil heaters
#> 197     ZAF    PCM           E      Final ktoe             Gas heaters
#> 198     ZAF    PCM           E      Final ktoe             Gas heaters
#> 199     ZAF    PCM           E      Final ktoe             Gas heaters
#> 200     ZAF    PCM           E      Final ktoe             Gas heaters
#> 201     ZAF    PCM           E      Final ktoe        Electric heaters
#> 202     ZAF    PCM           E      Final ktoe        Electric heaters
#> 203     ZAF    PCM           E      Final ktoe        Electric heaters
#> 204     ZAF    PCM           E      Final ktoe        Electric heaters
#> 205     ZAF    PCM           E      Final ktoe              LPG stoves
#> 206     ZAF    PCM           E      Final ktoe              LPG stoves
#> 207     ZAF    PCM           E      Final ktoe              LPG stoves
#> 208     ZAF    PCM           E      Final ktoe              LPG stoves
#> 209     ZAF    PCM           E      Final ktoe             Coal stoves
#> 210     ZAF    PCM           E      Final ktoe             Coal stoves
#> 211     ZAF    PCM           E      Final ktoe             Coal stoves
#> 212     ZAF    PCM           E      Final ktoe             Coal stoves
#> 213     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 214     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 215     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 216     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 217     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 218     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 219     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 220     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 221     ZAF    PCM           E      Final ktoe         Wood cookstoves
#> 222     ZAF    PCM           E      Final ktoe         Wood cookstoves
#> 223     ZAF    PCM           E      Final ktoe         Wood cookstoves
#> 224     ZAF    PCM           E      Final ktoe         Wood cookstoves
#> 225     ZAF    PCM           E      Final ktoe         Charcoal stoves
#> 226     ZAF    PCM           E      Final ktoe         Charcoal stoves
#> 227     ZAF    PCM           E      Final ktoe         Charcoal stoves
#> 228     ZAF    PCM           E      Final ktoe         Charcoal stoves
#> 229     ZAF    PCM           E      Final ktoe          LPG cookstoves
#> 230     ZAF    PCM           E      Final ktoe          LPG cookstoves
#> 231     ZAF    PCM           E      Final ktoe          LPG cookstoves
#> 232     ZAF    PCM           E      Final ktoe          LPG cookstoves
#> 233     ZAF    PCM           E      Final ktoe              LPG stoves
#> 234     ZAF    PCM           E      Final ktoe              LPG stoves
#> 235     ZAF    PCM           E      Final ktoe              LPG stoves
#> 236     ZAF    PCM           E      Final ktoe              LPG stoves
#> 237     ZAF    PCM           E      Final ktoe             Coal stoves
#> 238     ZAF    PCM           E      Final ktoe             Coal stoves
#> 239     ZAF    PCM           E      Final ktoe             Coal stoves
#> 240     ZAF    PCM           E      Final ktoe             Coal stoves
#> 241     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 242     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 243     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 244     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 245     ZAF    PCM           E      Final ktoe             Gas heaters
#> 246     ZAF    PCM           E      Final ktoe             Gas heaters
#> 247     ZAF    PCM           E      Final ktoe             Gas heaters
#> 248     ZAF    PCM           E      Final ktoe             Gas heaters
#> 249     ZAF    PCM           E      Final ktoe             Coal stoves
#> 250     ZAF    PCM           E      Final ktoe             Coal stoves
#> 251     ZAF    PCM           E      Final ktoe             Coal stoves
#> 252     ZAF    PCM           E      Final ktoe             Coal stoves
#> 253     ZAF    PCM           E      Final ktoe           Wood furnaces
#> 254     ZAF    PCM           E      Final ktoe           Wood furnaces
#> 255     ZAF    PCM           E      Final ktoe           Wood furnaces
#> 256     ZAF    PCM           E      Final ktoe           Wood furnaces
#> 257     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 258     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 259     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 260     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 261     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 262     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 263     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 264     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 265     ZAF    PCM           E      Final ktoe              LPG stoves
#> 266     ZAF    PCM           E      Final ktoe              LPG stoves
#> 267     ZAF    PCM           E      Final ktoe              LPG stoves
#> 268     ZAF    PCM           E      Final ktoe              LPG stoves
#> 269     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 270     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 271     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 272     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 273     ZAF    PCM           E      Final ktoe  Non-energy consumption
#> 274     ZAF    PCM           E      Final ktoe  Non-energy consumption
#> 275     ZAF    PCM           E      Final ktoe  Non-energy consumption
#> 276     ZAF    PCM           E      Final ktoe  Non-energy consumption
#>     Eu.product          Quantity Maximum.values         1971         2000
#> 1           MD     E.dot_machine   6.192510e+02 2.172315e+02 6.192510e+02
#> 2           MD E.dot_machine [%]   1.156388e-01 8.391282e-02 1.156388e-01
#> 3           MD            eta.fu             NA 3.000000e-01 3.100000e-01
#> 4           MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 5           MD     E.dot_machine   1.877078e+02 5.946665e+01 1.877078e+02
#> 6           MD E.dot_machine [%]   3.505250e-02 2.297095e-02 3.505250e-02
#> 7           MD            eta.fu             NA 3.000000e-01 3.000000e-01
#> 8           MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 9           MD     E.dot_machine   1.877078e+02 5.946665e+01 1.877078e+02
#> 10          MD E.dot_machine [%]   3.505250e-02 2.297095e-02 3.505250e-02
#> 11          MD            eta.fu             NA 3.000000e-01 3.000000e-01
#> 12          MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 13          MD     E.dot_machine   1.870163e+02 6.383495e+01 1.870163e+02
#> 14          MD E.dot_machine [%]   3.492338e-02 2.465835e-02 3.492338e-02
#> 15          MD            eta.fu             NA 8.500000e-01 9.500000e-01
#> 16          MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 17          MD     E.dot_machine   7.032580e+01 1.499595e+01 7.032580e+01
#> 18          MD E.dot_machine [%]   1.313262e-02 5.792678e-03 1.313262e-02
#> 19          MD            eta.fu             NA 3.000000e-01 3.100000e-01
#> 20          MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 21          MD     E.dot_machine   6.316280e+01 1.226230e+01 6.316280e+01
#> 22          MD E.dot_machine [%]   1.179500e-02 4.736716e-03 1.179500e-02
#> 23          MD            eta.fu             NA 3.000000e-01 3.100000e-01
#> 24          MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 25          MD     E.dot_machine   2.585510e+01 1.514280e+01 2.585510e+01
#> 26          MD E.dot_machine [%]   5.849404e-03 5.849404e-03 4.828174e-03
#> 27          MD            eta.fu             NA 3.000000e-01 3.100000e-01
#> 28          MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 29          MD     E.dot_machine   1.907567e+01 2.115225e+00 1.907567e+01
#> 30          MD E.dot_machine [%]   3.562184e-03 8.170751e-04 3.562184e-03
#> 31          MD            eta.fu             NA 1.000000e-01 1.200000e-01
#> 32          MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 33       Light     E.dot_machine   1.589037e+02 1.411179e+02 1.589037e+02
#> 34       Light E.dot_machine [%]   5.451143e-02 5.451143e-02 2.967363e-02
#> 35       Light            eta.fu             NA 2.000000e-01 2.000000e-01
#> 36       Light             phi.u             NA 1.261201e-01 1.830161e-01
#> 37   HTH.600.C     E.dot_machine   2.642350e+01 1.953040e+01 2.642350e+01
#> 38   HTH.600.C E.dot_machine [%]   7.544259e-03 7.544259e-03 4.934317e-03
#> 39   HTH.600.C            eta.fu             NA 8.500000e-01 8.600000e-01
#> 40   HTH.600.C             phi.u             NA 6.585352e-01 6.585352e-01
#> 41   HTH.600.C     E.dot_machine   1.584695e+01 1.584695e+01           NA
#> 42   HTH.600.C E.dot_machine [%]   6.121405e-03 6.121405e-03           NA
#> 43   HTH.600.C            eta.fu             NA 8.500000e-01           NA
#> 44   HTH.600.C             phi.u             NA 6.585352e-01           NA
#> 45   MTH.200.C     E.dot_machine   6.852728e+02 1.456960e+02 6.852728e+02
#> 46   MTH.200.C E.dot_machine [%]   1.279677e-01 5.627987e-02 1.279677e-01
#> 47   MTH.200.C            eta.fu             NA 8.500000e-01 8.600000e-01
#> 48   MTH.200.C             phi.u             NA 3.698616e-01 3.698616e-01
#> 49   MTH.200.C     E.dot_machine   2.166498e+02 1.369423e+02 2.166498e+02
#> 50   MTH.200.C E.dot_machine [%]   5.289846e-02 5.289846e-02 4.045712e-02
#> 51   MTH.200.C            eta.fu             NA 8.500000e-01 8.600000e-01
#> 52   MTH.200.C             phi.u             NA 3.698616e-01 3.698616e-01
#> 53   MTH.200.C     E.dot_machine   1.480998e+02 9.355120e+00 1.480998e+02
#> 54   MTH.200.C E.dot_machine [%]   2.765611e-02 3.613722e-03 2.765611e-02
#> 55   MTH.200.C            eta.fu             NA 9.700000e-01 9.900000e-01
#> 56   MTH.200.C             phi.u             NA 3.698616e-01 3.698616e-01
#> 57   MTH.200.C     E.dot_machine   1.907567e+01 2.115225e+00 1.907567e+01
#> 58   MTH.200.C E.dot_machine [%]   3.562184e-03 8.170751e-04 3.562184e-03
#> 59   MTH.200.C            eta.fu             NA 8.500000e-01 8.500000e-01
#> 60   MTH.200.C             phi.u             NA 3.698616e-01 3.698616e-01
#> 61   MTH.200.C     E.dot_machine   4.519000e+00           NA 4.519000e+00
#> 62   MTH.200.C E.dot_machine [%]   8.438768e-04           NA 8.438768e-04
#> 63   MTH.200.C            eta.fu             NA           NA 4.800000e-01
#> 64   MTH.200.C             phi.u             NA           NA 3.698616e-01
#> 65   MTH.200.C     E.dot_machine   1.160800e+00 1.160800e+00           NA
#> 66   MTH.200.C E.dot_machine [%]   4.483971e-04 4.483971e-04           NA
#> 67   MTH.200.C            eta.fu             NA 6.500000e-01           NA
#> 68   MTH.200.C             phi.u             NA 3.698616e-01           NA
#> 69   MTH.100.C     E.dot_machine   2.111732e+03 1.464125e+03 2.111732e+03
#> 70   MTH.100.C E.dot_machine [%]   5.655665e-01 5.655665e-01 3.943443e-01
#> 71   MTH.100.C            eta.fu             NA 1.200000e-01 1.200000e-01
#> 72   MTH.100.C             phi.u             NA 2.009916e-01 2.009916e-01
#> 73   MTH.100.C     E.dot_machine   5.156874e+02 1.191745e+02 5.156874e+02
#> 74   MTH.100.C E.dot_machine [%]   9.629932e-02 4.603507e-02 9.629932e-02
#> 75   MTH.100.C            eta.fu             NA 1.400000e-01 1.400000e-01
#> 76   MTH.100.C             phi.u             NA 2.009916e-01 2.009916e-01
#> 77   MTH.100.C     E.dot_machine   8.369160e+01 8.369160e+01 5.335340e+01
#> 78   MTH.100.C E.dot_machine [%]   3.232863e-02 3.232863e-02 9.963199e-03
#> 79   MTH.100.C            eta.fu             NA 1.600000e-01 1.600000e-01
#> 80   MTH.100.C             phi.u             NA 2.009916e-01 2.009916e-01
#> 81   MTH.100.C     E.dot_machine   3.615170e+01 3.389200e+00 3.615170e+01
#> 82   MTH.100.C E.dot_machine [%]   6.750959e-03 1.309190e-03 6.750959e-03
#> 83   MTH.100.C            eta.fu             NA 1.800000e-01 1.800000e-01
#> 84   MTH.100.C             phi.u             NA 2.009916e-01 2.009916e-01
#> 85   MTH.100.C     E.dot_machine   1.907567e+01 2.115225e+00 1.907567e+01
#> 86   MTH.100.C E.dot_machine [%]   3.562184e-03 8.170751e-04 3.562184e-03
#> 87   MTH.100.C            eta.fu             NA 7.000000e-01 7.000000e-01
#> 88   MTH.100.C             phi.u             NA 2.009916e-01 2.009916e-01
#> 89         NEU     E.dot_machine   6.262540e+01 1.805670e+01 6.821440e+01
#> 90         NEU E.dot_machine [%]   1.154755e-02 6.926680e-03 1.257811e-02
#> 91         NEU            eta.fu             NA 1.000000e-06 1.000000e-06
#> 92         NEU             phi.u             NA 1.000000e+00 1.000000e+00
#> 93          MD     E.dot_machine   3.295513e+03 3.295513e+03 1.926513e+02
#> 94          MD E.dot_machine [%]   9.922186e-02 9.922186e-02 3.670736e-03
#> 95          MD            eta.fu             NA 6.000000e-01 6.200000e-01
#> 96          MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 97          MD     E.dot_machine   5.168693e+03 1.030447e+03 5.168693e+03
#> 98          MD E.dot_machine [%]   9.848315e-02 3.102487e-02 9.848315e-02
#> 99          MD            eta.fu             NA 8.700000e-01 9.600000e-01
#> 100         MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 101         MD     E.dot_machine   4.109393e+03 1.850616e+03 4.109393e+03
#> 102         MD E.dot_machine [%]   7.829947e-02 5.571866e-02 7.829947e-02
#> 103         MD            eta.fu             NA 3.000000e-01 3.100000e-01
#> 104         MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 105         MD     E.dot_machine   4.029254e+03 2.305665e+03 4.029254e+03
#> 106         MD E.dot_machine [%]   7.677253e-02 6.941935e-02 7.677253e-02
#> 107         MD            eta.fu             NA 3.000000e-01 3.100000e-01
#> 108         MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 109         MD     E.dot_machine   2.496819e+03 1.190714e+03 2.496819e+03
#> 110         MD E.dot_machine [%]   4.757384e-02 3.585020e-02 4.757384e-02
#> 111         MD            eta.fu             NA 3.000000e-01 3.100000e-01
#> 112         MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 113         MD     E.dot_machine   1.673811e+03 7.115315e+02 1.673811e+03
#> 114         MD E.dot_machine [%]   3.189242e-02 2.142291e-02 3.189242e-02
#> 115         MD            eta.fu             NA 3.000000e-01 3.100000e-01
#> 116         MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 117         MD     E.dot_machine   1.672812e+03 7.115315e+02 1.672812e+03
#> 118         MD E.dot_machine [%]   3.187339e-02 2.142291e-02 3.187339e-02
#> 119         MD            eta.fu             NA 3.000000e-01 3.100000e-01
#> 120         MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 121         MD     E.dot_machine   7.398108e+02 2.086328e+02 7.398108e+02
#> 122         MD E.dot_machine [%]   1.409619e-02 6.281552e-03 1.409619e-02
#> 123         MD            eta.fu             NA 1.200000e-01 1.200000e-01
#> 124         MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 125         MD     E.dot_machine   6.141134e+02 1.715057e+02 6.141134e+02
#> 126         MD E.dot_machine [%]   1.170118e-02 5.163722e-03 1.170118e-02
#> 127         MD            eta.fu             NA 3.000000e-01 3.100000e-01
#> 128         MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 129         MD     E.dot_machine   5.185681e+02           NA 5.185681e+02
#> 130         MD E.dot_machine [%]   9.880684e-03           NA 9.880684e-03
#> 131         MD            eta.fu             NA           NA 3.100000e-01
#> 132         MD             phi.u             NA           NA 1.000000e+00
#> 133         MD     E.dot_machine   1.620809e+02 5.176270e+01 1.620809e+02
#> 134         MD E.dot_machine [%]   3.088254e-03 1.558480e-03 3.088254e-03
#> 135         MD            eta.fu             NA 4.500000e-01 5.000000e-01
#> 136         MD             phi.u             NA 1.000000e+00 1.000000e+00
#> 137         MD     E.dot_machine   1.026154e+02           NA 1.026154e+02
#> 138         MD E.dot_machine [%]   1.955211e-03           NA 1.955211e-03
#> 139         MD            eta.fu             NA           NA 3.100000e-01
#> 140         MD             phi.u             NA           NA 1.000000e+00
#> 141         MD     E.dot_machine   3.366390e+01           NA 3.366390e+01
#> 142         MD E.dot_machine [%]   6.414246e-04           NA 6.414246e-04
#> 143         MD            eta.fu             NA           NA 3.100000e-01
#> 144         MD             phi.u             NA           NA 1.000000e+00
#> 145      Light     E.dot_machine   6.263618e+03 1.413929e+03 6.263618e+03
#> 146      Light E.dot_machine [%]   1.193456e-01 4.257082e-02 1.193456e-01
#> 147      Light            eta.fu             NA 2.000000e-01 2.000000e-01
#> 148      Light             phi.u             NA 1.261201e-01 1.830161e-01
#> 149      Light     E.dot_machine   1.497862e+02           NA 1.497862e+02
#> 150      Light E.dot_machine [%]   2.853994e-03           NA 2.853994e-03
#> 151      Light            eta.fu             NA           NA 2.000000e-01
#> 152      Light             phi.u             NA           NA 1.830161e-01
#> 153      Light     E.dot_machine   1.154460e+02 4.325020e+01 1.154460e+02
#> 154      Light E.dot_machine [%]   2.199684e-03 1.302184e-03 2.199684e-03
#> 155      Light            eta.fu             NA 2.000000e-01 2.000000e-01
#> 156      Light             phi.u             NA 1.261201e-01 1.830161e-01
#> 157  HTH.600.C     E.dot_machine   7.157481e+03 7.157481e+03 4.612028e+03
#> 158  HTH.600.C E.dot_machine [%]   2.154986e-01 2.154986e-01 8.787657e-02
#> 159  HTH.600.C            eta.fu             NA 8.500000e-01 9.000000e-01
#> 160  HTH.600.C             phi.u             NA 6.585352e-01 6.585352e-01
#> 161  HTH.600.C     E.dot_machine   1.683990e+03 7.889936e+01 1.683990e+03
#> 162  HTH.600.C E.dot_machine [%]   3.208637e-02 2.375515e-03 3.208637e-02
#> 163  HTH.600.C            eta.fu             NA 9.000000e-01 9.500000e-01
#> 164  HTH.600.C             phi.u             NA 6.585352e-01 6.585352e-01
#> 165  HTH.600.C     E.dot_machine   1.052345e+03           NA 1.052345e+03
#> 166  HTH.600.C E.dot_machine [%]   2.005116e-02           NA 2.005116e-02
#> 167  HTH.600.C            eta.fu             NA           NA 8.500000e-01
#> 168  HTH.600.C             phi.u             NA           NA 6.585352e-01
#> 169  HTH.600.C     E.dot_machine   5.799369e+02 5.799369e+02           NA
#> 170  HTH.600.C E.dot_machine [%]   1.746084e-02 1.746084e-02           NA
#> 171  HTH.600.C            eta.fu             NA 8.000000e-01           NA
#> 172  HTH.600.C             phi.u             NA 6.585352e-01           NA
#> 173  HTH.600.C     E.dot_machine   1.706363e+02 2.914880e+01 1.706363e+02
#> 174  HTH.600.C E.dot_machine [%]   3.251267e-03 8.776168e-04 3.251267e-03
#> 175  HTH.600.C            eta.fu             NA 8.000000e-01 8.500000e-01
#> 176  HTH.600.C             phi.u             NA 6.585352e-01 6.585352e-01
#> 177  MTH.200.C     E.dot_machine   3.459107e+03 2.080060e+03 3.459107e+03
#> 178  MTH.200.C E.dot_machine [%]   6.590907e-02 6.262680e-02 6.590907e-02
#> 179  MTH.200.C            eta.fu             NA 8.000000e-01 8.500000e-01
#> 180  MTH.200.C             phi.u             NA 3.698616e-01 3.698616e-01
#> 181  MTH.200.C     E.dot_machine   1.508431e+03 7.643069e+02 1.508431e+03
#> 182  MTH.200.C E.dot_machine [%]   2.874132e-02 2.301188e-02 2.874132e-02
#> 183  MTH.200.C            eta.fu             NA 2.000000e-01 2.500000e-01
#> 184  MTH.200.C             phi.u             NA 3.698616e-01 3.698616e-01
#> 185  MTH.200.C     E.dot_machine   1.233018e+03 3.477214e+02 1.233018e+03
#> 186  MTH.200.C E.dot_machine [%]   2.349366e-02 1.046925e-02 2.349366e-02
#> 187  MTH.200.C            eta.fu             NA 3.000000e-01 3.500000e-01
#> 188  MTH.200.C             phi.u             NA 3.698616e-01 3.698616e-01
#> 189  MTH.200.C     E.dot_machine   5.596876e+02 5.596876e+02 5.165000e+00
#> 190  MTH.200.C E.dot_machine [%]   1.685117e-02 1.685117e-02 9.841278e-05
#> 191  MTH.200.C            eta.fu             NA 1.400000e-01 1.400000e-01
#> 192  MTH.200.C             phi.u             NA 3.698616e-01 3.698616e-01
#> 193  MTH.200.C     E.dot_machine   5.235502e+02 5.235502e+02           NA
#> 194  MTH.200.C E.dot_machine [%]   1.576314e-02 1.576314e-02           NA
#> 195  MTH.200.C            eta.fu             NA 4.000000e-01           NA
#> 196  MTH.200.C             phi.u             NA 3.698616e-01           NA
#> 197  MTH.200.C     E.dot_machine   5.746345e+02 6.605760e+01 5.746345e+02
#> 198  MTH.200.C E.dot_machine [%]   1.094896e-02 1.988873e-03 1.094896e-02
#> 199  MTH.200.C            eta.fu             NA 4.000000e-01 4.500000e-01
#> 200  MTH.200.C             phi.u             NA 3.698616e-01 3.698616e-01
#> 201  MTH.200.C     E.dot_machine   4.777397e+02 4.315563e+01 4.777397e+02
#> 202  MTH.200.C E.dot_machine [%]   9.102749e-03 1.299337e-03 9.102749e-03
#> 203  MTH.200.C            eta.fu             NA 4.000000e-01 4.500000e-01
#> 204  MTH.200.C             phi.u             NA 3.698616e-01 3.698616e-01
#> 205  MTH.200.C     E.dot_machine   1.197526e+02 1.197526e+02 2.680800e+00
#> 206  MTH.200.C E.dot_machine [%]   3.605531e-03 3.605531e-03 5.107938e-05
#> 207  MTH.200.C            eta.fu             NA 3.000000e-01 3.500000e-01
#> 208  MTH.200.C             phi.u             NA 3.698616e-01 3.698616e-01
#> 209  MTH.200.C     E.dot_machine   7.672980e+01 7.672980e+01           NA
#> 210  MTH.200.C E.dot_machine [%]   2.310193e-03 2.310193e-03           NA
#> 211  MTH.200.C            eta.fu             NA 2.500000e-01           NA
#> 212  MTH.200.C             phi.u             NA 3.698616e-01           NA
#> 213  MTH.200.C     E.dot_machine   1.098900e+01           NA 1.098900e+01
#> 214  MTH.200.C E.dot_machine [%]   2.093820e-04           NA 2.093820e-04
#> 215  MTH.200.C            eta.fu             NA           NA 9.000000e-01
#> 216  MTH.200.C             phi.u             NA           NA 3.698616e-01
#> 217  MTH.200.C     E.dot_machine   3.075000e+00           NA 3.075000e+00
#> 218  MTH.200.C E.dot_machine [%]   5.859038e-05           NA 5.859038e-05
#> 219  MTH.200.C            eta.fu             NA           NA 9.000000e-01
#> 220  MTH.200.C             phi.u             NA           NA 3.698616e-01
#> 221  MTH.100.C     E.dot_machine   2.787475e+03 1.934652e+03 2.787475e+03
#> 222  MTH.100.C E.dot_machine [%]   5.824881e-02 5.824881e-02 5.311194e-02
#> 223  MTH.100.C            eta.fu             NA 1.200000e-01 1.200000e-01
#> 224  MTH.100.C             phi.u             NA 2.009916e-01 2.009916e-01
#> 225  MTH.100.C     E.dot_machine   6.584026e+02 1.691980e+01 6.584026e+02
#> 226  MTH.100.C E.dot_machine [%]   1.254506e-02 5.094241e-04 1.254506e-02
#> 227  MTH.100.C            eta.fu             NA 1.400000e-01 1.400000e-01
#> 228  MTH.100.C             phi.u             NA 2.009916e-01 2.009916e-01
#> 229  MTH.100.C     E.dot_machine   1.452114e+02           NA 1.452114e+02
#> 230  MTH.100.C E.dot_machine [%]   2.766826e-03           NA 2.766826e-03
#> 231  MTH.100.C            eta.fu             NA           NA 1.600000e-01
#> 232  MTH.100.C             phi.u             NA           NA 2.009916e-01
#> 233  MTH.100.C     E.dot_machine   4.691440e+01           NA 4.691440e+01
#> 234  MTH.100.C E.dot_machine [%]   8.938968e-04           NA 8.938968e-04
#> 235  MTH.100.C            eta.fu             NA           NA 1.600000e-01
#> 236  MTH.100.C             phi.u             NA           NA 2.009916e-01
#> 237  MTH.100.C     E.dot_machine   4.449040e+01           NA 4.449040e+01
#> 238  MTH.100.C E.dot_machine [%]   8.477104e-04           NA 8.477104e-04
#> 239  MTH.100.C            eta.fu             NA           NA 1.600000e-01
#> 240  MTH.100.C             phi.u             NA           NA 2.009916e-01
#> 241  MTH.100.C     E.dot_machine   3.099000e+00           NA 3.099000e+00
#> 242  MTH.100.C E.dot_machine [%]   5.904767e-05           NA 5.904767e-05
#> 243  MTH.100.C            eta.fu             NA           NA 1.600000e-01
#> 244  MTH.100.C             phi.u             NA           NA 2.009916e-01
#> 245  MTH.100.C     E.dot_machine   5.223600e-01           NA 5.223600e-01
#> 246  MTH.100.C E.dot_machine [%]   9.952934e-06           NA 9.952934e-06
#> 247  MTH.100.C            eta.fu             NA           NA 9.000000e-01
#> 248  MTH.100.C             phi.u             NA           NA 2.009916e-01
#> 249   LTH.20.C     E.dot_machine   3.140119e+03 3.140119e+03 1.490106e+03
#> 250   LTH.20.C E.dot_machine [%]   9.454323e-02 9.454323e-02 2.839215e-02
#> 251   LTH.20.C            eta.fu             NA 1.600000e-01 1.600000e-01
#> 252   LTH.20.C             phi.u             NA 1.677008e-02 1.677008e-02
#> 253   LTH.20.C     E.dot_machine   2.787475e+03 1.934652e+03 2.787475e+03
#> 254   LTH.20.C E.dot_machine [%]   5.824881e-02 5.824881e-02 5.311194e-02
#> 255   LTH.20.C            eta.fu             NA 8.500000e-01 8.500000e-01
#> 256   LTH.20.C             phi.u             NA 1.677008e-02 1.677008e-02
#> 257   LTH.20.C     E.dot_machine   9.715247e+02 7.756520e+02 9.715247e+02
#> 258   LTH.20.C E.dot_machine [%]   2.335346e-02 2.335346e-02 1.851122e-02
#> 259   LTH.20.C            eta.fu             NA 8.500000e-01 8.500000e-01
#> 260   LTH.20.C             phi.u             NA 1.677008e-02 1.677008e-02
#> 261   LTH.20.C     E.dot_machine   5.516266e+02           NA 5.516266e+02
#> 262   LTH.20.C E.dot_machine [%]   1.051057e-02           NA 1.051057e-02
#> 263   LTH.20.C            eta.fu             NA           NA 3.000000e-01
#> 264   LTH.20.C             phi.u             NA           NA 1.677008e-02
#> 265   LTH.20.C     E.dot_machine   1.452114e+02           NA 1.452114e+02
#> 266   LTH.20.C E.dot_machine [%]   2.766826e-03           NA 2.766826e-03
#> 267   LTH.20.C            eta.fu             NA           NA 3.200000e-01
#> 268   LTH.20.C             phi.u             NA           NA 1.677008e-02
#> 269   LTH.20.C     E.dot_machine   1.398600e+01           NA 1.398600e+01
#> 270   LTH.20.C E.dot_machine [%]   2.664862e-04           NA 2.664862e-04
#> 271   LTH.20.C            eta.fu             NA           NA 8.500000e-01
#> 272   LTH.20.C             phi.u             NA           NA 1.677008e-02
#> 273        NEU     E.dot_machine   5.654902e+03 9.323913e+02 5.654902e+03
#> 274        NEU E.dot_machine [%]   9.726590e-02 2.730604e-02 9.726590e-02
#> 275        NEU            eta.fu             NA 1.000000e-06 1.000000e-06
#> 276        NEU             phi.u             NA 1.000000e+00 1.000000e+00