Load final-to-useful machine efficiency data

When performing extending an energy conversion chain from useful energy to final energy, efficiencies of final-to-useful energy conversion machines are defined by the analyst. The Excel file at path contains those allocations.

load_eta_fu_data(
  path = sample_eta_fu_table_path(),
  eta_fu_tab_name = IEATools::fu_analysis_file_info$eta_fu_tab_name
)

Arguments

path
the path from which final-to-useful machine efficiency data will be loaded. Default is the path to sample efficiency data supplied with this package.

eta_fu_tab_name
the tab in path that contains the final-to-useful machine efficiency data. Default is "FU etas".

Value

the eta_fu_tab_name tab in path as a data frame.

Details

A final-to-useful machine efficiencies template can be generated using eta_fu_template() and write_eta_fu_template().

A filled example can be loaded with the default value of path.

Examples

load_eta_fu_data()

#>     Country Method Energy.type Last.stage Unit                 Machine
#> 1       GHA    PCM           E      Final ktoe             Automobiles
#> 2       GHA    PCM           E      Final ktoe             Automobiles
#> 3       GHA    PCM           E      Final ktoe             Automobiles
#> 4       GHA    PCM           E      Final ktoe             Automobiles
#> 5       GHA    PCM           E      Final ktoe             Diesel cars
#> 6       GHA    PCM           E      Final ktoe             Diesel cars
#> 7       GHA    PCM           E      Final ktoe             Diesel cars
#> 8       GHA    PCM           E      Final ktoe             Diesel cars
#> 9       GHA    PCM           E      Final ktoe           Diesel trucks
#> 10      GHA    PCM           E      Final ktoe           Diesel trucks
#> 11      GHA    PCM           E      Final ktoe           Diesel trucks
#> 12      GHA    PCM           E      Final ktoe           Diesel trucks
#> 13      GHA    PCM           E      Final ktoe         Electric motors
#> 14      GHA    PCM           E      Final ktoe         Electric motors
#> 15      GHA    PCM           E      Final ktoe         Electric motors
#> 16      GHA    PCM           E      Final ktoe         Electric motors
#> 17      GHA    PCM           E      Final ktoe                  Trucks
#> 18      GHA    PCM           E      Final ktoe                  Trucks
#> 19      GHA    PCM           E      Final ktoe                  Trucks
#> 20      GHA    PCM           E      Final ktoe                  Trucks
#> 21      GHA    PCM           E      Final ktoe            Boat engines
#> 22      GHA    PCM           E      Final ktoe            Boat engines
#> 23      GHA    PCM           E      Final ktoe            Boat engines
#> 24      GHA    PCM           E      Final ktoe            Boat engines
#> 25      GHA    PCM           E      Final ktoe           Diesel trains
#> 26      GHA    PCM           E      Final ktoe           Diesel trains
#> 27      GHA    PCM           E      Final ktoe           Diesel trains
#> 28      GHA    PCM           E      Final ktoe           Diesel trains
#> 29      GHA    PCM           E      Final ktoe                    Fans
#> 30      GHA    PCM           E      Final ktoe                    Fans
#> 31      GHA    PCM           E      Final ktoe                    Fans
#> 32      GHA    PCM           E      Final ktoe                    Fans
#> 33      GHA    PCM           E      Final ktoe         Electric lights
#> 34      GHA    PCM           E      Final ktoe         Electric lights
#> 35      GHA    PCM           E      Final ktoe         Electric lights
#> 36      GHA    PCM           E      Final ktoe         Electric lights
#> 37      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 38      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 39      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 40      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 41      GHA    PCM           E      Final ktoe        Electric heaters
#> 42      GHA    PCM           E      Final ktoe        Electric heaters
#> 43      GHA    PCM           E      Final ktoe        Electric heaters
#> 44      GHA    PCM           E      Final ktoe        Electric heaters
#> 45      GHA    PCM           E      Final ktoe             Wood stoves
#> 46      GHA    PCM           E      Final ktoe             Wood stoves
#> 47      GHA    PCM           E      Final ktoe             Wood stoves
#> 48      GHA    PCM           E      Final ktoe             Wood stoves
#> 49      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 50      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 51      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 52      GHA    PCM           E      Final ktoe     Industrial furnaces
#> 53      GHA    PCM           E      Final ktoe        Electric heaters
#> 54      GHA    PCM           E      Final ktoe        Electric heaters
#> 55      GHA    PCM           E      Final ktoe        Electric heaters
#> 56      GHA    PCM           E      Final ktoe        Electric heaters
#> 57      GHA    PCM           E      Final ktoe                   Irons
#> 58      GHA    PCM           E      Final ktoe                   Irons
#> 59      GHA    PCM           E      Final ktoe                   Irons
#> 60      GHA    PCM           E      Final ktoe                   Irons
#> 61      GHA    PCM           E      Final ktoe              LPG stoves
#> 62      GHA    PCM           E      Final ktoe              LPG stoves
#> 63      GHA    PCM           E      Final ktoe              LPG stoves
#> 64      GHA    PCM           E      Final ktoe              LPG stoves
#> 65      GHA    PCM           E      Final ktoe         Electric stoves
#> 66      GHA    PCM           E      Final ktoe         Electric stoves
#> 67      GHA    PCM           E      Final ktoe         Electric stoves
#> 68      GHA    PCM           E      Final ktoe         Electric stoves
#> 69      GHA    PCM           E      Final ktoe         Wood cookstoves
#> 70      GHA    PCM           E      Final ktoe         Wood cookstoves
#> 71      GHA    PCM           E      Final ktoe         Wood cookstoves
#> 72      GHA    PCM           E      Final ktoe         Wood cookstoves
#> 73      GHA    PCM           E      Final ktoe         Charcoal stoves
#> 74      GHA    PCM           E      Final ktoe         Charcoal stoves
#> 75      GHA    PCM           E      Final ktoe         Charcoal stoves
#> 76      GHA    PCM           E      Final ktoe         Charcoal stoves
#> 77      GHA    PCM           E      Final ktoe         Kerosene stoves
#> 78      GHA    PCM           E      Final ktoe         Kerosene stoves
#> 79      GHA    PCM           E      Final ktoe         Kerosene stoves
#> 80      GHA    PCM           E      Final ktoe         Kerosene stoves
#> 81      GHA    PCM           E      Final ktoe              LPG stoves
#> 82      GHA    PCM           E      Final ktoe              LPG stoves
#> 83      GHA    PCM           E      Final ktoe              LPG stoves
#> 84      GHA    PCM           E      Final ktoe              LPG stoves
#> 85      GHA    PCM           E      Final ktoe         Electric stoves
#> 86      GHA    PCM           E      Final ktoe         Electric stoves
#> 87      GHA    PCM           E      Final ktoe         Electric stoves
#> 88      GHA    PCM           E      Final ktoe         Electric stoves
#> 89      GHA    PCM           E      Final ktoe              Non-energy
#> 90      GHA    PCM           E      Final ktoe              Non-energy
#> 91      GHA    PCM           E      Final ktoe              Non-energy
#> 92      GHA    PCM           E      Final ktoe              Non-energy
#> 93      GHA    PCM           E      Final ktoe              Non-energy
#> 94      GHA    PCM           E      Final ktoe              Non-energy
#> 95      GHA    PCM           E      Final ktoe              Non-energy
#> 96      GHA    PCM           E      Final ktoe              Non-energy
#> 97      GHA    PCM           E      Final ktoe              Non-energy
#> 98      GHA    PCM           E      Final ktoe              Non-energy
#> 99      GHA    PCM           E      Final ktoe              Non-energy
#> 100     GHA    PCM           E      Final ktoe              Non-energy
#> 101     ZAF    PCM           E      Final ktoe  Steam engines (trains)
#> 102     ZAF    PCM           E      Final ktoe  Steam engines (trains)
#> 103     ZAF    PCM           E      Final ktoe  Steam engines (trains)
#> 104     ZAF    PCM           E      Final ktoe  Steam engines (trains)
#> 105     ZAF    PCM           E      Final ktoe         Electric motors
#> 106     ZAF    PCM           E      Final ktoe         Electric motors
#> 107     ZAF    PCM           E      Final ktoe         Electric motors
#> 108     ZAF    PCM           E      Final ktoe         Electric motors
#> 109     ZAF    PCM           E      Final ktoe                  Trucks
#> 110     ZAF    PCM           E      Final ktoe                  Trucks
#> 111     ZAF    PCM           E      Final ktoe                  Trucks
#> 112     ZAF    PCM           E      Final ktoe                  Trucks
#> 113     ZAF    PCM           E      Final ktoe             Automobiles
#> 114     ZAF    PCM           E      Final ktoe             Automobiles
#> 115     ZAF    PCM           E      Final ktoe             Automobiles
#> 116     ZAF    PCM           E      Final ktoe             Automobiles
#> 117     ZAF    PCM           E      Final ktoe  Industry static motors
#> 118     ZAF    PCM           E      Final ktoe  Industry static motors
#> 119     ZAF    PCM           E      Final ktoe  Industry static motors
#> 120     ZAF    PCM           E      Final ktoe  Industry static motors
#> 121     ZAF    PCM           E      Final ktoe           Diesel trucks
#> 122     ZAF    PCM           E      Final ktoe           Diesel trucks
#> 123     ZAF    PCM           E      Final ktoe           Diesel trucks
#> 124     ZAF    PCM           E      Final ktoe           Diesel trucks
#> 125     ZAF    PCM           E      Final ktoe      Diesel automobiles
#> 126     ZAF    PCM           E      Final ktoe      Diesel automobiles
#> 127     ZAF    PCM           E      Final ktoe      Diesel automobiles
#> 128     ZAF    PCM           E      Final ktoe      Diesel automobiles
#> 129     ZAF    PCM           E      Final ktoe                    Fans
#> 130     ZAF    PCM           E      Final ktoe                    Fans
#> 131     ZAF    PCM           E      Final ktoe                    Fans
#> 132     ZAF    PCM           E      Final ktoe                    Fans
#> 133     ZAF    PCM           E      Final ktoe               Airplanes
#> 134     ZAF    PCM           E      Final ktoe               Airplanes
#> 135     ZAF    PCM           E      Final ktoe               Airplanes
#> 136     ZAF    PCM           E      Final ktoe               Airplanes
#> 137     ZAF    PCM           E      Final ktoe Industry static engines
#> 138     ZAF    PCM           E      Final ktoe Industry static engines
#> 139     ZAF    PCM           E      Final ktoe Industry static engines
#> 140     ZAF    PCM           E      Final ktoe Industry static engines
#> 141     ZAF    PCM           E      Final ktoe    Electric automobiles
#> 142     ZAF    PCM           E      Final ktoe    Electric automobiles
#> 143     ZAF    PCM           E      Final ktoe    Electric automobiles
#> 144     ZAF    PCM           E      Final ktoe    Electric automobiles
#> 145     ZAF    PCM           E      Final ktoe            Boat engines
#> 146     ZAF    PCM           E      Final ktoe            Boat engines
#> 147     ZAF    PCM           E      Final ktoe            Boat engines
#> 148     ZAF    PCM           E      Final ktoe            Boat engines
#> 149     ZAF    PCM           E      Final ktoe              Generators
#> 150     ZAF    PCM           E      Final ktoe              Generators
#> 151     ZAF    PCM           E      Final ktoe              Generators
#> 152     ZAF    PCM           E      Final ktoe              Generators
#> 153     ZAF    PCM           E      Final ktoe         Electric lights
#> 154     ZAF    PCM           E      Final ktoe         Electric lights
#> 155     ZAF    PCM           E      Final ktoe         Electric lights
#> 156     ZAF    PCM           E      Final ktoe         Electric lights
#> 157     ZAF    PCM           E      Final ktoe         Kerosene lights
#> 158     ZAF    PCM           E      Final ktoe         Kerosene lights
#> 159     ZAF    PCM           E      Final ktoe         Kerosene lights
#> 160     ZAF    PCM           E      Final ktoe         Kerosene lights
#> 161     ZAF    PCM           E      Final ktoe              Gas lights
#> 162     ZAF    PCM           E      Final ktoe              Gas lights
#> 163     ZAF    PCM           E      Final ktoe              Gas lights
#> 164     ZAF    PCM           E      Final ktoe              Gas lights
#> 165     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 166     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 167     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 168     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 169     ZAF    PCM           E      Final ktoe        Electric heaters
#> 170     ZAF    PCM           E      Final ktoe        Electric heaters
#> 171     ZAF    PCM           E      Final ktoe        Electric heaters
#> 172     ZAF    PCM           E      Final ktoe        Electric heaters
#> 173     ZAF    PCM           E      Final ktoe             Gas heaters
#> 174     ZAF    PCM           E      Final ktoe             Gas heaters
#> 175     ZAF    PCM           E      Final ktoe             Gas heaters
#> 176     ZAF    PCM           E      Final ktoe             Gas heaters
#> 177     ZAF    PCM           E      Final ktoe             Oil heaters
#> 178     ZAF    PCM           E      Final ktoe             Oil heaters
#> 179     ZAF    PCM           E      Final ktoe             Oil heaters
#> 180     ZAF    PCM           E      Final ktoe             Oil heaters
#> 181     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 182     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 183     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 184     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 185     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 186     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 187     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 188     ZAF    PCM           E      Final ktoe     Industrial furnaces
#> 189     ZAF    PCM           E      Final ktoe             Wood stoves
#> 190     ZAF    PCM           E      Final ktoe             Wood stoves
#> 191     ZAF    PCM           E      Final ktoe             Wood stoves
#> 192     ZAF    PCM           E      Final ktoe             Wood stoves
#> 193     ZAF    PCM           E      Final ktoe                   Irons
#> 194     ZAF    PCM           E      Final ktoe                   Irons
#> 195     ZAF    PCM           E      Final ktoe                   Irons
#> 196     ZAF    PCM           E      Final ktoe                   Irons
#> 197     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 198     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 199     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 200     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 201     ZAF    PCM           E      Final ktoe             Oil heaters
#> 202     ZAF    PCM           E      Final ktoe             Oil heaters
#> 203     ZAF    PCM           E      Final ktoe             Oil heaters
#> 204     ZAF    PCM           E      Final ktoe             Oil heaters
#> 205     ZAF    PCM           E      Final ktoe             Gas heaters
#> 206     ZAF    PCM           E      Final ktoe             Gas heaters
#> 207     ZAF    PCM           E      Final ktoe             Gas heaters
#> 208     ZAF    PCM           E      Final ktoe             Gas heaters
#> 209     ZAF    PCM           E      Final ktoe        Electric heaters
#> 210     ZAF    PCM           E      Final ktoe        Electric heaters
#> 211     ZAF    PCM           E      Final ktoe        Electric heaters
#> 212     ZAF    PCM           E      Final ktoe        Electric heaters
#> 213     ZAF    PCM           E      Final ktoe              LPG stoves
#> 214     ZAF    PCM           E      Final ktoe              LPG stoves
#> 215     ZAF    PCM           E      Final ktoe              LPG stoves
#> 216     ZAF    PCM           E      Final ktoe              LPG stoves
#> 217     ZAF    PCM           E      Final ktoe             Coal stoves
#> 218     ZAF    PCM           E      Final ktoe             Coal stoves
#> 219     ZAF    PCM           E      Final ktoe             Coal stoves
#> 220     ZAF    PCM           E      Final ktoe             Coal stoves
#> 221     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 222     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 223     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 224     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 225     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 226     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 227     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 228     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 229     ZAF    PCM           E      Final ktoe         Wood cookstoves
#> 230     ZAF    PCM           E      Final ktoe         Wood cookstoves
#> 231     ZAF    PCM           E      Final ktoe         Wood cookstoves
#> 232     ZAF    PCM           E      Final ktoe         Wood cookstoves
#> 233     ZAF    PCM           E      Final ktoe         Charcoal stoves
#> 234     ZAF    PCM           E      Final ktoe         Charcoal stoves
#> 235     ZAF    PCM           E      Final ktoe         Charcoal stoves
#> 236     ZAF    PCM           E      Final ktoe         Charcoal stoves
#> 237     ZAF    PCM           E      Final ktoe          LPG cookstoves
#> 238     ZAF    PCM           E      Final ktoe          LPG cookstoves
#> 239     ZAF    PCM           E      Final ktoe          LPG cookstoves
#> 240     ZAF    PCM           E      Final ktoe          LPG cookstoves
#> 241     ZAF    PCM           E      Final ktoe              LPG stoves
#> 242     ZAF    PCM           E      Final ktoe              LPG stoves
#> 243     ZAF    PCM           E      Final ktoe              LPG stoves
#> 244     ZAF    PCM           E      Final ktoe              LPG stoves
#> 245     ZAF    PCM           E      Final ktoe             Coal stoves
#> 246     ZAF    PCM           E      Final ktoe             Coal stoves
#> 247     ZAF    PCM           E      Final ktoe             Coal stoves
#> 248     ZAF    PCM           E      Final ktoe             Coal stoves
#> 249     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 250     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 251     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 252     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 253     ZAF    PCM           E      Final ktoe             Gas heaters
#> 254     ZAF    PCM           E      Final ktoe             Gas heaters
#> 255     ZAF    PCM           E      Final ktoe             Gas heaters
#> 256     ZAF    PCM           E      Final ktoe             Gas heaters
#> 257     ZAF    PCM           E      Final ktoe             Coal stoves
#> 258     ZAF    PCM           E      Final ktoe             Coal stoves
#> 259     ZAF    PCM           E      Final ktoe             Coal stoves
#> 260     ZAF    PCM           E      Final ktoe             Coal stoves
#> 261     ZAF    PCM           E      Final ktoe           Wood furnaces
#> 262     ZAF    PCM           E      Final ktoe           Wood furnaces
#> 263     ZAF    PCM           E      Final ktoe           Wood furnaces
#> 264     ZAF    PCM           E      Final ktoe           Wood furnaces
#> 265     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 266     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 267     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 268     ZAF    PCM           E      Final ktoe            Gas furnaces
#> 269     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 270     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 271     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 272     ZAF    PCM           E      Final ktoe         Kerosene stoves
#> 273     ZAF    PCM           E      Final ktoe              LPG stoves
#> 274     ZAF    PCM           E      Final ktoe              LPG stoves
#> 275     ZAF    PCM           E      Final ktoe              LPG stoves
#> 276     ZAF    PCM           E      Final ktoe              LPG stoves
#> 277     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 278     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 279     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 280     ZAF    PCM           E      Final ktoe            Oil furnaces
#> 281     ZAF    PCM           E      Final ktoe              Non-energy
#> 282     ZAF    PCM           E      Final ktoe              Non-energy
#> 283     ZAF    PCM           E      Final ktoe              Non-energy
#> 284     ZAF    PCM           E      Final ktoe              Non-energy
#> 285     ZAF    PCM           E      Final ktoe              Non-energy
#> 286     ZAF    PCM           E      Final ktoe              Non-energy
#> 287     ZAF    PCM           E      Final ktoe              Non-energy
#> 288     ZAF    PCM           E      Final ktoe              Non-energy
#> 289     ZAF    PCM           E      Final ktoe              Non-energy
#> 290     ZAF    PCM           E      Final ktoe              Non-energy
#> 291     ZAF    PCM           E      Final ktoe              Non-energy
#> 292     ZAF    PCM           E      Final ktoe              Non-energy
#> 293     ZAF    PCM           E      Final ktoe              Non-energy
#> 294     ZAF    PCM           E      Final ktoe              Non-energy
#> 295     ZAF    PCM           E      Final ktoe              Non-energy
#> 296     ZAF    PCM           E      Final ktoe              Non-energy
#> 297     ZAF    PCM           E      Final ktoe              Non-energy
#> 298     ZAF    PCM           E      Final ktoe              Non-energy
#> 299     ZAF    PCM           E      Final ktoe              Non-energy
#> 300     ZAF    PCM           E      Final ktoe              Non-energy
#> 301     ZAF    PCM           E      Final ktoe              Non-energy
#> 302     ZAF    PCM           E      Final ktoe              Non-energy
#> 303     ZAF    PCM           E      Final ktoe              Non-energy
#> 304     ZAF    PCM           E      Final ktoe              Non-energy
#>                   Eu.product          Quantity Maximum.values         1971
#> 1                         MD     E.dot_machine   6.192510e+02 2.172315e+02
#> 2                         MD E.dot_machine [%]   1.156388e-01 8.391282e-02
#> 3                         MD            eta.fu             NA 3.000000e-01
#> 4                         MD             phi.u             NA 1.000000e+00
#> 5                         MD     E.dot_machine   1.877078e+02 5.946665e+01
#> 6                         MD E.dot_machine [%]   3.505250e-02 2.297095e-02
#> 7                         MD            eta.fu             NA 3.000000e-01
#> 8                         MD             phi.u             NA 1.000000e+00
#> 9                         MD     E.dot_machine   1.877078e+02 5.946665e+01
#> 10                        MD E.dot_machine [%]   3.505250e-02 2.297095e-02
#> 11                        MD            eta.fu             NA 3.000000e-01
#> 12                        MD             phi.u             NA 1.000000e+00
#> 13                        MD     E.dot_machine   1.870163e+02 6.383495e+01
#> 14                        MD E.dot_machine [%]   3.492338e-02 2.465835e-02
#> 15                        MD            eta.fu             NA 8.500000e-01
#> 16                        MD             phi.u             NA 1.000000e+00
#> 17                        MD     E.dot_machine   7.032580e+01 1.499595e+01
#> 18                        MD E.dot_machine [%]   1.313262e-02 5.792678e-03
#> 19                        MD            eta.fu             NA 3.000000e-01
#> 20                        MD             phi.u             NA 1.000000e+00
#> 21                        MD     E.dot_machine   6.316280e+01 1.226230e+01
#> 22                        MD E.dot_machine [%]   1.179500e-02 4.736716e-03
#> 23                        MD            eta.fu             NA 3.000000e-01
#> 24                        MD             phi.u             NA 1.000000e+00
#> 25                        MD     E.dot_machine   2.585510e+01 1.514280e+01
#> 26                        MD E.dot_machine [%]   5.849404e-03 5.849404e-03
#> 27                        MD            eta.fu             NA 3.000000e-01
#> 28                        MD             phi.u             NA 1.000000e+00
#> 29                        MD     E.dot_machine   1.907567e+01 2.115225e+00
#> 30                        MD E.dot_machine [%]   3.562184e-03 8.170751e-04
#> 31                        MD            eta.fu             NA 1.000000e-01
#> 32                        MD             phi.u             NA 1.000000e+00
#> 33                     Light     E.dot_machine   1.589037e+02 1.411179e+02
#> 34                     Light E.dot_machine [%]   5.451143e-02 5.451143e-02
#> 35                     Light            eta.fu             NA 2.000000e-01
#> 36                     Light             phi.u             NA 1.261201e-01
#> 37                 HTH.600.C     E.dot_machine   2.642350e+01 1.953040e+01
#> 38                 HTH.600.C E.dot_machine [%]   7.544259e-03 7.544259e-03
#> 39                 HTH.600.C            eta.fu             NA 8.500000e-01
#> 40                 HTH.600.C             phi.u             NA 6.585352e-01
#> 41                 HTH.600.C     E.dot_machine   1.584695e+01 1.584695e+01
#> 42                 HTH.600.C E.dot_machine [%]   6.121405e-03 6.121405e-03
#> 43                 HTH.600.C            eta.fu             NA 8.500000e-01
#> 44                 HTH.600.C             phi.u             NA 6.585352e-01
#> 45                 MTH.200.C     E.dot_machine   6.852728e+02 1.456960e+02
#> 46                 MTH.200.C E.dot_machine [%]   1.279677e-01 5.627987e-02
#> 47                 MTH.200.C            eta.fu             NA 8.500000e-01
#> 48                 MTH.200.C             phi.u             NA 3.698616e-01
#> 49                 MTH.200.C     E.dot_machine   2.166498e+02 1.369423e+02
#> 50                 MTH.200.C E.dot_machine [%]   5.289846e-02 5.289846e-02
#> 51                 MTH.200.C            eta.fu             NA 8.500000e-01
#> 52                 MTH.200.C             phi.u             NA 3.698616e-01
#> 53                 MTH.200.C     E.dot_machine   1.480998e+02 9.355120e+00
#> 54                 MTH.200.C E.dot_machine [%]   2.765611e-02 3.613722e-03
#> 55                 MTH.200.C            eta.fu             NA 9.700000e-01
#> 56                 MTH.200.C             phi.u             NA 3.698616e-01
#> 57                 MTH.200.C     E.dot_machine   1.907567e+01 2.115225e+00
#> 58                 MTH.200.C E.dot_machine [%]   3.562184e-03 8.170751e-04
#> 59                 MTH.200.C            eta.fu             NA 8.500000e-01
#> 60                 MTH.200.C             phi.u             NA 3.698616e-01
#> 61                 MTH.200.C     E.dot_machine   4.519000e+00           NA
#> 62                 MTH.200.C E.dot_machine [%]   8.438768e-04           NA
#> 63                 MTH.200.C            eta.fu             NA           NA
#> 64                 MTH.200.C             phi.u             NA           NA
#> 65                 MTH.200.C     E.dot_machine   1.160800e+00 1.160800e+00
#> 66                 MTH.200.C E.dot_machine [%]   4.483971e-04 4.483971e-04
#> 67                 MTH.200.C            eta.fu             NA 6.500000e-01
#> 68                 MTH.200.C             phi.u             NA 3.698616e-01
#> 69                 MTH.100.C     E.dot_machine   2.111732e+03 1.464125e+03
#> 70                 MTH.100.C E.dot_machine [%]   5.655665e-01 5.655665e-01
#> 71                 MTH.100.C            eta.fu             NA 1.200000e-01
#> 72                 MTH.100.C             phi.u             NA 2.009916e-01
#> 73                 MTH.100.C     E.dot_machine   5.156874e+02 1.191745e+02
#> 74                 MTH.100.C E.dot_machine [%]   9.629932e-02 4.603507e-02
#> 75                 MTH.100.C            eta.fu             NA 1.400000e-01
#> 76                 MTH.100.C             phi.u             NA 2.009916e-01
#> 77                 MTH.100.C     E.dot_machine   8.369160e+01 8.369160e+01
#> 78                 MTH.100.C E.dot_machine [%]   3.232863e-02 3.232863e-02
#> 79                 MTH.100.C            eta.fu             NA 1.600000e-01
#> 80                 MTH.100.C             phi.u             NA 2.009916e-01
#> 81                 MTH.100.C     E.dot_machine   3.615170e+01 3.389200e+00
#> 82                 MTH.100.C E.dot_machine [%]   6.750959e-03 1.309190e-03
#> 83                 MTH.100.C            eta.fu             NA 1.800000e-01
#> 84                 MTH.100.C             phi.u             NA 2.009916e-01
#> 85                 MTH.100.C     E.dot_machine   1.907567e+01 2.115225e+00
#> 86                 MTH.100.C E.dot_machine [%]   3.562184e-03 8.170751e-04
#> 87                 MTH.100.C            eta.fu             NA 7.000000e-01
#> 88                 MTH.100.C             phi.u             NA 2.009916e-01
#> 89                   Bitumen     E.dot_machine   5.589000e+00           NA
#> 90                   Bitumen E.dot_machine [%]   1.030561e-03           NA
#> 91                   Bitumen            eta.fu             NA           NA
#> 92                   Bitumen             phi.u             NA           NA
#> 93                Lubricants     E.dot_machine   3.109770e+01 1.805670e+01
#> 94                Lubricants E.dot_machine [%]   6.926680e-03 6.926680e-03
#> 95                Lubricants            eta.fu             NA 1.000000e-06
#> 96                Lubricants             phi.u             NA 1.000000e+00
#> 97        Other oil products     E.dot_machine   3.152770e+01           NA
#> 98        Other oil products E.dot_machine [%]   5.813421e-03           NA
#> 99        Other oil products            eta.fu             NA           NA
#> 100       Other oil products             phi.u             NA           NA
#> 101                       MD     E.dot_machine   3.295513e+03 3.295513e+03
#> 102                       MD E.dot_machine [%]   9.922186e-02 9.922186e-02
#> 103                       MD            eta.fu             NA 6.000000e-01
#> 104                       MD             phi.u             NA 1.000000e+00
#> 105                       MD     E.dot_machine   5.168693e+03 1.030447e+03
#> 106                       MD E.dot_machine [%]   9.848315e-02 3.102487e-02
#> 107                       MD            eta.fu             NA 8.700000e-01
#> 108                       MD             phi.u             NA 1.000000e+00
#> 109                       MD     E.dot_machine   4.109393e+03 1.850616e+03
#> 110                       MD E.dot_machine [%]   7.829947e-02 5.571866e-02
#> 111                       MD            eta.fu             NA 3.000000e-01
#> 112                       MD             phi.u             NA 1.000000e+00
#> 113                       MD     E.dot_machine   4.029254e+03 2.305665e+03
#> 114                       MD E.dot_machine [%]   7.677253e-02 6.941935e-02
#> 115                       MD            eta.fu             NA 3.000000e-01
#> 116                       MD             phi.u             NA 1.000000e+00
#> 117                       MD     E.dot_machine   2.496819e+03 1.190714e+03
#> 118                       MD E.dot_machine [%]   4.757384e-02 3.585020e-02
#> 119                       MD            eta.fu             NA 3.000000e-01
#> 120                       MD             phi.u             NA 1.000000e+00
#> 121                       MD     E.dot_machine   1.673811e+03 7.115315e+02
#> 122                       MD E.dot_machine [%]   3.189242e-02 2.142291e-02
#> 123                       MD            eta.fu             NA 3.000000e-01
#> 124                       MD             phi.u             NA 1.000000e+00
#> 125                       MD     E.dot_machine   1.672812e+03 7.115315e+02
#> 126                       MD E.dot_machine [%]   3.187339e-02 2.142291e-02
#> 127                       MD            eta.fu             NA 3.000000e-01
#> 128                       MD             phi.u             NA 1.000000e+00
#> 129                       MD     E.dot_machine   7.398108e+02 2.086328e+02
#> 130                       MD E.dot_machine [%]   1.409619e-02 6.281552e-03
#> 131                       MD            eta.fu             NA 1.200000e-01
#> 132                       MD             phi.u             NA 1.000000e+00
#> 133                       MD     E.dot_machine   6.141134e+02 1.715057e+02
#> 134                       MD E.dot_machine [%]   1.170118e-02 5.163722e-03
#> 135                       MD            eta.fu             NA 3.000000e-01
#> 136                       MD             phi.u             NA 1.000000e+00
#> 137                       MD     E.dot_machine   5.185681e+02           NA
#> 138                       MD E.dot_machine [%]   9.880684e-03           NA
#> 139                       MD            eta.fu             NA           NA
#> 140                       MD             phi.u             NA           NA
#> 141                       MD     E.dot_machine   1.620809e+02 5.176270e+01
#> 142                       MD E.dot_machine [%]   3.088254e-03 1.558480e-03
#> 143                       MD            eta.fu             NA 4.500000e-01
#> 144                       MD             phi.u             NA 1.000000e+00
#> 145                       MD     E.dot_machine   1.026154e+02           NA
#> 146                       MD E.dot_machine [%]   1.955211e-03           NA
#> 147                       MD            eta.fu             NA           NA
#> 148                       MD             phi.u             NA           NA
#> 149                       MD     E.dot_machine   3.366390e+01           NA
#> 150                       MD E.dot_machine [%]   6.414246e-04           NA
#> 151                       MD            eta.fu             NA           NA
#> 152                       MD             phi.u             NA           NA
#> 153                    Light     E.dot_machine   6.263618e+03 1.413929e+03
#> 154                    Light E.dot_machine [%]   1.193456e-01 4.257082e-02
#> 155                    Light            eta.fu             NA 2.000000e-01
#> 156                    Light             phi.u             NA 1.261201e-01
#> 157                    Light     E.dot_machine   1.497862e+02           NA
#> 158                    Light E.dot_machine [%]   2.853994e-03           NA
#> 159                    Light            eta.fu             NA           NA
#> 160                    Light             phi.u             NA           NA
#> 161                    Light     E.dot_machine   1.154460e+02 4.325020e+01
#> 162                    Light E.dot_machine [%]   2.199684e-03 1.302184e-03
#> 163                    Light            eta.fu             NA 2.000000e-01
#> 164                    Light             phi.u             NA 1.261201e-01
#> 165                HTH.600.C     E.dot_machine   7.157481e+03 7.157481e+03
#> 166                HTH.600.C E.dot_machine [%]   2.154986e-01 2.154986e-01
#> 167                HTH.600.C            eta.fu             NA 8.500000e-01
#> 168                HTH.600.C             phi.u             NA 6.585352e-01
#> 169                HTH.600.C     E.dot_machine   1.683990e+03 7.889936e+01
#> 170                HTH.600.C E.dot_machine [%]   3.208637e-02 2.375515e-03
#> 171                HTH.600.C            eta.fu             NA 9.000000e-01
#> 172                HTH.600.C             phi.u             NA 6.585352e-01
#> 173                HTH.600.C     E.dot_machine   1.052345e+03           NA
#> 174                HTH.600.C E.dot_machine [%]   2.005116e-02           NA
#> 175                HTH.600.C            eta.fu             NA           NA
#> 176                HTH.600.C             phi.u             NA           NA
#> 177                HTH.600.C     E.dot_machine   5.799369e+02 5.799369e+02
#> 178                HTH.600.C E.dot_machine [%]   1.746084e-02 1.746084e-02
#> 179                HTH.600.C            eta.fu             NA 8.000000e-01
#> 180                HTH.600.C             phi.u             NA 6.585352e-01
#> 181                HTH.600.C     E.dot_machine   1.706363e+02 2.914880e+01
#> 182                HTH.600.C E.dot_machine [%]   3.251267e-03 8.776168e-04
#> 183                HTH.600.C            eta.fu             NA 8.000000e-01
#> 184                HTH.600.C             phi.u             NA 6.585352e-01
#> 185                MTH.200.C     E.dot_machine   3.459107e+03 2.080060e+03
#> 186                MTH.200.C E.dot_machine [%]   6.590907e-02 6.262680e-02
#> 187                MTH.200.C            eta.fu             NA 8.000000e-01
#> 188                MTH.200.C             phi.u             NA 3.698616e-01
#> 189                MTH.200.C     E.dot_machine   1.508431e+03 7.643069e+02
#> 190                MTH.200.C E.dot_machine [%]   2.874132e-02 2.301188e-02
#> 191                MTH.200.C            eta.fu             NA 2.000000e-01
#> 192                MTH.200.C             phi.u             NA 3.698616e-01
#> 193                MTH.200.C     E.dot_machine   1.233018e+03 3.477214e+02
#> 194                MTH.200.C E.dot_machine [%]   2.349366e-02 1.046925e-02
#> 195                MTH.200.C            eta.fu             NA 3.000000e-01
#> 196                MTH.200.C             phi.u             NA 3.698616e-01
#> 197                MTH.200.C     E.dot_machine   5.596876e+02 5.596876e+02
#> 198                MTH.200.C E.dot_machine [%]   1.685117e-02 1.685117e-02
#> 199                MTH.200.C            eta.fu             NA 1.400000e-01
#> 200                MTH.200.C             phi.u             NA 3.698616e-01
#> 201                MTH.200.C     E.dot_machine   5.235502e+02 5.235502e+02
#> 202                MTH.200.C E.dot_machine [%]   1.576314e-02 1.576314e-02
#> 203                MTH.200.C            eta.fu             NA 4.000000e-01
#> 204                MTH.200.C             phi.u             NA 3.698616e-01
#> 205                MTH.200.C     E.dot_machine   5.746345e+02 6.605760e+01
#> 206                MTH.200.C E.dot_machine [%]   1.094896e-02 1.988873e-03
#> 207                MTH.200.C            eta.fu             NA 4.000000e-01
#> 208                MTH.200.C             phi.u             NA 3.698616e-01
#> 209                MTH.200.C     E.dot_machine   4.777397e+02 4.315563e+01
#> 210                MTH.200.C E.dot_machine [%]   9.102749e-03 1.299337e-03
#> 211                MTH.200.C            eta.fu             NA 4.000000e-01
#> 212                MTH.200.C             phi.u             NA 3.698616e-01
#> 213                MTH.200.C     E.dot_machine   1.197526e+02 1.197526e+02
#> 214                MTH.200.C E.dot_machine [%]   3.605531e-03 3.605531e-03
#> 215                MTH.200.C            eta.fu             NA 3.000000e-01
#> 216                MTH.200.C             phi.u             NA 3.698616e-01
#> 217                MTH.200.C     E.dot_machine   7.672980e+01 7.672980e+01
#> 218                MTH.200.C E.dot_machine [%]   2.310193e-03 2.310193e-03
#> 219                MTH.200.C            eta.fu             NA 2.500000e-01
#> 220                MTH.200.C             phi.u             NA 3.698616e-01
#> 221                MTH.200.C     E.dot_machine   1.098900e+01           NA
#> 222                MTH.200.C E.dot_machine [%]   2.093820e-04           NA
#> 223                MTH.200.C            eta.fu             NA           NA
#> 224                MTH.200.C             phi.u             NA           NA
#> 225                MTH.200.C     E.dot_machine   3.075000e+00           NA
#> 226                MTH.200.C E.dot_machine [%]   5.859038e-05           NA
#> 227                MTH.200.C            eta.fu             NA           NA
#> 228                MTH.200.C             phi.u             NA           NA
#> 229                MTH.100.C     E.dot_machine   2.787475e+03 1.934652e+03
#> 230                MTH.100.C E.dot_machine [%]   5.824881e-02 5.824881e-02
#> 231                MTH.100.C            eta.fu             NA 1.200000e-01
#> 232                MTH.100.C             phi.u             NA 2.009916e-01
#> 233                MTH.100.C     E.dot_machine   6.584026e+02 1.691980e+01
#> 234                MTH.100.C E.dot_machine [%]   1.254506e-02 5.094241e-04
#> 235                MTH.100.C            eta.fu             NA 1.400000e-01
#> 236                MTH.100.C             phi.u             NA 2.009916e-01
#> 237                MTH.100.C     E.dot_machine   1.452114e+02           NA
#> 238                MTH.100.C E.dot_machine [%]   2.766826e-03           NA
#> 239                MTH.100.C            eta.fu             NA           NA
#> 240                MTH.100.C             phi.u             NA           NA
#> 241                MTH.100.C     E.dot_machine   4.691440e+01           NA
#> 242                MTH.100.C E.dot_machine [%]   8.938968e-04           NA
#> 243                MTH.100.C            eta.fu             NA           NA
#> 244                MTH.100.C             phi.u             NA           NA
#> 245                MTH.100.C     E.dot_machine   4.449040e+01           NA
#> 246                MTH.100.C E.dot_machine [%]   8.477104e-04           NA
#> 247                MTH.100.C            eta.fu             NA           NA
#> 248                MTH.100.C             phi.u             NA           NA
#> 249                MTH.100.C     E.dot_machine   3.099000e+00           NA
#> 250                MTH.100.C E.dot_machine [%]   5.904767e-05           NA
#> 251                MTH.100.C            eta.fu             NA           NA
#> 252                MTH.100.C             phi.u             NA           NA
#> 253                MTH.100.C     E.dot_machine   5.223600e-01           NA
#> 254                MTH.100.C E.dot_machine [%]   9.952934e-06           NA
#> 255                MTH.100.C            eta.fu             NA           NA
#> 256                MTH.100.C             phi.u             NA           NA
#> 257                 LTH.20.C     E.dot_machine   3.140119e+03 3.140119e+03
#> 258                 LTH.20.C E.dot_machine [%]   9.454323e-02 9.454323e-02
#> 259                 LTH.20.C            eta.fu             NA 1.600000e-01
#> 260                 LTH.20.C             phi.u             NA 1.677008e-02
#> 261                 LTH.20.C     E.dot_machine   2.787475e+03 1.934652e+03
#> 262                 LTH.20.C E.dot_machine [%]   5.824881e-02 5.824881e-02
#> 263                 LTH.20.C            eta.fu             NA 8.500000e-01
#> 264                 LTH.20.C             phi.u             NA 1.677008e-02
#> 265                 LTH.20.C     E.dot_machine   9.715247e+02 7.756520e+02
#> 266                 LTH.20.C E.dot_machine [%]   2.335346e-02 2.335346e-02
#> 267                 LTH.20.C            eta.fu             NA 8.500000e-01
#> 268                 LTH.20.C             phi.u             NA 1.677008e-02
#> 269                 LTH.20.C     E.dot_machine   5.516266e+02           NA
#> 270                 LTH.20.C E.dot_machine [%]   1.051057e-02           NA
#> 271                 LTH.20.C            eta.fu             NA           NA
#> 272                 LTH.20.C             phi.u             NA           NA
#> 273                 LTH.20.C     E.dot_machine   1.452114e+02           NA
#> 274                 LTH.20.C E.dot_machine [%]   2.766826e-03           NA
#> 275                 LTH.20.C            eta.fu             NA           NA
#> 276                 LTH.20.C             phi.u             NA           NA
#> 277                 LTH.20.C     E.dot_machine   1.398600e+01           NA
#> 278                 LTH.20.C E.dot_machine [%]   2.664862e-04           NA
#> 279                 LTH.20.C            eta.fu             NA           NA
#> 280                 LTH.20.C             phi.u             NA           NA
#> 281                  Bitumen     E.dot_machine   2.142448e+02 1.844368e+02
#> 282                  Bitumen E.dot_machine [%]   5.401422e-03 5.401422e-03
#> 283                  Bitumen            eta.fu             NA 1.000000e-06
#> 284                  Bitumen             phi.u             NA 1.000000e+00
#> 285 Hard coal (if no detail)     E.dot_machine   7.479545e+02 7.479545e+02
#> 286 Hard coal (if no detail) E.dot_machine [%]   2.190462e-02 2.190462e-02
#> 287 Hard coal (if no detail)            eta.fu             NA 1.000000e-06
#> 288 Hard coal (if no detail)             phi.u             NA 1.000000e+00
#> 289               Lubricants     E.dot_machine   7.924910e+01           NA
#> 290               Lubricants E.dot_machine [%]   1.363107e-03           NA
#> 291               Lubricants            eta.fu             NA           NA
#> 292               Lubricants             phi.u             NA           NA
#> 293    Other bituminous coal     E.dot_machine   5.284686e+03           NA
#> 294    Other bituminous coal E.dot_machine [%]   9.089807e-02           NA
#> 295    Other bituminous coal            eta.fu             NA           NA
#> 296    Other bituminous coal             phi.u             NA           NA
#> 297           Paraffin waxes     E.dot_machine   6.687700e+00           NA
#> 298           Paraffin waxes E.dot_machine [%]   1.150303e-04           NA
#> 299           Paraffin waxes            eta.fu             NA           NA
#> 300           Paraffin waxes             phi.u             NA           NA
#> 301       White spirit & SBP     E.dot_machine   7.003500e+01           NA
#> 302       White spirit & SBP E.dot_machine [%]   1.204622e-03           NA
#> 303       White spirit & SBP            eta.fu             NA           NA
#> 304       White spirit & SBP             phi.u             NA           NA
#>             2000
#> 1   6.192510e+02
#> 2   1.156388e-01
#> 3   3.100000e-01
#> 4   1.000000e+00
#> 5   1.877078e+02
#> 6   3.505250e-02
#> 7   3.000000e-01
#> 8   1.000000e+00
#> 9   1.877078e+02
#> 10  3.505250e-02
#> 11  3.000000e-01
#> 12  1.000000e+00
#> 13  1.870163e+02
#> 14  3.492338e-02
#> 15  9.500000e-01
#> 16  1.000000e+00
#> 17  7.032580e+01
#> 18  1.313262e-02
#> 19  3.100000e-01
#> 20  1.000000e+00
#> 21  6.316280e+01
#> 22  1.179500e-02
#> 23  3.100000e-01
#> 24  1.000000e+00
#> 25  2.585510e+01
#> 26  4.828174e-03
#> 27  3.100000e-01
#> 28  1.000000e+00
#> 29  1.907567e+01
#> 30  3.562184e-03
#> 31  1.200000e-01
#> 32  1.000000e+00
#> 33  1.589037e+02
#> 34  2.967363e-02
#> 35  2.000000e-01
#> 36  1.830161e-01
#> 37  2.642350e+01
#> 38  4.934317e-03
#> 39  8.600000e-01
#> 40  6.585352e-01
#> 41            NA
#> 42            NA
#> 43            NA
#> 44            NA
#> 45  6.852728e+02
#> 46  1.279677e-01
#> 47  8.600000e-01
#> 48  3.698616e-01
#> 49  2.166498e+02
#> 50  4.045712e-02
#> 51  8.600000e-01
#> 52  3.698616e-01
#> 53  1.480998e+02
#> 54  2.765611e-02
#> 55  9.900000e-01
#> 56  3.698616e-01
#> 57  1.907567e+01
#> 58  3.562184e-03
#> 59  8.500000e-01
#> 60  3.698616e-01
#> 61  4.519000e+00
#> 62  8.438768e-04
#> 63  4.800000e-01
#> 64  3.698616e-01
#> 65            NA
#> 66            NA
#> 67            NA
#> 68            NA
#> 69  2.111732e+03
#> 70  3.943443e-01
#> 71  1.200000e-01
#> 72  2.009916e-01
#> 73  5.156874e+02
#> 74  9.629932e-02
#> 75  1.400000e-01
#> 76  2.009916e-01
#> 77  5.335340e+01
#> 78  9.963199e-03
#> 79  1.600000e-01
#> 80  2.009916e-01
#> 81  3.615170e+01
#> 82  6.750959e-03
#> 83  1.800000e-01
#> 84  2.009916e-01
#> 85  1.907567e+01
#> 86  3.562184e-03
#> 87  7.000000e-01
#> 88  2.009916e-01
#> 89  5.589000e+00
#> 90  1.030561e-03
#> 91  1.000000e-06
#> 92  1.000000e+00
#> 93  3.109770e+01
#> 94  5.734133e-03
#> 95  1.000000e-06
#> 96  1.000000e+00
#> 97  3.152770e+01
#> 98  5.813421e-03
#> 99  1.000000e-06
#> 100 1.000000e+00
#> 101 1.926513e+02
#> 102 3.670736e-03
#> 103 6.200000e-01
#> 104 1.000000e+00
#> 105 5.168693e+03
#> 106 9.848315e-02
#> 107 9.600000e-01
#> 108 1.000000e+00
#> 109 4.109393e+03
#> 110 7.829947e-02
#> 111 3.100000e-01
#> 112 1.000000e+00
#> 113 4.029254e+03
#> 114 7.677253e-02
#> 115 3.100000e-01
#> 116 1.000000e+00
#> 117 2.496819e+03
#> 118 4.757384e-02
#> 119 3.100000e-01
#> 120 1.000000e+00
#> 121 1.673811e+03
#> 122 3.189242e-02
#> 123 3.100000e-01
#> 124 1.000000e+00
#> 125 1.672812e+03
#> 126 3.187339e-02
#> 127 3.100000e-01
#> 128 1.000000e+00
#> 129 7.398108e+02
#> 130 1.409619e-02
#> 131 1.200000e-01
#> 132 1.000000e+00
#> 133 6.141134e+02
#> 134 1.170118e-02
#> 135 3.100000e-01
#> 136 1.000000e+00
#> 137 5.185681e+02
#> 138 9.880684e-03
#> 139 3.100000e-01
#> 140 1.000000e+00
#> 141 1.620809e+02
#> 142 3.088254e-03
#> 143 5.000000e-01
#> 144 1.000000e+00
#> 145 1.026154e+02
#> 146 1.955211e-03
#> 147 3.100000e-01
#> 148 1.000000e+00
#> 149 3.366390e+01
#> 150 6.414246e-04
#> 151 3.100000e-01
#> 152 1.000000e+00
#> 153 6.263618e+03
#> 154 1.193456e-01
#> 155 2.000000e-01
#> 156 1.830161e-01
#> 157 1.497862e+02
#> 158 2.853994e-03
#> 159 2.000000e-01
#> 160 1.830161e-01
#> 161 1.154460e+02
#> 162 2.199684e-03
#> 163 2.000000e-01
#> 164 1.830161e-01
#> 165 4.612028e+03
#> 166 8.787657e-02
#> 167 9.000000e-01
#> 168 6.585352e-01
#> 169 1.683990e+03
#> 170 3.208637e-02
#> 171 9.500000e-01
#> 172 6.585352e-01
#> 173 1.052345e+03
#> 174 2.005116e-02
#> 175 8.500000e-01
#> 176 6.585352e-01
#> 177           NA
#> 178           NA
#> 179           NA
#> 180           NA
#> 181 1.706363e+02
#> 182 3.251267e-03
#> 183 8.500000e-01
#> 184 6.585352e-01
#> 185 3.459107e+03
#> 186 6.590907e-02
#> 187 8.500000e-01
#> 188 3.698616e-01
#> 189 1.508431e+03
#> 190 2.874132e-02
#> 191 2.500000e-01
#> 192 3.698616e-01
#> 193 1.233018e+03
#> 194 2.349366e-02
#> 195 3.500000e-01
#> 196 3.698616e-01
#> 197 5.165000e+00
#> 198 9.841278e-05
#> 199 1.400000e-01
#> 200 3.698616e-01
#> 201           NA
#> 202           NA
#> 203           NA
#> 204           NA
#> 205 5.746345e+02
#> 206 1.094896e-02
#> 207 4.500000e-01
#> 208 3.698616e-01
#> 209 4.777397e+02
#> 210 9.102749e-03
#> 211 4.500000e-01
#> 212 3.698616e-01
#> 213 2.680800e+00
#> 214 5.107938e-05
#> 215 3.500000e-01
#> 216 3.698616e-01
#> 217           NA
#> 218           NA
#> 219           NA
#> 220           NA
#> 221 1.098900e+01
#> 222 2.093820e-04
#> 223 9.000000e-01
#> 224 3.698616e-01
#> 225 3.075000e+00
#> 226 5.859038e-05
#> 227 9.000000e-01
#> 228 3.698616e-01
#> 229 2.787475e+03
#> 230 5.311194e-02
#> 231 1.200000e-01
#> 232 2.009916e-01
#> 233 6.584026e+02
#> 234 1.254506e-02
#> 235 1.400000e-01
#> 236 2.009916e-01
#> 237 1.452114e+02
#> 238 2.766826e-03
#> 239 1.600000e-01
#> 240 2.009916e-01
#> 241 4.691440e+01
#> 242 8.938968e-04
#> 243 1.600000e-01
#> 244 2.009916e-01
#> 245 4.449040e+01
#> 246 8.477104e-04
#> 247 1.600000e-01
#> 248 2.009916e-01
#> 249 3.099000e+00
#> 250 5.904767e-05
#> 251 1.600000e-01
#> 252 2.009916e-01
#> 253 5.223600e-01
#> 254 9.952934e-06
#> 255 9.000000e-01
#> 256 2.009916e-01
#> 257 1.490106e+03
#> 258 2.839215e-02
#> 259 1.600000e-01
#> 260 1.677008e-02
#> 261 2.787475e+03
#> 262 5.311194e-02
#> 263 8.500000e-01
#> 264 1.677008e-02
#> 265 9.715247e+02
#> 266 1.851122e-02
#> 267 8.500000e-01
#> 268 1.677008e-02
#> 269 5.516266e+02
#> 270 1.051057e-02
#> 271 3.000000e-01
#> 272 1.677008e-02
#> 273 1.452114e+02
#> 274 2.766826e-03
#> 275 3.200000e-01
#> 276 1.677008e-02
#> 277 1.398600e+01
#> 278 2.664862e-04
#> 279 8.500000e-01
#> 280 1.677008e-02
#> 281 2.142448e+02
#> 282 3.685070e-03
#> 283 1.000000e-06
#> 284 1.000000e+00
#> 285           NA
#> 286           NA
#> 287           NA
#> 288           NA
#> 289 7.924910e+01
#> 290 1.363107e-03
#> 291 1.000000e-06
#> 292 1.000000e+00
#> 293 5.284686e+03
#> 294 9.089807e-02
#> 295 1.000000e-06
#> 296 1.000000e+00
#> 297 6.687700e+00
#> 298 1.150303e-04
#> 299 1.000000e-06
#> 300 1.000000e+00
#> 301 7.003500e+01
#> 302 1.204622e-03
#> 303 1.000000e-06
#> 304 1.000000e+00