Load final-to-useful machine efficiency data

When performing extending an energy conversion chain from useful energy to final energy, efficiencies of final-to-useful energy conversion machines are defined by the analyst. The Excel file at path contains those allocations.

Usage

load_eta_fu_data(
  path = sample_eta_fu_table_path(),
  eta_fu_tab_name = IEATools::fu_analysis_file_info$eta_fu_tab_name
)

Arguments

path

the path from which final-to-useful machine efficiency data will be loaded. Default is the path to sample efficiency data supplied with this package.

eta_fu_tab_name

the tab in path that contains the final-to-useful machine efficiency data. Default is "FU etas".

Value

the eta_fu_tab_name tab in path as a data frame.

Details

A final-to-useful machine efficiencies template can be generated using eta_fu_template() and write_eta_fu_template().

A filled example can be loaded with the default value of path.

Examples

load_eta_fu_data()
#>     Country Method EnergyType LastStage Unit                 Machine EuProduct
#> 1       GHA    PCM          E     Final ktoe             Automobiles        MD
#> 2       GHA    PCM          E     Final ktoe             Automobiles        MD
#> 3       GHA    PCM          E     Final ktoe             Automobiles        MD
#> 4       GHA    PCM          E     Final ktoe             Automobiles        MD
#> 5       GHA    PCM          E     Final ktoe             Diesel cars        MD
#> 6       GHA    PCM          E     Final ktoe             Diesel cars        MD
#> 7       GHA    PCM          E     Final ktoe             Diesel cars        MD
#> 8       GHA    PCM          E     Final ktoe             Diesel cars        MD
#> 9       GHA    PCM          E     Final ktoe           Diesel trucks        MD
#> 10      GHA    PCM          E     Final ktoe           Diesel trucks        MD
#> 11      GHA    PCM          E     Final ktoe           Diesel trucks        MD
#> 12      GHA    PCM          E     Final ktoe           Diesel trucks        MD
#> 13      GHA    PCM          E     Final ktoe         Electric motors        MD
#> 14      GHA    PCM          E     Final ktoe         Electric motors        MD
#> 15      GHA    PCM          E     Final ktoe         Electric motors        MD
#> 16      GHA    PCM          E     Final ktoe         Electric motors        MD
#> 17      GHA    PCM          E     Final ktoe                  Trucks        MD
#> 18      GHA    PCM          E     Final ktoe                  Trucks        MD
#> 19      GHA    PCM          E     Final ktoe                  Trucks        MD
#> 20      GHA    PCM          E     Final ktoe                  Trucks        MD
#> 21      GHA    PCM          E     Final ktoe            Boat engines        MD
#> 22      GHA    PCM          E     Final ktoe            Boat engines        MD
#> 23      GHA    PCM          E     Final ktoe            Boat engines        MD
#> 24      GHA    PCM          E     Final ktoe            Boat engines        MD
#> 25      GHA    PCM          E     Final ktoe           Diesel trains        MD
#> 26      GHA    PCM          E     Final ktoe           Diesel trains        MD
#> 27      GHA    PCM          E     Final ktoe           Diesel trains        MD
#> 28      GHA    PCM          E     Final ktoe           Diesel trains        MD
#> 29      GHA    PCM          E     Final ktoe                    Fans        MD
#> 30      GHA    PCM          E     Final ktoe                    Fans        MD
#> 31      GHA    PCM          E     Final ktoe                    Fans        MD
#> 32      GHA    PCM          E     Final ktoe                    Fans        MD
#> 33      GHA    PCM          E     Final ktoe         Electric lights     Light
#> 34      GHA    PCM          E     Final ktoe         Electric lights     Light
#> 35      GHA    PCM          E     Final ktoe         Electric lights     Light
#> 36      GHA    PCM          E     Final ktoe         Electric lights     Light
#> 37      GHA    PCM          E     Final ktoe     Industrial furnaces HTH.600.C
#> 38      GHA    PCM          E     Final ktoe     Industrial furnaces HTH.600.C
#> 39      GHA    PCM          E     Final ktoe     Industrial furnaces HTH.600.C
#> 40      GHA    PCM          E     Final ktoe     Industrial furnaces HTH.600.C
#> 41      GHA    PCM          E     Final ktoe        Electric heaters HTH.600.C
#> 42      GHA    PCM          E     Final ktoe        Electric heaters HTH.600.C
#> 43      GHA    PCM          E     Final ktoe        Electric heaters HTH.600.C
#> 44      GHA    PCM          E     Final ktoe        Electric heaters HTH.600.C
#> 45      GHA    PCM          E     Final ktoe             Wood stoves MTH.200.C
#> 46      GHA    PCM          E     Final ktoe             Wood stoves MTH.200.C
#> 47      GHA    PCM          E     Final ktoe             Wood stoves MTH.200.C
#> 48      GHA    PCM          E     Final ktoe             Wood stoves MTH.200.C
#> 49      GHA    PCM          E     Final ktoe     Industrial furnaces MTH.200.C
#> 50      GHA    PCM          E     Final ktoe     Industrial furnaces MTH.200.C
#> 51      GHA    PCM          E     Final ktoe     Industrial furnaces MTH.200.C
#> 52      GHA    PCM          E     Final ktoe     Industrial furnaces MTH.200.C
#> 53      GHA    PCM          E     Final ktoe        Electric heaters MTH.200.C
#> 54      GHA    PCM          E     Final ktoe        Electric heaters MTH.200.C
#> 55      GHA    PCM          E     Final ktoe        Electric heaters MTH.200.C
#> 56      GHA    PCM          E     Final ktoe        Electric heaters MTH.200.C
#> 57      GHA    PCM          E     Final ktoe                   Irons MTH.200.C
#> 58      GHA    PCM          E     Final ktoe                   Irons MTH.200.C
#> 59      GHA    PCM          E     Final ktoe                   Irons MTH.200.C
#> 60      GHA    PCM          E     Final ktoe                   Irons MTH.200.C
#> 61      GHA    PCM          E     Final ktoe              LPG stoves MTH.200.C
#> 62      GHA    PCM          E     Final ktoe              LPG stoves MTH.200.C
#> 63      GHA    PCM          E     Final ktoe              LPG stoves MTH.200.C
#> 64      GHA    PCM          E     Final ktoe              LPG stoves MTH.200.C
#> 65      GHA    PCM          E     Final ktoe         Electric stoves MTH.200.C
#> 66      GHA    PCM          E     Final ktoe         Electric stoves MTH.200.C
#> 67      GHA    PCM          E     Final ktoe         Electric stoves MTH.200.C
#> 68      GHA    PCM          E     Final ktoe         Electric stoves MTH.200.C
#> 69      GHA    PCM          E     Final ktoe         Wood cookstoves MTH.100.C
#> 70      GHA    PCM          E     Final ktoe         Wood cookstoves MTH.100.C
#> 71      GHA    PCM          E     Final ktoe         Wood cookstoves MTH.100.C
#> 72      GHA    PCM          E     Final ktoe         Wood cookstoves MTH.100.C
#> 73      GHA    PCM          E     Final ktoe         Charcoal stoves MTH.100.C
#> 74      GHA    PCM          E     Final ktoe         Charcoal stoves MTH.100.C
#> 75      GHA    PCM          E     Final ktoe         Charcoal stoves MTH.100.C
#> 76      GHA    PCM          E     Final ktoe         Charcoal stoves MTH.100.C
#> 77      GHA    PCM          E     Final ktoe         Kerosene stoves MTH.100.C
#> 78      GHA    PCM          E     Final ktoe         Kerosene stoves MTH.100.C
#> 79      GHA    PCM          E     Final ktoe         Kerosene stoves MTH.100.C
#> 80      GHA    PCM          E     Final ktoe         Kerosene stoves MTH.100.C
#> 81      GHA    PCM          E     Final ktoe              LPG stoves MTH.100.C
#> 82      GHA    PCM          E     Final ktoe              LPG stoves MTH.100.C
#> 83      GHA    PCM          E     Final ktoe              LPG stoves MTH.100.C
#> 84      GHA    PCM          E     Final ktoe              LPG stoves MTH.100.C
#> 85      GHA    PCM          E     Final ktoe         Electric stoves MTH.100.C
#> 86      GHA    PCM          E     Final ktoe         Electric stoves MTH.100.C
#> 87      GHA    PCM          E     Final ktoe         Electric stoves MTH.100.C
#> 88      GHA    PCM          E     Final ktoe         Electric stoves MTH.100.C
#> 89      GHA    PCM          E     Final ktoe  Non-energy consumption       NEU
#> 90      GHA    PCM          E     Final ktoe  Non-energy consumption       NEU
#> 91      GHA    PCM          E     Final ktoe  Non-energy consumption       NEU
#> 92      GHA    PCM          E     Final ktoe  Non-energy consumption       NEU
#> 93      ZAF    PCM          E     Final ktoe  Steam engines (trains)        MD
#> 94      ZAF    PCM          E     Final ktoe  Steam engines (trains)        MD
#> 95      ZAF    PCM          E     Final ktoe  Steam engines (trains)        MD
#> 96      ZAF    PCM          E     Final ktoe  Steam engines (trains)        MD
#> 97      ZAF    PCM          E     Final ktoe         Electric motors        MD
#> 98      ZAF    PCM          E     Final ktoe         Electric motors        MD
#> 99      ZAF    PCM          E     Final ktoe         Electric motors        MD
#> 100     ZAF    PCM          E     Final ktoe         Electric motors        MD
#> 101     ZAF    PCM          E     Final ktoe                  Trucks        MD
#> 102     ZAF    PCM          E     Final ktoe                  Trucks        MD
#> 103     ZAF    PCM          E     Final ktoe                  Trucks        MD
#> 104     ZAF    PCM          E     Final ktoe                  Trucks        MD
#> 105     ZAF    PCM          E     Final ktoe             Automobiles        MD
#> 106     ZAF    PCM          E     Final ktoe             Automobiles        MD
#> 107     ZAF    PCM          E     Final ktoe             Automobiles        MD
#> 108     ZAF    PCM          E     Final ktoe             Automobiles        MD
#> 109     ZAF    PCM          E     Final ktoe  Industry static motors        MD
#> 110     ZAF    PCM          E     Final ktoe  Industry static motors        MD
#> 111     ZAF    PCM          E     Final ktoe  Industry static motors        MD
#> 112     ZAF    PCM          E     Final ktoe  Industry static motors        MD
#> 113     ZAF    PCM          E     Final ktoe           Diesel trucks        MD
#> 114     ZAF    PCM          E     Final ktoe           Diesel trucks        MD
#> 115     ZAF    PCM          E     Final ktoe           Diesel trucks        MD
#> 116     ZAF    PCM          E     Final ktoe           Diesel trucks        MD
#> 117     ZAF    PCM          E     Final ktoe      Diesel automobiles        MD
#> 118     ZAF    PCM          E     Final ktoe      Diesel automobiles        MD
#> 119     ZAF    PCM          E     Final ktoe      Diesel automobiles        MD
#> 120     ZAF    PCM          E     Final ktoe      Diesel automobiles        MD
#> 121     ZAF    PCM          E     Final ktoe                    Fans        MD
#> 122     ZAF    PCM          E     Final ktoe                    Fans        MD
#> 123     ZAF    PCM          E     Final ktoe                    Fans        MD
#> 124     ZAF    PCM          E     Final ktoe                    Fans        MD
#> 125     ZAF    PCM          E     Final ktoe               Airplanes        MD
#> 126     ZAF    PCM          E     Final ktoe               Airplanes        MD
#> 127     ZAF    PCM          E     Final ktoe               Airplanes        MD
#> 128     ZAF    PCM          E     Final ktoe               Airplanes        MD
#> 129     ZAF    PCM          E     Final ktoe Industry static engines        MD
#> 130     ZAF    PCM          E     Final ktoe Industry static engines        MD
#> 131     ZAF    PCM          E     Final ktoe Industry static engines        MD
#> 132     ZAF    PCM          E     Final ktoe Industry static engines        MD
#> 133     ZAF    PCM          E     Final ktoe    Electric automobiles        MD
#> 134     ZAF    PCM          E     Final ktoe    Electric automobiles        MD
#> 135     ZAF    PCM          E     Final ktoe    Electric automobiles        MD
#> 136     ZAF    PCM          E     Final ktoe    Electric automobiles        MD
#> 137     ZAF    PCM          E     Final ktoe            Boat engines        MD
#> 138     ZAF    PCM          E     Final ktoe            Boat engines        MD
#> 139     ZAF    PCM          E     Final ktoe            Boat engines        MD
#> 140     ZAF    PCM          E     Final ktoe            Boat engines        MD
#> 141     ZAF    PCM          E     Final ktoe              Generators        MD
#> 142     ZAF    PCM          E     Final ktoe              Generators        MD
#> 143     ZAF    PCM          E     Final ktoe              Generators        MD
#> 144     ZAF    PCM          E     Final ktoe              Generators        MD
#> 145     ZAF    PCM          E     Final ktoe         Electric lights     Light
#> 146     ZAF    PCM          E     Final ktoe         Electric lights     Light
#> 147     ZAF    PCM          E     Final ktoe         Electric lights     Light
#> 148     ZAF    PCM          E     Final ktoe         Electric lights     Light
#> 149     ZAF    PCM          E     Final ktoe         Kerosene lights     Light
#> 150     ZAF    PCM          E     Final ktoe         Kerosene lights     Light
#> 151     ZAF    PCM          E     Final ktoe         Kerosene lights     Light
#> 152     ZAF    PCM          E     Final ktoe         Kerosene lights     Light
#> 153     ZAF    PCM          E     Final ktoe              Gas lights     Light
#> 154     ZAF    PCM          E     Final ktoe              Gas lights     Light
#> 155     ZAF    PCM          E     Final ktoe              Gas lights     Light
#> 156     ZAF    PCM          E     Final ktoe              Gas lights     Light
#> 157     ZAF    PCM          E     Final ktoe     Industrial furnaces HTH.600.C
#> 158     ZAF    PCM          E     Final ktoe     Industrial furnaces HTH.600.C
#> 159     ZAF    PCM          E     Final ktoe     Industrial furnaces HTH.600.C
#> 160     ZAF    PCM          E     Final ktoe     Industrial furnaces HTH.600.C
#> 161     ZAF    PCM          E     Final ktoe        Electric heaters HTH.600.C
#> 162     ZAF    PCM          E     Final ktoe        Electric heaters HTH.600.C
#> 163     ZAF    PCM          E     Final ktoe        Electric heaters HTH.600.C
#> 164     ZAF    PCM          E     Final ktoe        Electric heaters HTH.600.C
#> 165     ZAF    PCM          E     Final ktoe             Gas heaters HTH.600.C
#> 166     ZAF    PCM          E     Final ktoe             Gas heaters HTH.600.C
#> 167     ZAF    PCM          E     Final ktoe             Gas heaters HTH.600.C
#> 168     ZAF    PCM          E     Final ktoe             Gas heaters HTH.600.C
#> 169     ZAF    PCM          E     Final ktoe             Oil heaters HTH.600.C
#> 170     ZAF    PCM          E     Final ktoe             Oil heaters HTH.600.C
#> 171     ZAF    PCM          E     Final ktoe             Oil heaters HTH.600.C
#> 172     ZAF    PCM          E     Final ktoe             Oil heaters HTH.600.C
#> 173     ZAF    PCM          E     Final ktoe            Gas furnaces HTH.600.C
#> 174     ZAF    PCM          E     Final ktoe            Gas furnaces HTH.600.C
#> 175     ZAF    PCM          E     Final ktoe            Gas furnaces HTH.600.C
#> 176     ZAF    PCM          E     Final ktoe            Gas furnaces HTH.600.C
#> 177     ZAF    PCM          E     Final ktoe     Industrial furnaces MTH.200.C
#> 178     ZAF    PCM          E     Final ktoe     Industrial furnaces MTH.200.C
#> 179     ZAF    PCM          E     Final ktoe     Industrial furnaces MTH.200.C
#> 180     ZAF    PCM          E     Final ktoe     Industrial furnaces MTH.200.C
#> 181     ZAF    PCM          E     Final ktoe             Wood stoves MTH.200.C
#> 182     ZAF    PCM          E     Final ktoe             Wood stoves MTH.200.C
#> 183     ZAF    PCM          E     Final ktoe             Wood stoves MTH.200.C
#> 184     ZAF    PCM          E     Final ktoe             Wood stoves MTH.200.C
#> 185     ZAF    PCM          E     Final ktoe                   Irons MTH.200.C
#> 186     ZAF    PCM          E     Final ktoe                   Irons MTH.200.C
#> 187     ZAF    PCM          E     Final ktoe                   Irons MTH.200.C
#> 188     ZAF    PCM          E     Final ktoe                   Irons MTH.200.C
#> 189     ZAF    PCM          E     Final ktoe         Kerosene stoves MTH.200.C
#> 190     ZAF    PCM          E     Final ktoe         Kerosene stoves MTH.200.C
#> 191     ZAF    PCM          E     Final ktoe         Kerosene stoves MTH.200.C
#> 192     ZAF    PCM          E     Final ktoe         Kerosene stoves MTH.200.C
#> 193     ZAF    PCM          E     Final ktoe             Oil heaters MTH.200.C
#> 194     ZAF    PCM          E     Final ktoe             Oil heaters MTH.200.C
#> 195     ZAF    PCM          E     Final ktoe             Oil heaters MTH.200.C
#> 196     ZAF    PCM          E     Final ktoe             Oil heaters MTH.200.C
#> 197     ZAF    PCM          E     Final ktoe             Gas heaters MTH.200.C
#> 198     ZAF    PCM          E     Final ktoe             Gas heaters MTH.200.C
#> 199     ZAF    PCM          E     Final ktoe             Gas heaters MTH.200.C
#> 200     ZAF    PCM          E     Final ktoe             Gas heaters MTH.200.C
#> 201     ZAF    PCM          E     Final ktoe        Electric heaters MTH.200.C
#> 202     ZAF    PCM          E     Final ktoe        Electric heaters MTH.200.C
#> 203     ZAF    PCM          E     Final ktoe        Electric heaters MTH.200.C
#> 204     ZAF    PCM          E     Final ktoe        Electric heaters MTH.200.C
#> 205     ZAF    PCM          E     Final ktoe              LPG stoves MTH.200.C
#> 206     ZAF    PCM          E     Final ktoe              LPG stoves MTH.200.C
#> 207     ZAF    PCM          E     Final ktoe              LPG stoves MTH.200.C
#> 208     ZAF    PCM          E     Final ktoe              LPG stoves MTH.200.C
#> 209     ZAF    PCM          E     Final ktoe             Coal stoves MTH.200.C
#> 210     ZAF    PCM          E     Final ktoe             Coal stoves MTH.200.C
#> 211     ZAF    PCM          E     Final ktoe             Coal stoves MTH.200.C
#> 212     ZAF    PCM          E     Final ktoe             Coal stoves MTH.200.C
#> 213     ZAF    PCM          E     Final ktoe            Oil furnaces MTH.200.C
#> 214     ZAF    PCM          E     Final ktoe            Oil furnaces MTH.200.C
#> 215     ZAF    PCM          E     Final ktoe            Oil furnaces MTH.200.C
#> 216     ZAF    PCM          E     Final ktoe            Oil furnaces MTH.200.C
#> 217     ZAF    PCM          E     Final ktoe            Gas furnaces MTH.200.C
#> 218     ZAF    PCM          E     Final ktoe            Gas furnaces MTH.200.C
#> 219     ZAF    PCM          E     Final ktoe            Gas furnaces MTH.200.C
#> 220     ZAF    PCM          E     Final ktoe            Gas furnaces MTH.200.C
#> 221     ZAF    PCM          E     Final ktoe         Wood cookstoves MTH.100.C
#> 222     ZAF    PCM          E     Final ktoe         Wood cookstoves MTH.100.C
#> 223     ZAF    PCM          E     Final ktoe         Wood cookstoves MTH.100.C
#> 224     ZAF    PCM          E     Final ktoe         Wood cookstoves MTH.100.C
#> 225     ZAF    PCM          E     Final ktoe         Charcoal stoves MTH.100.C
#> 226     ZAF    PCM          E     Final ktoe         Charcoal stoves MTH.100.C
#> 227     ZAF    PCM          E     Final ktoe         Charcoal stoves MTH.100.C
#> 228     ZAF    PCM          E     Final ktoe         Charcoal stoves MTH.100.C
#> 229     ZAF    PCM          E     Final ktoe          LPG cookstoves MTH.100.C
#> 230     ZAF    PCM          E     Final ktoe          LPG cookstoves MTH.100.C
#> 231     ZAF    PCM          E     Final ktoe          LPG cookstoves MTH.100.C
#> 232     ZAF    PCM          E     Final ktoe          LPG cookstoves MTH.100.C
#> 233     ZAF    PCM          E     Final ktoe              LPG stoves MTH.100.C
#> 234     ZAF    PCM          E     Final ktoe              LPG stoves MTH.100.C
#> 235     ZAF    PCM          E     Final ktoe              LPG stoves MTH.100.C
#> 236     ZAF    PCM          E     Final ktoe              LPG stoves MTH.100.C
#> 237     ZAF    PCM          E     Final ktoe             Coal stoves MTH.100.C
#> 238     ZAF    PCM          E     Final ktoe             Coal stoves MTH.100.C
#> 239     ZAF    PCM          E     Final ktoe             Coal stoves MTH.100.C
#> 240     ZAF    PCM          E     Final ktoe             Coal stoves MTH.100.C
#> 241     ZAF    PCM          E     Final ktoe         Kerosene stoves MTH.100.C
#> 242     ZAF    PCM          E     Final ktoe         Kerosene stoves MTH.100.C
#> 243     ZAF    PCM          E     Final ktoe         Kerosene stoves MTH.100.C
#> 244     ZAF    PCM          E     Final ktoe         Kerosene stoves MTH.100.C
#> 245     ZAF    PCM          E     Final ktoe             Gas heaters MTH.100.C
#> 246     ZAF    PCM          E     Final ktoe             Gas heaters MTH.100.C
#> 247     ZAF    PCM          E     Final ktoe             Gas heaters MTH.100.C
#> 248     ZAF    PCM          E     Final ktoe             Gas heaters MTH.100.C
#> 249     ZAF    PCM          E     Final ktoe             Coal stoves  LTH.20.C
#> 250     ZAF    PCM          E     Final ktoe             Coal stoves  LTH.20.C
#> 251     ZAF    PCM          E     Final ktoe             Coal stoves  LTH.20.C
#> 252     ZAF    PCM          E     Final ktoe             Coal stoves  LTH.20.C
#> 253     ZAF    PCM          E     Final ktoe           Wood furnaces  LTH.20.C
#> 254     ZAF    PCM          E     Final ktoe           Wood furnaces  LTH.20.C
#> 255     ZAF    PCM          E     Final ktoe           Wood furnaces  LTH.20.C
#> 256     ZAF    PCM          E     Final ktoe           Wood furnaces  LTH.20.C
#> 257     ZAF    PCM          E     Final ktoe            Gas furnaces  LTH.20.C
#> 258     ZAF    PCM          E     Final ktoe            Gas furnaces  LTH.20.C
#> 259     ZAF    PCM          E     Final ktoe            Gas furnaces  LTH.20.C
#> 260     ZAF    PCM          E     Final ktoe            Gas furnaces  LTH.20.C
#> 261     ZAF    PCM          E     Final ktoe         Kerosene stoves  LTH.20.C
#> 262     ZAF    PCM          E     Final ktoe         Kerosene stoves  LTH.20.C
#> 263     ZAF    PCM          E     Final ktoe         Kerosene stoves  LTH.20.C
#> 264     ZAF    PCM          E     Final ktoe         Kerosene stoves  LTH.20.C
#> 265     ZAF    PCM          E     Final ktoe              LPG stoves  LTH.20.C
#> 266     ZAF    PCM          E     Final ktoe              LPG stoves  LTH.20.C
#> 267     ZAF    PCM          E     Final ktoe              LPG stoves  LTH.20.C
#> 268     ZAF    PCM          E     Final ktoe              LPG stoves  LTH.20.C
#> 269     ZAF    PCM          E     Final ktoe            Oil furnaces  LTH.20.C
#> 270     ZAF    PCM          E     Final ktoe            Oil furnaces  LTH.20.C
#> 271     ZAF    PCM          E     Final ktoe            Oil furnaces  LTH.20.C
#> 272     ZAF    PCM          E     Final ktoe            Oil furnaces  LTH.20.C
#> 273     ZAF    PCM          E     Final ktoe  Non-energy consumption       NEU
#> 274     ZAF    PCM          E     Final ktoe  Non-energy consumption       NEU
#> 275     ZAF    PCM          E     Final ktoe  Non-energy consumption       NEU
#> 276     ZAF    PCM          E     Final ktoe  Non-energy consumption       NEU
#>             Quantity MaximumValues         1971         2000
#> 1       Edot_machine  6.192510e+02 2.172315e+02 6.192510e+02
#> 2   Edot_machine [%]  1.156388e-01 8.391282e-02 1.156388e-01
#> 3              etafu            NA 3.000000e-01 3.100000e-01
#> 4               phiu            NA 1.000000e+00 1.000000e+00
#> 5       Edot_machine  1.877078e+02 5.946665e+01 1.877078e+02
#> 6   Edot_machine [%]  3.505250e-02 2.297095e-02 3.505250e-02
#> 7              etafu            NA 3.000000e-01 3.000000e-01
#> 8               phiu            NA 1.000000e+00 1.000000e+00
#> 9       Edot_machine  1.877078e+02 5.946665e+01 1.877078e+02
#> 10  Edot_machine [%]  3.505250e-02 2.297095e-02 3.505250e-02
#> 11             etafu            NA 3.000000e-01 3.000000e-01
#> 12              phiu            NA 1.000000e+00 1.000000e+00
#> 13      Edot_machine  1.870163e+02 6.383495e+01 1.870163e+02
#> 14  Edot_machine [%]  3.492338e-02 2.465835e-02 3.492338e-02
#> 15             etafu            NA 8.500000e-01 9.500000e-01
#> 16              phiu            NA 1.000000e+00 1.000000e+00
#> 17      Edot_machine  7.032580e+01 1.499595e+01 7.032580e+01
#> 18  Edot_machine [%]  1.313262e-02 5.792678e-03 1.313262e-02
#> 19             etafu            NA 3.000000e-01 3.100000e-01
#> 20              phiu            NA 1.000000e+00 1.000000e+00
#> 21      Edot_machine  6.316280e+01 1.226230e+01 6.316280e+01
#> 22  Edot_machine [%]  1.179500e-02 4.736716e-03 1.179500e-02
#> 23             etafu            NA 3.000000e-01 3.100000e-01
#> 24              phiu            NA 1.000000e+00 1.000000e+00
#> 25      Edot_machine  2.585510e+01 1.514280e+01 2.585510e+01
#> 26  Edot_machine [%]  5.849404e-03 5.849404e-03 4.828174e-03
#> 27             etafu            NA 3.000000e-01 3.100000e-01
#> 28              phiu            NA 1.000000e+00 1.000000e+00
#> 29      Edot_machine  1.907567e+01 2.115225e+00 1.907567e+01
#> 30  Edot_machine [%]  3.562184e-03 8.170751e-04 3.562184e-03
#> 31             etafu            NA 1.000000e-01 1.200000e-01
#> 32              phiu            NA 1.000000e+00 1.000000e+00
#> 33      Edot_machine  1.589037e+02 1.411179e+02 1.589037e+02
#> 34  Edot_machine [%]  5.451143e-02 5.451143e-02 2.967363e-02
#> 35             etafu            NA 2.000000e-01 2.000000e-01
#> 36              phiu            NA 1.261201e-01 1.830161e-01
#> 37      Edot_machine  2.642350e+01 1.953040e+01 2.642350e+01
#> 38  Edot_machine [%]  7.544259e-03 7.544259e-03 4.934317e-03
#> 39             etafu            NA 8.500000e-01 8.600000e-01
#> 40              phiu            NA 6.585352e-01 6.585352e-01
#> 41      Edot_machine  1.584695e+01 1.584695e+01           NA
#> 42  Edot_machine [%]  6.121405e-03 6.121405e-03           NA
#> 43             etafu            NA 8.500000e-01           NA
#> 44              phiu            NA 6.585352e-01           NA
#> 45      Edot_machine  6.852728e+02 1.456960e+02 6.852728e+02
#> 46  Edot_machine [%]  1.279677e-01 5.627987e-02 1.279677e-01
#> 47             etafu            NA 8.500000e-01 8.600000e-01
#> 48              phiu            NA 3.698616e-01 3.698616e-01
#> 49      Edot_machine  2.166498e+02 1.369423e+02 2.166498e+02
#> 50  Edot_machine [%]  5.289846e-02 5.289846e-02 4.045712e-02
#> 51             etafu            NA 8.500000e-01 8.600000e-01
#> 52              phiu            NA 3.698616e-01 3.698616e-01
#> 53      Edot_machine  1.480998e+02 9.355120e+00 1.480998e+02
#> 54  Edot_machine [%]  2.765611e-02 3.613722e-03 2.765611e-02
#> 55             etafu            NA 9.700000e-01 9.900000e-01
#> 56              phiu            NA 3.698616e-01 3.698616e-01
#> 57      Edot_machine  1.907567e+01 2.115225e+00 1.907567e+01
#> 58  Edot_machine [%]  3.562184e-03 8.170751e-04 3.562184e-03
#> 59             etafu            NA 8.500000e-01 8.500000e-01
#> 60              phiu            NA 3.698616e-01 3.698616e-01
#> 61      Edot_machine  4.519000e+00           NA 4.519000e+00
#> 62  Edot_machine [%]  8.438768e-04           NA 8.438768e-04
#> 63             etafu            NA           NA 4.800000e-01
#> 64              phiu            NA           NA 3.698616e-01
#> 65      Edot_machine  1.160800e+00 1.160800e+00           NA
#> 66  Edot_machine [%]  4.483971e-04 4.483971e-04           NA
#> 67             etafu            NA 6.500000e-01           NA
#> 68              phiu            NA 3.698616e-01           NA
#> 69      Edot_machine  2.111732e+03 1.464125e+03 2.111732e+03
#> 70  Edot_machine [%]  5.655665e-01 5.655665e-01 3.943443e-01
#> 71             etafu            NA 1.200000e-01 1.200000e-01
#> 72              phiu            NA 2.009916e-01 2.009916e-01
#> 73      Edot_machine  5.156874e+02 1.191745e+02 5.156874e+02
#> 74  Edot_machine [%]  9.629932e-02 4.603507e-02 9.629932e-02
#> 75             etafu            NA 1.400000e-01 1.400000e-01
#> 76              phiu            NA 2.009916e-01 2.009916e-01
#> 77      Edot_machine  8.369160e+01 8.369160e+01 5.335340e+01
#> 78  Edot_machine [%]  3.232863e-02 3.232863e-02 9.963199e-03
#> 79             etafu            NA 1.600000e-01 1.600000e-01
#> 80              phiu            NA 2.009916e-01 2.009916e-01
#> 81      Edot_machine  3.615170e+01 3.389200e+00 3.615170e+01
#> 82  Edot_machine [%]  6.750959e-03 1.309190e-03 6.750959e-03
#> 83             etafu            NA 1.800000e-01 1.800000e-01
#> 84              phiu            NA 2.009916e-01 2.009916e-01
#> 85      Edot_machine  1.907567e+01 2.115225e+00 1.907567e+01
#> 86  Edot_machine [%]  3.562184e-03 8.170751e-04 3.562184e-03
#> 87             etafu            NA 7.000000e-01 7.000000e-01
#> 88              phiu            NA 2.009916e-01 2.009916e-01
#> 89      Edot_machine  6.262540e+01 1.805670e+01 6.821440e+01
#> 90  Edot_machine [%]  1.154755e-02 6.926680e-03 1.257811e-02
#> 91             etafu            NA 1.000000e-06 1.000000e-06
#> 92              phiu            NA 1.000000e+00 1.000000e+00
#> 93      Edot_machine  3.295513e+03 3.295513e+03 1.926513e+02
#> 94  Edot_machine [%]  9.922186e-02 9.922186e-02 3.670736e-03
#> 95             etafu            NA 6.000000e-01 6.200000e-01
#> 96              phiu            NA 1.000000e+00 1.000000e+00
#> 97      Edot_machine  5.168693e+03 1.030447e+03 5.168693e+03
#> 98  Edot_machine [%]  9.848315e-02 3.102487e-02 9.848315e-02
#> 99             etafu            NA 8.700000e-01 9.600000e-01
#> 100             phiu            NA 1.000000e+00 1.000000e+00
#> 101     Edot_machine  4.109393e+03 1.850616e+03 4.109393e+03
#> 102 Edot_machine [%]  7.829947e-02 5.571866e-02 7.829947e-02
#> 103            etafu            NA 3.000000e-01 3.100000e-01
#> 104             phiu            NA 1.000000e+00 1.000000e+00
#> 105     Edot_machine  4.029254e+03 2.305665e+03 4.029254e+03
#> 106 Edot_machine [%]  7.677253e-02 6.941935e-02 7.677253e-02
#> 107            etafu            NA 3.000000e-01 3.100000e-01
#> 108             phiu            NA 1.000000e+00 1.000000e+00
#> 109     Edot_machine  2.496819e+03 1.190714e+03 2.496819e+03
#> 110 Edot_machine [%]  4.757384e-02 3.585020e-02 4.757384e-02
#> 111            etafu            NA 3.000000e-01 3.100000e-01
#> 112             phiu            NA 1.000000e+00 1.000000e+00
#> 113     Edot_machine  1.673811e+03 7.115315e+02 1.673811e+03
#> 114 Edot_machine [%]  3.189242e-02 2.142291e-02 3.189242e-02
#> 115            etafu            NA 3.000000e-01 3.100000e-01
#> 116             phiu            NA 1.000000e+00 1.000000e+00
#> 117     Edot_machine  1.672812e+03 7.115315e+02 1.672812e+03
#> 118 Edot_machine [%]  3.187339e-02 2.142291e-02 3.187339e-02
#> 119            etafu            NA 3.000000e-01 3.100000e-01
#> 120             phiu            NA 1.000000e+00 1.000000e+00
#> 121     Edot_machine  7.398108e+02 2.086328e+02 7.398108e+02
#> 122 Edot_machine [%]  1.409619e-02 6.281552e-03 1.409619e-02
#> 123            etafu            NA 1.200000e-01 1.200000e-01
#> 124             phiu            NA 1.000000e+00 1.000000e+00
#> 125     Edot_machine  6.141134e+02 1.715057e+02 6.141134e+02
#> 126 Edot_machine [%]  1.170118e-02 5.163722e-03 1.170118e-02
#> 127            etafu            NA 3.000000e-01 3.100000e-01
#> 128             phiu            NA 1.000000e+00 1.000000e+00
#> 129     Edot_machine  5.185681e+02           NA 5.185681e+02
#> 130 Edot_machine [%]  9.880684e-03           NA 9.880684e-03
#> 131            etafu            NA           NA 3.100000e-01
#> 132             phiu            NA           NA 1.000000e+00
#> 133     Edot_machine  1.620809e+02 5.176270e+01 1.620809e+02
#> 134 Edot_machine [%]  3.088254e-03 1.558480e-03 3.088254e-03
#> 135            etafu            NA 4.500000e-01 5.000000e-01
#> 136             phiu            NA 1.000000e+00 1.000000e+00
#> 137     Edot_machine  1.026154e+02           NA 1.026154e+02
#> 138 Edot_machine [%]  1.955211e-03           NA 1.955211e-03
#> 139            etafu            NA           NA 3.100000e-01
#> 140             phiu            NA           NA 1.000000e+00
#> 141     Edot_machine  3.366390e+01           NA 3.366390e+01
#> 142 Edot_machine [%]  6.414246e-04           NA 6.414246e-04
#> 143            etafu            NA           NA 3.100000e-01
#> 144             phiu            NA           NA 1.000000e+00
#> 145     Edot_machine  6.263618e+03 1.413929e+03 6.263618e+03
#> 146 Edot_machine [%]  1.193456e-01 4.257082e-02 1.193456e-01
#> 147            etafu            NA 2.000000e-01 2.000000e-01
#> 148             phiu            NA 1.261201e-01 1.830161e-01
#> 149     Edot_machine  1.497862e+02           NA 1.497862e+02
#> 150 Edot_machine [%]  2.853994e-03           NA 2.853994e-03
#> 151            etafu            NA           NA 2.000000e-01
#> 152             phiu            NA           NA 1.830161e-01
#> 153     Edot_machine  1.154460e+02 4.325020e+01 1.154460e+02
#> 154 Edot_machine [%]  2.199684e-03 1.302184e-03 2.199684e-03
#> 155            etafu            NA 2.000000e-01 2.000000e-01
#> 156             phiu            NA 1.261201e-01 1.830161e-01
#> 157     Edot_machine  7.157481e+03 7.157481e+03 4.612028e+03
#> 158 Edot_machine [%]  2.154986e-01 2.154986e-01 8.787657e-02
#> 159            etafu            NA 8.500000e-01 9.000000e-01
#> 160             phiu            NA 6.585352e-01 6.585352e-01
#> 161     Edot_machine  1.683990e+03 7.889936e+01 1.683990e+03
#> 162 Edot_machine [%]  3.208637e-02 2.375515e-03 3.208637e-02
#> 163            etafu            NA 9.000000e-01 9.500000e-01
#> 164             phiu            NA 6.585352e-01 6.585352e-01
#> 165     Edot_machine  1.052345e+03           NA 1.052345e+03
#> 166 Edot_machine [%]  2.005116e-02           NA 2.005116e-02
#> 167            etafu            NA           NA 8.500000e-01
#> 168             phiu            NA           NA 6.585352e-01
#> 169     Edot_machine  5.799369e+02 5.799369e+02           NA
#> 170 Edot_machine [%]  1.746084e-02 1.746084e-02           NA
#> 171            etafu            NA 8.000000e-01           NA
#> 172             phiu            NA 6.585352e-01           NA
#> 173     Edot_machine  1.706363e+02 2.914880e+01 1.706363e+02
#> 174 Edot_machine [%]  3.251267e-03 8.776168e-04 3.251267e-03
#> 175            etafu            NA 8.000000e-01 8.500000e-01
#> 176             phiu            NA 6.585352e-01 6.585352e-01
#> 177     Edot_machine  3.459107e+03 2.080060e+03 3.459107e+03
#> 178 Edot_machine [%]  6.590907e-02 6.262680e-02 6.590907e-02
#> 179            etafu            NA 8.000000e-01 8.500000e-01
#> 180             phiu            NA 3.698616e-01 3.698616e-01
#> 181     Edot_machine  1.508431e+03 7.643069e+02 1.508431e+03
#> 182 Edot_machine [%]  2.874132e-02 2.301188e-02 2.874132e-02
#> 183            etafu            NA 2.000000e-01 2.500000e-01
#> 184             phiu            NA 3.698616e-01 3.698616e-01
#> 185     Edot_machine  1.233018e+03 3.477214e+02 1.233018e+03
#> 186 Edot_machine [%]  2.349366e-02 1.046925e-02 2.349366e-02
#> 187            etafu            NA 3.000000e-01 3.500000e-01
#> 188             phiu            NA 3.698616e-01 3.698616e-01
#> 189     Edot_machine  5.596876e+02 5.596876e+02 5.165000e+00
#> 190 Edot_machine [%]  1.685117e-02 1.685117e-02 9.841278e-05
#> 191            etafu            NA 1.400000e-01 1.400000e-01
#> 192             phiu            NA 3.698616e-01 3.698616e-01
#> 193     Edot_machine  5.235502e+02 5.235502e+02           NA
#> 194 Edot_machine [%]  1.576314e-02 1.576314e-02           NA
#> 195            etafu            NA 4.000000e-01           NA
#> 196             phiu            NA 3.698616e-01           NA
#> 197     Edot_machine  5.746345e+02 6.605760e+01 5.746345e+02
#> 198 Edot_machine [%]  1.094896e-02 1.988873e-03 1.094896e-02
#> 199            etafu            NA 4.000000e-01 4.500000e-01
#> 200             phiu            NA 3.698616e-01 3.698616e-01
#> 201     Edot_machine  4.777397e+02 4.315563e+01 4.777397e+02
#> 202 Edot_machine [%]  9.102749e-03 1.299337e-03 9.102749e-03
#> 203            etafu            NA 4.000000e-01 4.500000e-01
#> 204             phiu            NA 3.698616e-01 3.698616e-01
#> 205     Edot_machine  1.197526e+02 1.197526e+02 2.680800e+00
#> 206 Edot_machine [%]  3.605531e-03 3.605531e-03 5.107938e-05
#> 207            etafu            NA 3.000000e-01 3.500000e-01
#> 208             phiu            NA 3.698616e-01 3.698616e-01
#> 209     Edot_machine  7.672980e+01 7.672980e+01           NA
#> 210 Edot_machine [%]  2.310193e-03 2.310193e-03           NA
#> 211            etafu            NA 2.500000e-01           NA
#> 212             phiu            NA 3.698616e-01           NA
#> 213     Edot_machine  1.098900e+01           NA 1.098900e+01
#> 214 Edot_machine [%]  2.093820e-04           NA 2.093820e-04
#> 215            etafu            NA           NA 9.000000e-01
#> 216             phiu            NA           NA 3.698616e-01
#> 217     Edot_machine  3.075000e+00           NA 3.075000e+00
#> 218 Edot_machine [%]  5.859038e-05           NA 5.859038e-05
#> 219            etafu            NA           NA 9.000000e-01
#> 220             phiu            NA           NA 3.698616e-01
#> 221     Edot_machine  2.787475e+03 1.934652e+03 2.787475e+03
#> 222 Edot_machine [%]  5.824881e-02 5.824881e-02 5.311194e-02
#> 223            etafu            NA 1.200000e-01 1.200000e-01
#> 224             phiu            NA 2.009916e-01 2.009916e-01
#> 225     Edot_machine  6.584026e+02 1.691980e+01 6.584026e+02
#> 226 Edot_machine [%]  1.254506e-02 5.094241e-04 1.254506e-02
#> 227            etafu            NA 1.400000e-01 1.400000e-01
#> 228             phiu            NA 2.009916e-01 2.009916e-01
#> 229     Edot_machine  1.452114e+02           NA 1.452114e+02
#> 230 Edot_machine [%]  2.766826e-03           NA 2.766826e-03
#> 231            etafu            NA           NA 1.600000e-01
#> 232             phiu            NA           NA 2.009916e-01
#> 233     Edot_machine  4.691440e+01           NA 4.691440e+01
#> 234 Edot_machine [%]  8.938968e-04           NA 8.938968e-04
#> 235            etafu            NA           NA 1.600000e-01
#> 236             phiu            NA           NA 2.009916e-01
#> 237     Edot_machine  4.449040e+01           NA 4.449040e+01
#> 238 Edot_machine [%]  8.477104e-04           NA 8.477104e-04
#> 239            etafu            NA           NA 1.600000e-01
#> 240             phiu            NA           NA 2.009916e-01
#> 241     Edot_machine  3.099000e+00           NA 3.099000e+00
#> 242 Edot_machine [%]  5.904767e-05           NA 5.904767e-05
#> 243            etafu            NA           NA 1.600000e-01
#> 244             phiu            NA           NA 2.009916e-01
#> 245     Edot_machine  5.223600e-01           NA 5.223600e-01
#> 246 Edot_machine [%]  9.952934e-06           NA 9.952934e-06
#> 247            etafu            NA           NA 9.000000e-01
#> 248             phiu            NA           NA 2.009916e-01
#> 249     Edot_machine  3.140119e+03 3.140119e+03 1.490106e+03
#> 250 Edot_machine [%]  9.454323e-02 9.454323e-02 2.839215e-02
#> 251            etafu            NA 1.600000e-01 1.600000e-01
#> 252             phiu            NA 1.677008e-02 1.677008e-02
#> 253     Edot_machine  2.787475e+03 1.934652e+03 2.787475e+03
#> 254 Edot_machine [%]  5.824881e-02 5.824881e-02 5.311194e-02
#> 255            etafu            NA 8.500000e-01 8.500000e-01
#> 256             phiu            NA 1.677008e-02 1.677008e-02
#> 257     Edot_machine  9.715247e+02 7.756520e+02 9.715247e+02
#> 258 Edot_machine [%]  2.335346e-02 2.335346e-02 1.851122e-02
#> 259            etafu            NA 8.500000e-01 8.500000e-01
#> 260             phiu            NA 1.677008e-02 1.677008e-02
#> 261     Edot_machine  5.516266e+02           NA 5.516266e+02
#> 262 Edot_machine [%]  1.051057e-02           NA 1.051057e-02
#> 263            etafu            NA           NA 3.000000e-01
#> 264             phiu            NA           NA 1.677008e-02
#> 265     Edot_machine  1.452114e+02           NA 1.452114e+02
#> 266 Edot_machine [%]  2.766826e-03           NA 2.766826e-03
#> 267            etafu            NA           NA 3.200000e-01
#> 268             phiu            NA           NA 1.677008e-02
#> 269     Edot_machine  1.398600e+01           NA 1.398600e+01
#> 270 Edot_machine [%]  2.664862e-04           NA 2.664862e-04
#> 271            etafu            NA           NA 8.500000e-01
#> 272             phiu            NA           NA 1.677008e-02
#> 273     Edot_machine  5.654902e+03 9.323913e+02 5.654902e+03
#> 274 Edot_machine [%]  9.726590e-02 2.730604e-02 9.726590e-02
#> 275            etafu            NA 1.000000e-06 1.000000e-06
#> 276             phiu            NA 1.000000e+00 1.000000e+00