Prepare for PSUT analysis

Converts a tidy IEA data frame into a PSUT data frame by collapsing the IEA data into PSUT matrices (R, U, V, and Y).

prep_psut(
  .tidy_iea_df,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  supply = IEATools::ledger_sides$supply,
  consumption = IEATools::ledger_sides$consumption,
  matnames = IEATools::mat_meta_cols$matnames,
  rownames = IEATools::mat_meta_cols$rownames,
  colnames = IEATools::mat_meta_cols$colnames,
  rowtypes = IEATools::mat_meta_cols$rowtypes,
  coltypes = IEATools::mat_meta_cols$coltypes,
  matvals = IEATools::psut_cols$matvals
)

Arguments

.tidy_iea_df
a tidy data frame that has been specified with specify_all().

ledger_side, flow_aggregation_point, flow, product, e_dot, unit
See IEATools::iea_cols.

supply, consumption
See IEATools::ledger_sides.

matnames, rownames, colnames, rowtypes, coltypes
See IEATools::mat_meta_cols.

matvals
See IEATools::psut_cols.

Value

a wide PSUT data frame with metadata columns and columns named for each type of matrix

Details

This function bundles several others:

add_psut_matnames()

add_row_col_meta()

collapse_to_tidy_psut()

Furthermore, it extracts S_units matrices using extract_S_units_from_tidy() and adds those matrices to the data frame.

Examples

library(dplyr)
library(tidyr)

#> 
#> Attaching package: ‘tidyr’

#> The following object is masked from ‘package:magrittr’:
#> 
#>     extract

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

Simple <- load_tidy_iea_df() %>%
  specify_all() %>%
  prep_psut() %>%
  pivot_longer(cols = c(R, U_EIOU, U_excl_EIOU, V, Y, S_units),
               names_to = "matnames",
               values_to = "matval_simple")
S_units <- load_tidy_iea_df() %>%
  specify_all() %>%
  extract_S_units_from_tidy()
Complicated <- load_tidy_iea_df() %>%
  specify_all() %>%
  add_psut_matnames() %>%
  add_row_col_meta() %>%
  collapse_to_tidy_psut() %>%
  spread(key = matnames, value = matvals) %>%
  full_join(S_units, by = c("Method", "Energy.type", "Last.stage",
                            "Country", "Year")) %>%
  gather(key = matnames, value = matvals, R, U_EIOU, U_excl_EIOU,
                                        V, Y, S_units) %>%
  rename(matval_complicated = matvals)
# Simple and Complicated are same.
full_join(Simple, Complicated, by = c("Method", "Energy.type",
                                      "Last.stage", "Country",
                                      "Year", "matnames")) %>%
  dplyr::mutate(
    same = matsbyname::equal_byname(matval_simple, matval_complicated)
  ) %>%
  magrittr::extract2("same") %>%
  as.logical() %>%
  all()

#> [1] TRUE