Identify all energy types supplied or consumed by Production, Transformation processes, or Energy industry own use

Sometimes, it is helpful to know all types of energy supplied or consumed by Production, Transformation processes, or Energy industry own use. This function (optionally) reads an IEA data file or loads an IEA data frame and builds a named list of energy types supplied or consumed by Production, Transformation processes, or Energy industry own use.

prod_tp_eiou_energy_carriers(
  file_path = sample_iea_data_path(),
  iea_df = IEATools::load_tidy_iea_df(file_path),
  side = c("Consumption", "Supply"),
  flow_aggregation_point = "Flow.aggregation.point",
  production = "Production",
  transformation_processes = "Transformation processes",
  eiou = "Energy industry own use",
  stage = c(production, transformation_processes, eiou),
  e_dot = "E.dot",
  flow = "Flow",
  product = "Product"
)

Arguments

file_path
the path to the IEA data file (optional)

iea_df
a data frame containing IEA data. Default is IEATools::load_tidy_iea_df(file_path).

side
refers to the "Consumptiion" or "Supply" side of Production, Transformation processes, or Energy industry own use. One of "Consumption" or "Supply". Default is "Consumption".

flow_aggregation_point
the flow aggregation point column in iea_df. Default is "Flow.aggregation.point".

production
the string indicating the production flow. Default is "Production".

transformation_processes
the string indicating the transformation process stage. Default is "Transformation processes".

eiou
the string indicating the energy industry own use flow. Default is "Energy industry own use".

stage
the string indicating the stage for the analysis. One of production, transformation_processes, or eiou. Default is production.

e_dot
the energy flow rate column in iea_df. Default is "E.dot".

flow
the flow column in iea_df. Default is "Flow".

product
the product column in iea_df. Default is "Product".

Value

a list of string vectors

Details

The names in the returned list are the Production, Transformation processes, or Energy industry own use industries in iea_df. The items in the returned list are vectors of energy types produced or consumed by the corresponding industries.

Examples

prod_tp_eiou_energy_carriers()

#> named list()
#> attr(,"ptype")
#> character(0)