Physical Supply-Use Table (PSUT) data frame column names

A string list containing named names of columns in PSUT data frames. Items in the list provide default values for column name function arguments throughout the IEATools package.

psut_cols

Format

A string list with 12 entries

resources,R

The name of a column in a wide data frame containing resource (R) matrices.

use,U

The name of a column in a wide data frame containing use (U) matrices.

U_excl_eiou

The name of a column in a wide data frame containing use (U) matrices that exclude energy industry own use.

U_EIOU

The name of a column in a wide data frame containing use (U) matrices that contain exclusively energy industry own use.

make,V

The name of a column in a wide data frame containing make (V) matrices.

final_demand,Y

The name of a column in a wide data frame containing final demand (Y) matrices.

s_units

The name of a column in a wide data frame containing unit summation (S_units) matrices.

matvals

The name of a column in a tidy data frame containing matrices.

Details

Note that some of the values are repeated, thereby providing synonyms. E.g., both resources and R point to the "R" column name.

Examples

psut_cols

#> $resources
#> [1] "R"
#> 
#> $R
#> [1] "R"
#> 
#> $use
#> [1] "U"
#> 
#> $U
#> [1] "U"
#> 
#> $U_excl_eiou
#> [1] "U_excl_EIOU"
#> 
#> $U_eiou
#> [1] "U_EIOU"
#> 
#> $make
#> [1] "V"
#> 
#> $V
#> [1] "V"
#> 
#> $final_demand
#> [1] "Y"
#> 
#> $Y
#> [1] "Y"
#> 
#> $s_units
#> [1] "S_units"
#> 
#> $matvals
#> [1] "matvals"
#>