Physical Supply-Use Table (PSUT) data frame column names

A string list containing named names of columns in PSUT data frames. Items in the list provide default values for column name function arguments throughout the IEATools package.

Usage

psut_cols

Format

A string list with 14 entries.

resources,R

The name of a column in a wide-by-matrices data frame containing resource (R) matrices.

U_feed

The name of a column in a wide-by-matrices data frame containing use (U) matrices that exclude energy industry own use.

U_eiou

The name of a column in a wide-by-matrices data frame containing use (U) matrices that contain exclusively energy industry own use.

U

The name of a column in a wide data-by-matrices frame containing use (U) matrices that are the sum of U_feed and U_eiou matrices.

r_eiou

The name of a column in a wide-by-matrices data frame containing the ratio of U_eiou and U matrices.

make,V

The name of a column in a wide-by-matrices data frame containing make (V) matrices.

final_demand,Y

The name of a column in a wide-by-matrices data frame containing final demand (Y) matrices.

s_units

The name of a column in a wide-by-matrices data frame containing unit summation (S_units) matrices.

matvals

The name of a column in a tidy data frame containing matrices.

Details

Note that some of the values are repeated, thereby providing synonyms. E.g., both resources and R point to the "R" column name.

Examples

psut_cols
#> $resources
#> [1] "R"
#> 
#> $R
#> [1] "R"
#> 
#> $use
#> [1] "U"
#> 
#> $U
#> [1] "U"
#> 
#> $U_feed
#> [1] "U_feed"
#> 
#> $U_eiou
#> [1] "U_EIOU"
#> 
#> $r_eiou
#> [1] "r_EIOU"
#> 
#> $make
#> [1] "V"
#> 
#> $V
#> [1] "V"
#> 
#> $final_demand
#> [1] "Y"
#> 
#> $Y
#> [1] "Y"
#> 
#> $B
#> [1] "B"
#> 
#> $s_units
#> [1] "S_units"
#> 
#> $matvals
#> [1] "matvals"
#>