Remove specification suffixes from a column

Flow and Product columns of IEA data frames may have been "specified" to contain a suffix of the form " [of Natural gas\"], for example. This function strips away the suffix.

Identification of parenthetical notation delimiters is determined by the notations argument.

remove_suffix_specifications(
  .df,
  col,
  unsuffixed_col,
  notations = list(IEATools::of_notation, IEATools::from_notation)
)

Arguments

.df
The data frame in which col exists.

col
The string name of the column in .df to be de-specified.

unsuffixed_col
The string name of the column in the output data frame to contain the un-suffixed version of col.

notations
The notations used for row and column names. See matsbyname::notation_vec(). Default is list(RCLabels::of_notation, RCLabels::from_notation), because both RCLabels::of_notation and RCLabels::from_notation can be used in the Flow column of an IEA data frame.

Value

A version of .df with suffixes removed from the col column and the result placed in the despecified_col column.

Examples

library(dplyr)
load_tidy_iea_df() %>% 
  specify_all() %>% 
  remove_suffix_specifications(col = "Flow", unsuffixed_col = "clean_Flow") %>% 
  select(Flow, Product, E.dot, clean_Flow) %>% 
  filter(endsWith(Flow, bracket_notation[["suff_end"]]))

#> # A tibble: 101 × 4
#>    Flow                            Product              E.dot clean_Flow        
#>    <chr>                           <chr>                <dbl> <chr>             
#>  1 Exports [of Fuel oil]           Fuel oil           -178.   Exports           
#>  2 Exports [of Motor gasoline exc… Motor gasoline ex…   -4.28 Exports           
#>  3 Exports [of Other kerosene]     Other kerosene       -4.18 Exports           
#>  4 Imports [of Aviation gasoline]  Aviation gasoline     1.07 Imports           
#>  5 Imports [of Crude oil]          Crude oil           916.   Imports           
#>  6 Imports [of Kerosene type jet … Kerosene type jet…   21.3  Imports           
#>  7 Imports [of Lubricants]         Lubricants           18.1  Imports           
#>  8 Imports [of Other kerosene]     Other kerosene        1.05 Imports           
#>  9 Exports to World aviation bunk… Aviation gasoline    -1.07 Exports to World …
#> 10 Exports to World aviation bunk… Kerosene type jet…  -43.7  Exports to World …
#> # … with 91 more rows