Slurp an IEA extended energy balance data file

This is an internal helper function. This function reads an IEA extended energy balances .csv file and converts it to a data frame with appropriately-labeled columns. One of iea_file or text must be specified, but not both. The first line of iea_file or text is expected to start with expected_start_1st_line, and the second line is expected to start with expected_2nd_line_start, and it may have any number of commas appended. (The extra commas might come from opening and re-saving the file in Excel.) Alternatively, the file may have a first line of expected_simple_start. If none of these conditions are not met, execution is halted, and an error message is provided. Files should have a return character at the end of their final line.

slurp_iea_to_raw_df(
  .iea_file = NULL,
  text = NULL,
  expected_1st_line_start = ",,TIME",
  expected_2nd_line_start = "COUNTRY,FLOW,PRODUCT",
  expected_simple_start = expected_2nd_line_start
)

Arguments

.iea_file
the path to the raw IEA data file for which quality assurance is desired

text
a string containing text to be parsed as an IEA file.

expected_1st_line_start
the expected start of the first line of iea_file. Default is ",,TIME".

expected_2nd_line_start
the expected start of the second line of iea_file. Default is "COUNTRY,FLOW,PRODUCT".

expected_simple_start
the expected starting of the first line of iea_file. Default is the value of expected_2nd_line_start. Note that expected_simple_start is sometimes encountered in data supplied by the IEA. Furthermore, expected_simple_start could be the format of the file when somebody "helpfully" fiddles with the raw data from the IEA.

Value

a raw data frame of IEA extended energy balance data with appropriate column titles

Details

This function is designed to work even as more years are added in columns at the right of .iea_file, because column names in the output are constructed from the header line(s) of .iea_file (which contain years and country, flow, product information).

Examples

# 2018 and earlier file format
slurp_iea_to_raw_df(text = paste0(",,TIME,1960,1961\n",
                     "COUNTRY,FLOW,PRODUCT\n",
                     "World,Production,Hard coal (if no detail),42,43"))

#>    COUNTRY       FLOW                  PRODUCT 1960 1961
#> 1:   World Production Hard coal (if no detail)   42   43

# With extra commas on the 2nd line
slurp_iea_to_raw_df(text = paste0(",,TIME,1960,1961\n",
                     "COUNTRY,FLOW,PRODUCT,,,\n",
                     "World,Production,Hard coal (if no detail),42,43"))

#>    COUNTRY       FLOW                  PRODUCT 1960 1961
#> 1:   World Production Hard coal (if no detail)   42   43

# With a clean first line (2019 file format)
slurp_iea_to_raw_df(text = paste0("COUNTRY,FLOW,PRODUCT,1960,1961\n",
                     "World,Production,Hard coal (if no detail),42,43"))

#>    COUNTRY       FLOW                  PRODUCT 1960 1961
#> 1:   World Production Hard coal (if no detail)   42   43