Specify all industries

This is a convenience function. It bundles several others:

specify_primary_production()

specify_production_to_resources()

specify_tp_eiou()

specify_interface_industries()

tp_sinks_to_nonenergy()

specify_all(.tidy_iea_df)

Arguments

.tidy_iea_df
a tidy data frame containing IEA extended energy balance data

Value

an enhanced and corrected version of .tidy_iea_df that is ready for physical supply-use table (PSUT) analysis.

Details

Each bundled function is called in turn using default arguments. See examples for two ways to achieve the same result.

Examples

# Simple
load_tidy_iea_df() %>%
  specify_all()

#> # A tibble: 413 x 11
#>    Country Method Energy.type Last.stage  Year Ledger.side Flow.aggregatio…
#>    <chr>   <chr>  <chr>       <chr>      <dbl> <chr>       <chr>           
#>  1 GHA     PCM    E           Final       1971 Consumption Industry        
#>  2 GHA     PCM    E           Final       1971 Consumption Industry        
#>  3 GHA     PCM    E           Final       1971 Consumption Industry        
#>  4 GHA     PCM    E           Final       1971 Consumption Industry        
#>  5 GHA     PCM    E           Final       1971 Consumption Industry        
#>  6 GHA     PCM    E           Final       1971 Consumption Industry        
#>  7 GHA     PCM    E           Final       1971 Consumption Industry        
#>  8 GHA     PCM    E           Final       1971 Consumption Other           
#>  9 GHA     PCM    E           Final       1971 Consumption Other           
#> 10 GHA     PCM    E           Final       1971 Consumption Other           
#> # … with 403 more rows, and 4 more variables: Flow <chr>, Product <chr>,
#> #   Unit <chr>, E.dot <dbl>

# Complicated
load_tidy_iea_df() %>%
  specify_primary_production() %>%
  specify_production_to_resources() %>%
  specify_tp_eiou() %>%
  specify_interface_industries() %>%
  tp_sinks_to_nonenergy()

#> # A tibble: 413 x 11
#>    Country Method Energy.type Last.stage  Year Ledger.side Flow.aggregatio…
#>    <chr>   <chr>  <chr>       <chr>      <dbl> <chr>       <chr>           
#>  1 GHA     PCM    E           Final       1971 Consumption Industry        
#>  2 GHA     PCM    E           Final       1971 Consumption Industry        
#>  3 GHA     PCM    E           Final       1971 Consumption Industry        
#>  4 GHA     PCM    E           Final       1971 Consumption Industry        
#>  5 GHA     PCM    E           Final       1971 Consumption Industry        
#>  6 GHA     PCM    E           Final       1971 Consumption Industry        
#>  7 GHA     PCM    E           Final       1971 Consumption Industry        
#>  8 GHA     PCM    E           Final       1971 Consumption Other           
#>  9 GHA     PCM    E           Final       1971 Consumption Other           
#> 10 GHA     PCM    E           Final       1971 Consumption Other           
#> # … with 403 more rows, and 4 more variables: Flow <chr>, Product <chr>,
#> #   Unit <chr>, E.dot <dbl>