Specify interface industries

An interface industry is one that moves energy carriers into or out of a country. When Flow is any of the interface industries, we need to be more specific. If we don't separate these Flows by Product, we run into trouble with upstream swims (e.g., all Products are produced even if only one is needed) and embodied energy calculations (many types of energy are embodied, even if only one should be). This function adds a suffix [of Product] to each of these interface industries.

specify_interface_industries(
  .tidy_iea_df,
  flow = "Flow",
  int_industries = IEATools::interface_industries,
  product = "Product",
  notation = IEATools::of_notation
)

Arguments

.tidy_iea_df
a tidy data frame containing IEA extended energy balance data

flow
the name of the flow column in .tidy_iea_df. Default is "Flow".

int_industries
a string vector of industries involved in exchanges with other countries, bunkers, or stock changes. Default is IEATools::interface_industries.

product
the name of the product column in .tidy_iea_df. Default is "Product".

notation
a list of specification notations. Default is IEATools::of_notation.

Value

a modified version of .tidy_iea_df with specified interface industries

Details

Note that "Production" also needs to be specified, but that is accomplished in the specify_primary_production() and specify_production_to_resources() functions.

Examples

load_tidy_iea_df() %>% 
  specify_interface_industries()

#> # A tibble: 402 × 11
#>    Country Method Energy.type Last.stage  Year Ledger.side Flow.aggregation.poi…
#>    <chr>   <chr>  <chr>       <chr>      <dbl> <chr>       <chr>                
#>  1 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  2 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  3 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  4 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  5 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  6 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  7 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  8 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  9 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#> 10 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#> # … with 392 more rows, and 4 more variables: Flow <chr>, Product <chr>,
#> #   Unit <chr>, E.dot <dbl>