Specify primary production industries

The IEA extended energy balances include some Flows that identify Energy industry own use (EIOU), the consumption of energy by energy-producing industries. But some primary production industries that receive EIOU do not produce anything. For example, Coal mines receive electricity but there are no Coal mines that produce coal. Rather, the generic Production industry produces coal. This function solves that problem by replacing the generic Production industry with specific industries.

specify_primary_production(
  .tidy_iea_df,
  eiou_destinations = c("Coal mines", "Oil and gas extraction"),
  liquefaction_regas = "Liquefaction (LNG) / regasification plants",
  liquefaction_regas_reassign = "Oil and gas extraction",
  production_products = list(IEATools::primary_coal_products,
    c(IEATools::primary_oil_products, "Natural gas")),
  production_products_short_names = c("Coal", "Oil and natural gas"),
  flow_aggregation_point = "Flow.aggregation.point",
  eiou = "Energy industry own use",
  transformation_processes = "Transformation processes",
  flow = "Flow",
  resources = "Resources",
  production = "Production",
  e_dot = "E.dot",
  product = "Product",
  notation = IEATools::of_notation
)

Arguments

.tidy_iea_df
an IEA data frame whose columns have been renamed by rename_iea_df_cols()

eiou_destinations
a vector of destinations for EIOU for primary production of coal and coal products and oil and natural gas. Default is c("Coal mines", "Oil and gas extraction").

liquefaction_regas
a string identifying liquefaction and regasification plants. Default is "Liquefaction (LNG) / regasification plants".

liquefaction_regas_reassign
a string identifying the industry to which EIOU into liquefaction_regas will be reassigned. Default is "Oil and gas extraction".

production_products
a list of products for which we want to specify primary industries. Default is list(primary_coal_products, c(primary_oil_products, "Natural gas")).

production_products_short_names
a vector of short names for primary industries. Default is c("Coal", "Oil and natural gas").

flow_aggregation_point
the name of the flow aggregation point column in .tidy_iea_df. Default is "Flow.aggregation.point".

eiou
a string identifying energy industry own use in the Flow.aggregation.point column of .tidy_iea_df. Default is "Energy industry own use".

transformation_processes
a string identifying transformation processes in the flow column of .tidy_iea_df. Default is "Transformation processes".

flow
the name of the flow column in .tidy_iea_df. Default is "Flow".

resources
a string identifying resource industries to be added to .tidy_iea_df. Default is "Resources".

production
a string identifying production in the flow column. Default is "Production".

e_dot
the name of the energy column in .tidy_iea_df. Default is "E.dot".

product
the name of the product column in .tidy_iea_df. Default is "Product".

notation
a list of specification notations. Default is IEATools::of_notation.

Value

.tidy_iea_df with adjusted production information for primary energy for both coal and coal products and oil and gas extraction

Details

By default, the following changes are made to .tidy_iea_df:

Energy industry own use for Liquefaction (LNG) / regasification plants is reassigned to Oil and gas extraction.

The Production industry for coal_and_coal_products is replaced by Coal mines.

A Resources [of Coal] industry is created which becomes the source of all primary coal_and_coal_products in the energy conversion chain.

Flows from Resources [of Coal] to Coal mines are added.

The Coal mines (energy) EIOU Flow is replaced by Coal mines.

The Production industry for oil_and_oil_products and Natural gas is replaced by Oil and gas extraction.

A Resources [of Oil and natural gas] industry is created which becomes the source of all primary oil_and_oil_products and Natural gas in the energy conversion chain.

Flows from Resources [of Oil and natural gas] to Oil and gas extraction are added.

The Oil and gas extraction (energy) EIOU Flow is replaced by Oil and gas extraction.

Users can specify other changes by adjusting the default argument values.

Be sure to call this function after calling augment_iea_df() or load_tidy_iea_df().

Examples

library(dplyr)
load_tidy_iea_df() %>%
  specify_primary_production() %>%
  add_psut_matnames() %>%
  filter(Flow %in% c("Resources (Coal)", "Coal mines")) %>%
  select(-Method, -Last.stage, -Ledger.side, -Unit)

#> # A tibble: 8 x 8
#>   Country Energy.type  Year Flow.aggregation… Flow   Product      E.dot matnames
#>   <chr>   <chr>       <dbl> <chr>             <chr>  <chr>        <dbl> <chr>   
#> 1 ZAF     E            1971 Energy industry … Coal … Electrici… -5.24e1 U_EIOU  
#> 2 ZAF     E            2000 Energy industry … Coal … Electrici… -2.37e2 U_EIOU  
#> 3 ZAF     E            1971 Transformation p… Coal … Hard coal… -3.31e4 U_excl_…
#> 4 ZAF     E            2000 Transformation p… Coal … Coking co… -2.37e3 U_excl_…
#> 5 ZAF     E            2000 Transformation p… Coal … Other bit… -1.25e5 U_excl_…
#> 6 ZAF     E            1971 Transformation p… Coal … Hard coal…  3.31e4 V       
#> 7 ZAF     E            2000 Transformation p… Coal … Coking co…  2.37e3 V       
#> 8 ZAF     E            2000 Transformation p… Coal … Other bit…  1.25e5 V       

# EIOU by "Liquefaction (LNG) / regasification plants" is reassigned to "Oil and gas extraction"
data.frame(
  Flow.aggregation.point = c("Energy industry own use"),
  Flow = c("Liquefaction (LNG) / regasification plants"),
  Product = c("Natural gas"),
  E.dot = c(-42),
  stringsAsFactors = FALSE
) %>%
  specify_primary_production()

#> # A tibble: 1 x 4
#>   Flow.aggregation.point Flow              Product                         E.dot
#>   <chr>                  <chr>             <chr>                           <dbl>
#> 1 Energy industry own u… Oil and gas extr… Natural gas [of Oil and gas ex…   -42