Specify destinations for energy industry own use flows into transformation processes

The extended energy balance data from the IEA includes "Energy industry own use" (EIOU) for many transformation processes. Unfortunately, in some cases the EIOU flows into industries that aren't included in transformation processes. For example, "Electricity" is consumed by "Own use in electricity, CHP and heat plants", which is not a transformation process. We have to make some decisions to ensure that EIOU is routed to actual transformation processes. See details for a list of changes made to the .tidy_iea_df data frame.

Usage

specify_tp_eiou(
  .tidy_iea_df,
  split_own_use_elect_chp_heat_using_shares_of = c("input", "output"),
  route_non_specified_eiou = TRUE,
  route_non_specified_tp = TRUE,
  specify_renewable_plants = FALSE,
  specify_electricity_grid = FALSE,
  specify_distribution_industries = FALSE,
  ascribe_eiou_to_renewable_plants = FALSE,
  ascribe_eiou_to_nuclear = FALSE,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  eiou = IEATools::tfc_compare_flows$energy_industry_own_use,
  transformation_processes = IEATools::tfc_compare_flows$transformation_processes,
  flow = IEATools::iea_cols$flow,
  own_use_elect_chp_heat = IEATools::eiou_flows$own_use_elect_chp_heat_plants,
  pumped_storage = IEATools::eiou_flows$pumped_storage_plants,
  nuclear_industry = IEATools::eiou_flows$nuclear_industry,
  e_dot = IEATools::iea_cols$e_dot,
  negzeropos = ".negzeropos",
  main_act_producer_elect =
IEATools::eiou_flows$main_activity_producer_electricity_plants
)

Arguments

.tidy_iea_df

An IEA data frame whose columns have been renamed by rename_iea_df_cols().

split_own_use_elect_chp_heat_using_shares_of

Indicates whether the input or outputs to Main activity producer plants should be use for splitting the Own use in electricity, chp and heat plants EIOU flow. Default is "input".

route_non_specified_eiou

Boolean stating whether non-specified EIOU flows should be routed to existing industries. Default is TRUE.

route_non_specified_tp

Boolean stating whether non-specified transformation processes flows should be routed to existing industries. Default is TRUE.

specify_renewable_plants

Boolean stating whether renewable energy industries should be specified or not. Default is FALSE.

specify_electricity_grid

Boolean stating whether an electricity grid industry should be specified or not. Default is FALSE.

specify_distribution_industries

Boolean stating whether distribution industries should be specified or not. Default is FALSE.

ascribe_eiou_to_renewable_plants

A boolean defining whether a fraction of the EIOU of electricity, CHP and heat plants should be ascribed to the new renewable industries. Default is FALSE.

ascribe_eiou_to_nuclear

A boolean defining whether a fraction of the EIOU of electricity, CHP and heat plants should be ascribed to the new nuclear industry. Default is FALSE.

flow_aggregation_point

The name of the flow aggregation point column in .tidy_iea_df. Default is "Flow.aggregation.point".

eiou

A string identifying energy industry own use in the flow aggregation point column. Default is "Energy industry own use".

transformation_processes

A string identifying transformation processes in the flow aggregation point column. Default is "Transformation processes".

flow

The name of the flow column in .tidy_iea_df. Default is "Flow".

own_use_elect_chp_heat

A string identifying own use in electricity, CHP and heat plants in the flow column. Default is "Own use in electricity, CHP and heat plants".

pumped_storage

A string identifying pumped storage plants in the flow column. Default is "Pumped storage plants".

nuclear_industry

A string identifying nuclear plants in the flow column. Default is "Nuclear industry".

e_dot

The name of the energy flow column in .tidy_iea_df. Default is "Edot".

negzeropos

The name of a temporary column created in .tidy_iea_df. Default is ".negzeropos".

main_act_producer_elect

A string identifying main activity producer electricity plants. Default is "Main activity producer electricity plants".

Value

A modified version of .tidy_iea_df.

Details

The following changes are made to the .tidy_iea_df data frame:

EIOU classified as own_use_elect_chp_heat is sent to main_act_producer_elect.

EIOU classified as pumped_storage is sent to main_act_producer_elect.

EIOU classified as nuclear_industry is sent to main_act_producer_elect.

EIOU classified as non_spec_energy is sent to nonspecenergy_reclassify.

After the changes are made, reassigned EIOU may double-up pre-existing EIOU. For example, a country may already have "Electricity" EIOU flowing into "Main activity producer electricity plants". It may also have EIOU flowing into "Nuclear industry". When we switch the EIOU flow into "Nuclear industry" into "Main activity producer electricity plants", we now have two rows of electricity EIOU into "Main activity producer electricity plants". To avoid double rows, all like rows are summed before returning.

Examples

library(dplyr)
load_tidy_iea_df() %>% 
  specify_tp_eiou() %>% 
  filter(FlowAggregationPoint == "Energy industry own use" & 
           Flow == "Main activity producer electricity plants")
#> # A tibble: 4 × 11
#>   Country Method EnergyType LastStage  Year LedgerSide FlowAggregationPoint   
#>   <chr>   <chr>  <chr>      <chr>     <dbl> <chr>      <chr>                  
#> 1 GHA     PCM    E          Final      1971 Supply     Energy industry own use
#> 2 GHA     PCM    E          Final      2000 Supply     Energy industry own use
#> 3 ZAF     PCM    E          Final      1971 Supply     Energy industry own use
#> 4 ZAF     PCM    E          Final      2000 Supply     Energy industry own use
#> # ℹ 4 more variables: Flow <chr>, Product <chr>, Unit <chr>, Edot <dbl>