Separates EIOU flows of oil and gas extraction

This function separates the EIOU flows of the Oil and gas extraction industry into EIOU flows for the Oil extraction industry and EIOU flows for the Natural gas extraction industry. It uses the shares of production of each of these two industries to separate EIOU flows. As such, the EIOU consumed per unit of output will be the same for the Oil extraction and Natural gas extraction industries.

split_oil_gas_extraction_eiou(
  .tidy_iea_df,
  eiou = IEATools::aggregation_flows$energy_industry_own_use,
  country = IEATools::iea_cols$country,
  energy_type = IEATools::iea_cols$energy_type,
  method = IEATools::iea_cols$method,
  last_stage = IEATools::iea_cols$last_stage,
  ledger_side = IEATools::iea_cols$ledger_side,
  year = IEATools::iea_cols$year,
  flow = IEATools::iea_cols$flow,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  e_dot = IEATools::iea_cols$e_dot,
  oil_gas_extraction = IEATools::eiou_flows$oil_and_gas_extraction,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  oil_extraction = IEATools::industry_flows$oil_extraction,
  gas_extraction = IEATools::industry_flows$natural_gas_extraction,
  .share = ".share"
)

Arguments

.tidy_iea_df
The .tidy_iea_df on which oil and gas extraction EIOU flows need to be separated.

eiou
The name of the Energy industry own use flow aggregation point. Default is IEATools::aggregation_flows$energy_industry_own_use.

country, energy_type, method, last_stage, ledger_side, year, flow, flow_aggregation_point, e_dot
See IEATools::iea_cols.

oil_gas_extraction
The name of the Oil and gas extraction EIOU flow. Default is IEATools::eiou_flows$oil_and_gas_extraction.

transformation_processes
The name of the flow aggregation point referring to transformation processes. Default is IEATools::aggregation_flows$transformation_processes.

oil_extraction
The name of the Oil extraction industry. Default is IEATools::industry_flows$oil_extraction.

gas_extraction
The name of the Natural gas extraction industry. Default is IEATools::industry_flows$natural_gas_extraction.

.share
The name of a temporary column that is added to the data frame. Default is ".share".

Value

A .tidy_iea_df with "Oil and gas extraction" EIOU flows split into 'Oil extraction" and "Natural gas extraction" EIOU flows.

Examples

library(dplyr)
load_tidy_iea_df() %>% 
  split_oil_gas_extraction_eiou()

#> # A tibble: 402 × 11
#>    Country Method Energy.type Last.stage  Year Ledger.side Flow.aggregation.poi…
#>    <chr>   <chr>  <chr>       <chr>      <dbl> <chr>       <chr>                
#>  1 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  2 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  3 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  4 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  5 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  6 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  7 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  8 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#>  9 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#> 10 GHA     PCM    E           Final       1971 Supply      Total primary energy…
#> # … with 392 more rows, and 4 more variables: Flow <chr>, Product <chr>,
#> #   Unit <chr>, E.dot <dbl>