Final-to-useful template column names

A string list containing named names of columns in templates (either final-to-useful allocations or machine efficiency) for final-to-useful analysis. Items in the list provide default values for column and row name function arguments throughout the IEATools package.

Usage

template_cols

Format

A string list with 24 entries

ef_product

The name of the final energy product column in final-to-useful templates.

machine

The name of the machine column in final-to-useful templates.

eu_product

The name of the useful energy product column in final-to-useful templates.

destination

The name of the destination column in final-to-useful templates.

quantity

The name of the quantity column in final-to-useful templates.

maximum_values

The name of the maximum values column in final-to-useful templates.

C_eiou

The name of the EIOU allocation rows in final-to-useful templates.

C_Y

The name of the final demand allocation rows in final-to-useful templates.

C_perc

The name of the percentage allocation rows in final-to-useful templates.

e_dot_max

The name of the maximum E.dot column in final-to-useful templates.

e_dot_dest

The name of the column representing the destination for energy flows in final-to-useful templates.

e_dot_perc

The name of the energy flow percentage column in final-to-useful templates.

e_dot_machine

The name of the column representing energy flow into a machine in final-to-useful templates.

e_dot_machine_tot

The name of the column representing total energy flow into a machine in final-to-useful templates.

e_dot_machine_perc

The name of the machine energy flow percentage column in final-to-useful templates.

e_dot_machine_max_perc

The name of the machine energy flow maximum percentage column in final-to-useful templates.

eta_fu

The name of the final-to-useful machine efficiency column in final-to-useful templates.

phi_pf

The name of the exergy-to-energy ratio column various data frames templates.

phi_u

The name of the useful exergy-to-energy ratio column in final-to-useful templates.

phi

The name of the exergy-to-energy ratio column for primary, final, and useful stages.

c_source

The name of a column containing the source of final-to-useful allocation information.

eta_fu_source

The name of a column containing the source of final-to-useful efficiency data.

phi_source

The name of a column containing the source of exergy-to-energy ratios.

.values

The name of a temporary column that holds values.

Details

Some of these names may represent rows (instead of columns) at some stages of the final-to-useful analysis.

Examples

template_cols
#> $ef_product
#> [1] "Ef.product"
#> 
#> $machine
#> [1] "Machine"
#> 
#> $eu_product
#> [1] "Eu.product"
#> 
#> $destination
#> [1] "Destination"
#> 
#> $quantity
#> [1] "Quantity"
#> 
#> $maximum_values
#> [1] "Maximum.values"
#> 
#> $C_eiou
#> [1] "C_EIOU"
#> 
#> $C_Y
#> [1] "C_Y"
#> 
#> $C_perc
#> [1] "C.perc [%]"
#> 
#> $e_dot_max
#> [1] "E.dot_max"
#> 
#> $e_dot_perc
#> [1] "E.dot [%]"
#> 
#> $e_dot_dest
#> [1] "E.dot_dest"
#> 
#> $e_dot_machine
#> [1] "E.dot_machine"
#> 
#> $e_dot_machine_tot
#> [1] "E.dot_machine_tot"
#> 
#> $e_dot_machine_perc
#> [1] "E.dot_machine [%]"
#> 
#> $e_dot_machine_max_perc
#> [1] "E.dot_machine_max [%]"
#> 
#> $eta_fu
#> [1] "eta.fu"
#> 
#> $phi_pf
#> [1] "phi.pf"
#> 
#> $phi_u
#> [1] "phi.u"
#> 
#> $phi
#> [1] "phi"
#> 
#> $c_source
#> [1] "C.source"
#> 
#> $eta_fu_source
#> [1] "eta.fu.source"
#> 
#> $phi_source
#> [1] "phi.source"
#> 
#> $.values
#> [1] ".values"
#>