Transformation processes

A string vector containing names of Transformation processes Flows in the IEA extended energy balances database.

Usage

transformation_processes

Format

A string vector with 24 entries.

main_activity_producer_electricity_plants

The string that identifies main activity producer electricity plants.

autoproducer_electricity_plants

The string that identifies autoproducer electricity plants.

main_activity_producer_CHP_plants

The string that identifies main activity producer combined heat and power plants.

autoproducer_CHP_plants

The string that identifies autoproducer combined heat and power plants.

main_activity_producer_heat_plants

The string that identifies main activity producer heat plants.

autoproducer_heat_plants

The string that identifies autoproducer heat plants.

heat_pumps

The string that identifies heat pumps.

electric_boilers

The string that identifies electric boilers.

chemical_heat_for_electricity_production

The string that identifies chemical heat for electricity production.

blast_furnaces

The string that identifies blast furnaces.

gas_works

The string that identifies gas works.

coke_ovens

The string that identifies coke ovens.

patent_fuel_plants

The string that identifies patent fuel plants.

bkb_peat_briquette_plants

The string that identifies BKB/peat briquette plants.

oil_refineries

The string that identifies oil refineries.

petrochemical_plants

The string that identifies petrochemical plants.

coal_liquefaction_plants

The string that identifies coal liquefaction plants.

gas_to_liquid_gtl_plants

The string that identifies gas to liquid (GTL) plants.

for_blended_natural_gas

The string that identifies for blended natural gas.

charcoal_production_plants

The string that identifies charcoal production plants.

nuclear_industry

The string that identifies the Nuclear industry

non_specified_transformation

The string that identifies non-specified transformation.

non_specified_energy

The string that identifies non-specified transformation, for the IEA's extended energy balances after 2019.

manufacture

The string that identifies manufacture industries, i.e., those that convert feedstock directly to outputs with 100% efficiency. These industries are added to the energy conversion chain where appropriate.

Examples

transformation_processes
#> $main_activity_producer_electricity_plants
#> [1] "Main activity producer electricity plants"
#> 
#> $autoproducer_electricity_plants
#> [1] "Autoproducer electricity plants"
#> 
#> $main_activity_producer_CHP_plants
#> [1] "Main activity producer CHP plants"
#> 
#> $autoproducer_CHP_plants
#> [1] "Autoproducer CHP plants"
#> 
#> $main_activity_producer_heat_plants
#> [1] "Main activity producer heat plants"
#> 
#> $autoproducer_heat_plants
#> [1] "Autoproducer heat plants"
#> 
#> $heat_pumps
#> [1] "Heat pumps"
#> 
#> $electric_boilers
#> [1] "Electric boilers"
#> 
#> $chemical_heat_for_electricity_production
#> [1] "Chemical heat for electricity production"
#> 
#> $blast_furnaces
#> [1] "Blast furnaces"
#> 
#> $gas_works
#> [1] "Gas works"
#> 
#> $coke_ovens
#> [1] "Coke ovens"
#> 
#> $patent_fuel_plants
#> [1] "Patent fuel plants"
#> 
#> $bkb_peat_briquette_plants
#> [1] "BKB/peat briquette plants"
#> 
#> $oil_refineries
#> [1] "Oil refineries"
#> 
#> $petrochemical_plants
#> [1] "Petrochemical plants"
#> 
#> $coal_liquefaction_plants
#> [1] "Coal liquefaction plants"
#> 
#> $gas_to_liquid_gtl_plants
#> [1] "Gas-to-liquids (GTL) plants"
#> 
#> $for_blended_natural_gas
#> [1] "For blended natural gas"
#> 
#> $charcoal_production_plants
#> [1] "Charcoal production plants"
#> 
#> $nuclear_indsutry
#> [1] "Nuclear industry"
#> 
#> $non_specified_transformation
#> [1] "Non-specified (transformation)"
#> 
#> $non_specified_energy
#> [1] "Non-specified (energy)"
#> 
#> $manufacture
#> [1] "Manufacture"
#>